% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academicons.R
\name{download_academicons}
\alias{download_academicons}
\alias{academicons}
\title{Academicons icons}
\usage{
download_academicons(version = "dev")

academicons(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\value{
\code{download_academicons()} invisibly returns the \code{academicons} icon
set after downloading and installing its SVG files locally.

\code{academicons(name)} returns an \code{icons} vector (an SVG tag) for the
requested icon.
}
\description{
Academicons is a specialist icon font for academics. It contains icons for
websites and organisations related to academia that are often missing from
mainstream font packages. It can be used by itself, but its primary purpose
is to be used as a supplementary package alongside a larger icon set.
}
\section{License}{

Academicons is distributed under the \href{https://scripts.sil.org/OFL}{SIL Open Font License 1.1}
(see the \href{https://github.com/jpswalsh/academicons#license}{project README} for details).
}

