% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.R
\name{ai-alias}
\alias{ai-alias}
\alias{ai-alias}
\title{Academicons alias}
\description{
Academicons alias
}
\keyword{internal} \alias{ai_academia}
\keyword{internal} \alias{ai_academia_square}
\keyword{internal} \alias{ai_acclaim}
\keyword{internal} \alias{ai_acclaim_square}
\keyword{internal} \alias{ai_acm}
\keyword{internal} \alias{ai_acm_square}
\keyword{internal} \alias{ai_ads}
\keyword{internal} \alias{ai_ads_square}
\keyword{internal} \alias{ai_arxiv}
\keyword{internal} \alias{ai_arxiv_square}
\keyword{internal} \alias{ai_biorxiv}
\keyword{internal} \alias{ai_biorxiv_square}
\keyword{internal} \alias{ai_ceur}
\keyword{internal} \alias{ai_ceur_box}
\keyword{internal} \alias{ai_courser}
\keyword{internal} \alias{ai_courser_square}
\keyword{internal} \alias{ai_dataverse}
\keyword{internal} \alias{ai_dataverse_box}
\keyword{internal} \alias{ai_dblp}
\keyword{internal} \alias{ai_dblp_square}
\keyword{internal} \alias{ai_depsy}
\keyword{internal} \alias{ai_depsy_square}
\keyword{internal} \alias{ai_doi}
\keyword{internal} \alias{ai_doi_square}
\keyword{internal} \alias{ai_dryad}
\keyword{internal} \alias{ai_dryad_square}
\keyword{internal} \alias{ai_figshare}
\keyword{internal} \alias{ai_figshare_square}
\keyword{internal} \alias{ai_google_scholar}
\keyword{internal} \alias{ai_google_scholar_square}
\keyword{internal} \alias{ai_ieee}
\keyword{internal} \alias{ai_ieee_square}
\keyword{internal} \alias{ai_impactstory}
\keyword{internal} \alias{ai_impactstory_square}
\keyword{internal} \alias{ai_inspire}
\keyword{internal} \alias{ai_inspire_square}
\keyword{internal} \alias{ai_mendeley}
\keyword{internal} \alias{ai_mendeley_square}
\keyword{internal} \alias{ai_open_access}
\keyword{internal} \alias{ai_open_access_square}
\keyword{internal} \alias{ai_orcid}
\keyword{internal} \alias{ai_orcid_square}
\keyword{internal} \alias{ai_osf}
\keyword{internal} \alias{ai_osf_square}
\keyword{internal} \alias{ai_overleaf}
\keyword{internal} \alias{ai_overleaf_square}
\keyword{internal} \alias{ai_philpapers}
\keyword{internal} \alias{ai_philpapers_square}
\keyword{internal} \alias{ai_piazza}
\keyword{internal} \alias{ai_piazza_square}
\keyword{internal} \alias{ai_publons}
\keyword{internal} \alias{ai_publons_square}
\keyword{internal} \alias{ai_pubmed}
\keyword{internal} \alias{ai_pubmed_square}
\keyword{internal} \alias{ai_researchgate}
\keyword{internal} \alias{ai_researchgate_square}
\keyword{internal} \alias{ai_scirate}
\keyword{internal} \alias{ai_scirate_square}
\keyword{internal} \alias{ai_springer}
\keyword{internal} \alias{ai_springer_square}
\keyword{internal} \alias{ai_zotero}
\keyword{internal} \alias{ai_zotero_square}
