% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.R
\name{ai}
\alias{ai}
\title{Insert Icon From Academicons}
\usage{
ai(name = "open-access", size = 1, fixed_width = FALSE,
  animate = "still", rotate = 0, flip = "none", border = FALSE,
  pull = NULL, other = NULL)
}
\arguments{
\item{name}{The name of the icon}

\item{size}{Size of the icon relative to font size. Options are 1, lg (33%
increase), 2, 3, 4, or 5.}

\item{fixed_width}{If true, icons are set to a fixed width}

\item{animate}{Options are 'still', 'spin', or 'pulse'.}

\item{rotate}{Options are 0, 90, 180, or 270.}

\item{flip}{Options are 'none', 'horizontal', 'vertical'.}

\item{border}{If true, draws a border around the icon.}

\item{pull}{Pulls icon to either 'left' or 'right' and wraps proceeding text
around it.}

\item{other}{Character vector of other parameters directly added to the icon classes}
}
\description{
Insert Icon From Academicons
}
\alias{ai_academia}
\alias{ai_academia_square}
\alias{ai_acclaim}
\alias{ai_acclaim_square}
\alias{ai_acm}
\alias{ai_acm_square}
\alias{ai_ads}
\alias{ai_ads_square}
\alias{ai_arxiv}
\alias{ai_arxiv_square}
\alias{ai_biorxiv}
\alias{ai_biorxiv_square}
\alias{ai_ceur}
\alias{ai_ceur_box}
\alias{ai_courser}
\alias{ai_courser_square}
\alias{ai_dataverse}
\alias{ai_dataverse_box}
\alias{ai_dblp}
\alias{ai_dblp_square}
\alias{ai_depsy}
\alias{ai_depsy_square}
\alias{ai_doi}
\alias{ai_doi_square}
\alias{ai_dryad}
\alias{ai_dryad_square}
\alias{ai_figshare}
\alias{ai_figshare_square}
\alias{ai_google_scholar}
\alias{ai_google_scholar_square}
\alias{ai_ieee}
\alias{ai_ieee_square}
\alias{ai_impactstory}
\alias{ai_impactstory_square}
\alias{ai_inspire}
\alias{ai_inspire_square}
\alias{ai_mendeley}
\alias{ai_mendeley_square}
\alias{ai_open_access}
\alias{ai_open_access_square}
\alias{ai_orcid}
\alias{ai_orcid_square}
\alias{ai_osf}
\alias{ai_osf_square}
\alias{ai_overleaf}
\alias{ai_overleaf_square}
\alias{ai_philpapers}
\alias{ai_philpapers_square}
\alias{ai_piazza}
\alias{ai_piazza_square}
\alias{ai_publons}
\alias{ai_publons_square}
\alias{ai_pubmed}
\alias{ai_pubmed_square}
\alias{ai_researchgate}
\alias{ai_researchgate_square}
\alias{ai_scirate}
\alias{ai_scirate_square}
\alias{ai_springer}
\alias{ai_springer_square}
\alias{ai_zotero}
\alias{ai_zotero_square}
