% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{block_index}
\alias{block_index}
\title{Variable indices belonging to given communities}
\usage{
block_index(x, which = seq_along(x$sizes))
}
\arguments{
\item{x}{An \code{"icons_partition"} object.}

\item{which}{Integer vector of community numbers.  Defaults to all of them.}
}
\value{
A list with \code{index}, the original variable indices in block order,
and \code{community}, the matching community label.
}
\description{
Variable indices belonging to given communities
}
\examples{
data(sim)
part <- icons_detect(cor(sim), threshold = 0.6)
block_index(part, which = 1:2)$index

}
