% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense.R
\name{dense}
\alias{dense}
\title{Adaptive dense subgraph extraction}
\usage{
dense(W_original, threshold = 0.5, lambda = 0.5)
}
\arguments{
\item{W_original}{Input adjacency matrix}

\item{threshold}{Threshold value for filtering edges, default to 0.5}

\item{lambda}{Tuning parameter \eqn{\lambda}{lambda} for greedy peeling algorithm, default to 0.5}
}
\value{
A list containing:\tabular{ll}{
\code{W_dense} \tab Reordered adjacency matrix \cr
\tab \cr
\code{Clist} \tab Node index of the reordered adjacency matrix with detected dense subgraphs\cr
\tab \cr
\code{CID} \tab Number of nodes in each dense subgraph \cr
}
}
\description{
Extract dense subgraphs from the entire graph using the greedy algorithm with \eqn{l_0}{l_0} shrinkage.
}
\examples{
data(sim)
matrix <- cor(sim)
dense(matrix, 0.6, 0.5)
}
\references{
Chen, S., Zhang, Y., Wu, Q., Bi, C., Kochunov, P., & Hong, L. E. (2024).
Identifying covariate-related subnetworks for whole-brain connectome analysis.
Biostatistics (Oxford, England), 25(2), 541–558. https://doi.org/10.1093/biostatistics/kxad007

Wu, Q., Huang, X., Culbreth, A. J., Waltz, J. A., Hong, L. E., & Chen, S. (2022). Extracting brain disease‐related connectome subgraphs by adaptive dense subgraph discovery.
Biometrics, 78(4), 1566–1578. https://doi.org/10.1111/biom.13537
}
