% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{fa-alias}
\alias{fa-alias}
\alias{fa_r}
\title{Font awesome alias}
\description{
Font awesome alias
}
\keyword{internal} \alias{fa_500px}
\keyword{internal} \alias{fa_accessible_icon}
\keyword{internal} \alias{fa_accusoft}
\keyword{internal} \alias{fa_ad}
\keyword{internal} \alias{fa_address_book}
\keyword{internal} \alias{fa_address_card}
\keyword{internal} \alias{fa_adjust}
\keyword{internal} \alias{fa_adn}
\keyword{internal} \alias{fa_adversal}
\keyword{internal} \alias{fa_affiliatetheme}
\keyword{internal} \alias{fa_air_freshener}
\keyword{internal} \alias{fa_algolia}
\keyword{internal} \alias{fa_align_center}
\keyword{internal} \alias{fa_align_justify}
\keyword{internal} \alias{fa_align_left}
\keyword{internal} \alias{fa_align_right}
\keyword{internal} \alias{fa_alipay}
\keyword{internal} \alias{fa_allergies}
\keyword{internal} \alias{fa_amazon}
\keyword{internal} \alias{fa_amazon_pay}
\keyword{internal} \alias{fa_ambulance}
\keyword{internal} \alias{fa_american_sign_language_interpreting}
\keyword{internal} \alias{fa_amilia}
\keyword{internal} \alias{fa_anchor}
\keyword{internal} \alias{fa_android}
\keyword{internal} \alias{fa_angellist}
\keyword{internal} \alias{fa_angle_double_down}
\keyword{internal} \alias{fa_angle_double_left}
\keyword{internal} \alias{fa_angle_double_right}
\keyword{internal} \alias{fa_angle_double_up}
\keyword{internal} \alias{fa_angle_down}
\keyword{internal} \alias{fa_angle_left}
\keyword{internal} \alias{fa_angle_right}
\keyword{internal} \alias{fa_angle_up}
\keyword{internal} \alias{fa_angry}
\keyword{internal} \alias{fa_angrycreative}
\keyword{internal} \alias{fa_angular}
\keyword{internal} \alias{fa_ankh}
\keyword{internal} \alias{fa_app_store}
\keyword{internal} \alias{fa_app_store_ios}
\keyword{internal} \alias{fa_apper}
\keyword{internal} \alias{fa_apple}
\keyword{internal} \alias{fa_apple_alt}
\keyword{internal} \alias{fa_apple_pay}
\keyword{internal} \alias{fa_archive}
\keyword{internal} \alias{fa_archway}
\keyword{internal} \alias{fa_arrow_alt_circle_down}
\keyword{internal} \alias{fa_arrow_alt_circle_left}
\keyword{internal} \alias{fa_arrow_alt_circle_right}
\keyword{internal} \alias{fa_arrow_alt_circle_up}
\keyword{internal} \alias{fa_arrow_circle_down}
\keyword{internal} \alias{fa_arrow_circle_left}
\keyword{internal} \alias{fa_arrow_circle_right}
\keyword{internal} \alias{fa_arrow_circle_up}
\keyword{internal} \alias{fa_arrow_down}
\keyword{internal} \alias{fa_arrow_left}
\keyword{internal} \alias{fa_arrow_right}
\keyword{internal} \alias{fa_arrow_up}
\keyword{internal} \alias{fa_arrows_alt}
\keyword{internal} \alias{fa_arrows_alt_h}
\keyword{internal} \alias{fa_arrows_alt_v}
\keyword{internal} \alias{fa_assistive_listening_systems}
\keyword{internal} \alias{fa_asterisk}
\keyword{internal} \alias{fa_asymmetrik}
\keyword{internal} \alias{fa_at}
\keyword{internal} \alias{fa_atlas}
\keyword{internal} \alias{fa_atom}
\keyword{internal} \alias{fa_audible}
\keyword{internal} \alias{fa_audio_description}
\keyword{internal} \alias{fa_autoprefixer}
\keyword{internal} \alias{fa_avianex}
\keyword{internal} \alias{fa_aviato}
\keyword{internal} \alias{fa_award}
\keyword{internal} \alias{fa_aws}
\keyword{internal} \alias{fa_backspace}
\keyword{internal} \alias{fa_backward}
\keyword{internal} \alias{fa_balance_scale}
\keyword{internal} \alias{fa_ban}
\keyword{internal} \alias{fa_band_aid}
\keyword{internal} \alias{fa_bandcamp}
\keyword{internal} \alias{fa_barcode}
\keyword{internal} \alias{fa_bars}
\keyword{internal} \alias{fa_baseball_ball}
\keyword{internal} \alias{fa_basketball_ball}
\keyword{internal} \alias{fa_bath}
\keyword{internal} \alias{fa_battery_empty}
\keyword{internal} \alias{fa_battery_full}
\keyword{internal} \alias{fa_battery_half}
\keyword{internal} \alias{fa_battery_quarter}
\keyword{internal} \alias{fa_battery_three_quarters}
\keyword{internal} \alias{fa_bed}
\keyword{internal} \alias{fa_beer}
\keyword{internal} \alias{fa_behance}
\keyword{internal} \alias{fa_behance_square}
\keyword{internal} \alias{fa_bell}
\keyword{internal} \alias{fa_bell_slash}
\keyword{internal} \alias{fa_bezier_curve}
\keyword{internal} \alias{fa_bible}
\keyword{internal} \alias{fa_bicycle}
\keyword{internal} \alias{fa_bimobject}
\keyword{internal} \alias{fa_binoculars}
\keyword{internal} \alias{fa_birthday_cake}
\keyword{internal} \alias{fa_bitbucket}
\keyword{internal} \alias{fa_bitcoin}
\keyword{internal} \alias{fa_bity}
\keyword{internal} \alias{fa_black_tie}
\keyword{internal} \alias{fa_blackberry}
\keyword{internal} \alias{fa_blender}
\keyword{internal} \alias{fa_blind}
\keyword{internal} \alias{fa_blogger}
\keyword{internal} \alias{fa_blogger_b}
\keyword{internal} \alias{fa_bluetooth}
\keyword{internal} \alias{fa_bluetooth_b}
\keyword{internal} \alias{fa_bold}
\keyword{internal} \alias{fa_bolt}
\keyword{internal} \alias{fa_bomb}
\keyword{internal} \alias{fa_bone}
\keyword{internal} \alias{fa_bong}
\keyword{internal} \alias{fa_book}
\keyword{internal} \alias{fa_book_open}
\keyword{internal} \alias{fa_book_reader}
\keyword{internal} \alias{fa_bookmark}
\keyword{internal} \alias{fa_bowling_ball}
\keyword{internal} \alias{fa_box}
\keyword{internal} \alias{fa_box_open}
\keyword{internal} \alias{fa_boxes}
\keyword{internal} \alias{fa_braille}
\keyword{internal} \alias{fa_brain}
\keyword{internal} \alias{fa_briefcase}
\keyword{internal} \alias{fa_briefcase_medical}
\keyword{internal} \alias{fa_broadcast_tower}
\keyword{internal} \alias{fa_broom}
\keyword{internal} \alias{fa_brush}
\keyword{internal} \alias{fa_btc}
\keyword{internal} \alias{fa_bug}
\keyword{internal} \alias{fa_building}
\keyword{internal} \alias{fa_bullhorn}
\keyword{internal} \alias{fa_bullseye}
\keyword{internal} \alias{fa_burn}
\keyword{internal} \alias{fa_buromobelexperte}
\keyword{internal} \alias{fa_bus}
\keyword{internal} \alias{fa_bus_alt}
\keyword{internal} \alias{fa_business_time}
\keyword{internal} \alias{fa_buysellads}
\keyword{internal} \alias{fa_calculator}
\keyword{internal} \alias{fa_calendar}
\keyword{internal} \alias{fa_calendar_alt}
\keyword{internal} \alias{fa_calendar_check}
\keyword{internal} \alias{fa_calendar_minus}
\keyword{internal} \alias{fa_calendar_plus}
\keyword{internal} \alias{fa_calendar_times}
\keyword{internal} \alias{fa_camera}
\keyword{internal} \alias{fa_camera_retro}
\keyword{internal} \alias{fa_cannabis}
\keyword{internal} \alias{fa_capsules}
\keyword{internal} \alias{fa_car}
\keyword{internal} \alias{fa_car_alt}
\keyword{internal} \alias{fa_car_battery}
\keyword{internal} \alias{fa_car_crash}
\keyword{internal} \alias{fa_car_side}
\keyword{internal} \alias{fa_caret_down}
\keyword{internal} \alias{fa_caret_left}
\keyword{internal} \alias{fa_caret_right}
\keyword{internal} \alias{fa_caret_square_down}
\keyword{internal} \alias{fa_caret_square_left}
\keyword{internal} \alias{fa_caret_square_right}
\keyword{internal} \alias{fa_caret_square_up}
\keyword{internal} \alias{fa_caret_up}
\keyword{internal} \alias{fa_cart_arrow_down}
\keyword{internal} \alias{fa_cart_plus}
\keyword{internal} \alias{fa_cc_amazon_pay}
\keyword{internal} \alias{fa_cc_amex}
\keyword{internal} \alias{fa_cc_apple_pay}
\keyword{internal} \alias{fa_cc_diners_club}
\keyword{internal} \alias{fa_cc_discover}
\keyword{internal} \alias{fa_cc_jcb}
\keyword{internal} \alias{fa_cc_mastercard}
\keyword{internal} \alias{fa_cc_paypal}
\keyword{internal} \alias{fa_cc_stripe}
\keyword{internal} \alias{fa_cc_visa}
\keyword{internal} \alias{fa_centercode}
\keyword{internal} \alias{fa_certificate}
\keyword{internal} \alias{fa_chalkboard}
\keyword{internal} \alias{fa_chalkboard_teacher}
\keyword{internal} \alias{fa_charging_station}
\keyword{internal} \alias{fa_chart_area}
\keyword{internal} \alias{fa_chart_bar}
\keyword{internal} \alias{fa_chart_line}
\keyword{internal} \alias{fa_chart_pie}
\keyword{internal} \alias{fa_check}
\keyword{internal} \alias{fa_check_circle}
\keyword{internal} \alias{fa_check_double}
\keyword{internal} \alias{fa_check_square}
\keyword{internal} \alias{fa_chess}
\keyword{internal} \alias{fa_chess_bishop}
\keyword{internal} \alias{fa_chess_board}
\keyword{internal} \alias{fa_chess_king}
\keyword{internal} \alias{fa_chess_knight}
\keyword{internal} \alias{fa_chess_pawn}
\keyword{internal} \alias{fa_chess_queen}
\keyword{internal} \alias{fa_chess_rook}
\keyword{internal} \alias{fa_chevron_circle_down}
\keyword{internal} \alias{fa_chevron_circle_left}
\keyword{internal} \alias{fa_chevron_circle_right}
\keyword{internal} \alias{fa_chevron_circle_up}
\keyword{internal} \alias{fa_chevron_down}
\keyword{internal} \alias{fa_chevron_left}
\keyword{internal} \alias{fa_chevron_right}
\keyword{internal} \alias{fa_chevron_up}
\keyword{internal} \alias{fa_child}
\keyword{internal} \alias{fa_chrome}
\keyword{internal} \alias{fa_church}
\keyword{internal} \alias{fa_circle}
\keyword{internal} \alias{fa_circle_notch}
\keyword{internal} \alias{fa_city}
\keyword{internal} \alias{fa_clipboard}
\keyword{internal} \alias{fa_clipboard_check}
\keyword{internal} \alias{fa_clipboard_list}
\keyword{internal} \alias{fa_clock}
\keyword{internal} \alias{fa_clone}
\keyword{internal} \alias{fa_closed_captioning}
\keyword{internal} \alias{fa_cloud}
\keyword{internal} \alias{fa_cloud_download_alt}
\keyword{internal} \alias{fa_cloud_upload_alt}
\keyword{internal} \alias{fa_cloudscale}
\keyword{internal} \alias{fa_cloudsmith}
\keyword{internal} \alias{fa_cloudversify}
\keyword{internal} \alias{fa_cocktail}
\keyword{internal} \alias{fa_code}
\keyword{internal} \alias{fa_code_branch}
\keyword{internal} \alias{fa_codepen}
\keyword{internal} \alias{fa_codiepie}
\keyword{internal} \alias{fa_coffee}
\keyword{internal} \alias{fa_cog}
\keyword{internal} \alias{fa_cogs}
\keyword{internal} \alias{fa_coins}
\keyword{internal} \alias{fa_columns}
\keyword{internal} \alias{fa_comment}
\keyword{internal} \alias{fa_comment_alt}
\keyword{internal} \alias{fa_comment_dollar}
\keyword{internal} \alias{fa_comment_dots}
\keyword{internal} \alias{fa_comment_slash}
\keyword{internal} \alias{fa_comments}
\keyword{internal} \alias{fa_comments_dollar}
\keyword{internal} \alias{fa_compact_disc}
\keyword{internal} \alias{fa_compass}
\keyword{internal} \alias{fa_compress}
\keyword{internal} \alias{fa_concierge_bell}
\keyword{internal} \alias{fa_connectdevelop}
\keyword{internal} \alias{fa_contao}
\keyword{internal} \alias{fa_cookie}
\keyword{internal} \alias{fa_cookie_bite}
\keyword{internal} \alias{fa_copy}
\keyword{internal} \alias{fa_copyright}
\keyword{internal} \alias{fa_couch}
\keyword{internal} \alias{fa_cpanel}
\keyword{internal} \alias{fa_creative_commons}
\keyword{internal} \alias{fa_creative_commons_by}
\keyword{internal} \alias{fa_creative_commons_nc}
\keyword{internal} \alias{fa_creative_commons_nc_eu}
\keyword{internal} \alias{fa_creative_commons_nc_jp}
\keyword{internal} \alias{fa_creative_commons_nd}
\keyword{internal} \alias{fa_creative_commons_pd}
\keyword{internal} \alias{fa_creative_commons_pd_alt}
\keyword{internal} \alias{fa_creative_commons_remix}
\keyword{internal} \alias{fa_creative_commons_sa}
\keyword{internal} \alias{fa_creative_commons_sampling}
\keyword{internal} \alias{fa_creative_commons_sampling_plus}
\keyword{internal} \alias{fa_creative_commons_share}
\keyword{internal} \alias{fa_credit_card}
\keyword{internal} \alias{fa_crop}
\keyword{internal} \alias{fa_crop_alt}
\keyword{internal} \alias{fa_cross}
\keyword{internal} \alias{fa_crosshairs}
\keyword{internal} \alias{fa_crow}
\keyword{internal} \alias{fa_crown}
\keyword{internal} \alias{fa_css3}
\keyword{internal} \alias{fa_css3_alt}
\keyword{internal} \alias{fa_cube}
\keyword{internal} \alias{fa_cubes}
\keyword{internal} \alias{fa_cut}
\keyword{internal} \alias{fa_cuttlefish}
\keyword{internal} \alias{fa_d_and_d}
\keyword{internal} \alias{fa_dashcube}
\keyword{internal} \alias{fa_database}
\keyword{internal} \alias{fa_deaf}
\keyword{internal} \alias{fa_delicious}
\keyword{internal} \alias{fa_deploydog}
\keyword{internal} \alias{fa_deskpro}
\keyword{internal} \alias{fa_desktop}
\keyword{internal} \alias{fa_deviantart}
\keyword{internal} \alias{fa_dharmachakra}
\keyword{internal} \alias{fa_diagnoses}
\keyword{internal} \alias{fa_dice}
\keyword{internal} \alias{fa_dice_five}
\keyword{internal} \alias{fa_dice_four}
\keyword{internal} \alias{fa_dice_one}
\keyword{internal} \alias{fa_dice_six}
\keyword{internal} \alias{fa_dice_three}
\keyword{internal} \alias{fa_dice_two}
\keyword{internal} \alias{fa_digg}
\keyword{internal} \alias{fa_digital_ocean}
\keyword{internal} \alias{fa_digital_tachograph}
\keyword{internal} \alias{fa_directions}
\keyword{internal} \alias{fa_discord}
\keyword{internal} \alias{fa_discourse}
\keyword{internal} \alias{fa_divide}
\keyword{internal} \alias{fa_dizzy}
\keyword{internal} \alias{fa_dna}
\keyword{internal} \alias{fa_dochub}
\keyword{internal} \alias{fa_docker}
\keyword{internal} \alias{fa_dollar_sign}
\keyword{internal} \alias{fa_dolly}
\keyword{internal} \alias{fa_dolly_flatbed}
\keyword{internal} \alias{fa_donate}
\keyword{internal} \alias{fa_door_closed}
\keyword{internal} \alias{fa_door_open}
\keyword{internal} \alias{fa_dot_circle}
\keyword{internal} \alias{fa_dove}
\keyword{internal} \alias{fa_download}
\keyword{internal} \alias{fa_draft2digital}
\keyword{internal} \alias{fa_drafting_compass}
\keyword{internal} \alias{fa_draw_polygon}
\keyword{internal} \alias{fa_dribbble}
\keyword{internal} \alias{fa_dribbble_square}
\keyword{internal} \alias{fa_dropbox}
\keyword{internal} \alias{fa_drum}
\keyword{internal} \alias{fa_drum_steelpan}
\keyword{internal} \alias{fa_drupal}
\keyword{internal} \alias{fa_dumbbell}
\keyword{internal} \alias{fa_dyalog}
\keyword{internal} \alias{fa_earlybirds}
\keyword{internal} \alias{fa_ebay}
\keyword{internal} \alias{fa_edge}
\keyword{internal} \alias{fa_edit}
\keyword{internal} \alias{fa_eject}
\keyword{internal} \alias{fa_elementor}
\keyword{internal} \alias{fa_ellipsis_h}
\keyword{internal} \alias{fa_ellipsis_v}
\keyword{internal} \alias{fa_ello}
\keyword{internal} \alias{fa_ember}
\keyword{internal} \alias{fa_empire}
\keyword{internal} \alias{fa_envelope}
\keyword{internal} \alias{fa_envelope_open}
\keyword{internal} \alias{fa_envelope_open_text}
\keyword{internal} \alias{fa_envelope_square}
\keyword{internal} \alias{fa_envira}
\keyword{internal} \alias{fa_equals}
\keyword{internal} \alias{fa_eraser}
\keyword{internal} \alias{fa_erlang}
\keyword{internal} \alias{fa_ethereum}
\keyword{internal} \alias{fa_etsy}
\keyword{internal} \alias{fa_euro_sign}
\keyword{internal} \alias{fa_exchange_alt}
\keyword{internal} \alias{fa_exclamation}
\keyword{internal} \alias{fa_exclamation_circle}
\keyword{internal} \alias{fa_exclamation_triangle}
\keyword{internal} \alias{fa_expand}
\keyword{internal} \alias{fa_expand_arrows_alt}
\keyword{internal} \alias{fa_expeditedssl}
\keyword{internal} \alias{fa_external_link_alt}
\keyword{internal} \alias{fa_external_link_square_alt}
\keyword{internal} \alias{fa_eye}
\keyword{internal} \alias{fa_eye_dropper}
\keyword{internal} \alias{fa_eye_slash}
\keyword{internal} \alias{fa_facebook}
\keyword{internal} \alias{fa_facebook_f}
\keyword{internal} \alias{fa_facebook_messenger}
\keyword{internal} \alias{fa_facebook_square}
\keyword{internal} \alias{fa_fast_backward}
\keyword{internal} \alias{fa_fast_forward}
\keyword{internal} \alias{fa_fax}
\keyword{internal} \alias{fa_feather}
\keyword{internal} \alias{fa_feather_alt}
\keyword{internal} \alias{fa_female}
\keyword{internal} \alias{fa_fighter_jet}
\keyword{internal} \alias{fa_file}
\keyword{internal} \alias{fa_file_alt}
\keyword{internal} \alias{fa_file_archive}
\keyword{internal} \alias{fa_file_audio}
\keyword{internal} \alias{fa_file_code}
\keyword{internal} \alias{fa_file_contract}
\keyword{internal} \alias{fa_file_download}
\keyword{internal} \alias{fa_file_excel}
\keyword{internal} \alias{fa_file_export}
\keyword{internal} \alias{fa_file_image}
\keyword{internal} \alias{fa_file_import}
\keyword{internal} \alias{fa_file_invoice}
\keyword{internal} \alias{fa_file_invoice_dollar}
\keyword{internal} \alias{fa_file_medical}
\keyword{internal} \alias{fa_file_medical_alt}
\keyword{internal} \alias{fa_file_pdf}
\keyword{internal} \alias{fa_file_powerpoint}
\keyword{internal} \alias{fa_file_prescription}
\keyword{internal} \alias{fa_file_signature}
\keyword{internal} \alias{fa_file_upload}
\keyword{internal} \alias{fa_file_video}
\keyword{internal} \alias{fa_file_word}
\keyword{internal} \alias{fa_fill}
\keyword{internal} \alias{fa_fill_drip}
\keyword{internal} \alias{fa_film}
\keyword{internal} \alias{fa_filter}
\keyword{internal} \alias{fa_fingerprint}
\keyword{internal} \alias{fa_fire}
\keyword{internal} \alias{fa_fire_extinguisher}
\keyword{internal} \alias{fa_firefox}
\keyword{internal} \alias{fa_first_aid}
\keyword{internal} \alias{fa_first_order}
\keyword{internal} \alias{fa_first_order_alt}
\keyword{internal} \alias{fa_firstdraft}
\keyword{internal} \alias{fa_fish}
\keyword{internal} \alias{fa_flag}
\keyword{internal} \alias{fa_flag_checkered}
\keyword{internal} \alias{fa_flask}
\keyword{internal} \alias{fa_flickr}
\keyword{internal} \alias{fa_flipboard}
\keyword{internal} \alias{fa_flushed}
\keyword{internal} \alias{fa_fly}
\keyword{internal} \alias{fa_folder}
\keyword{internal} \alias{fa_folder_minus}
\keyword{internal} \alias{fa_folder_open}
\keyword{internal} \alias{fa_folder_plus}
\keyword{internal} \alias{fa_font}
\keyword{internal} \alias{fa_font_awesome}
\keyword{internal} \alias{fa_font_awesome_alt}
\keyword{internal} \alias{fa_font_awesome_flag}
\keyword{internal} \alias{fa_font_awesome_logo_full}
\keyword{internal} \alias{fa_fonticons}
\keyword{internal} \alias{fa_fonticons_fi}
\keyword{internal} \alias{fa_football_ball}
\keyword{internal} \alias{fa_fort_awesome}
\keyword{internal} \alias{fa_fort_awesome_alt}
\keyword{internal} \alias{fa_forumbee}
\keyword{internal} \alias{fa_forward}
\keyword{internal} \alias{fa_foursquare}
\keyword{internal} \alias{fa_free_code_camp}
\keyword{internal} \alias{fa_freebsd}
\keyword{internal} \alias{fa_frog}
\keyword{internal} \alias{fa_frown}
\keyword{internal} \alias{fa_frown_open}
\keyword{internal} \alias{fa_fulcrum}
\keyword{internal} \alias{fa_funnel_dollar}
\keyword{internal} \alias{fa_futbol}
\keyword{internal} \alias{fa_galactic_republic}
\keyword{internal} \alias{fa_galactic_senate}
\keyword{internal} \alias{fa_gamepad}
\keyword{internal} \alias{fa_gas_pump}
\keyword{internal} \alias{fa_gavel}
\keyword{internal} \alias{fa_gem}
\keyword{internal} \alias{fa_genderless}
\keyword{internal} \alias{fa_get_pocket}
\keyword{internal} \alias{fa_gg}
\keyword{internal} \alias{fa_gg_circle}
\keyword{internal} \alias{fa_gift}
\keyword{internal} \alias{fa_git}
\keyword{internal} \alias{fa_git_square}
\keyword{internal} \alias{fa_github}
\keyword{internal} \alias{fa_github_alt}
\keyword{internal} \alias{fa_github_square}
\keyword{internal} \alias{fa_gitkraken}
\keyword{internal} \alias{fa_gitlab}
\keyword{internal} \alias{fa_gitter}
\keyword{internal} \alias{fa_glass_martini}
\keyword{internal} \alias{fa_glass_martini_alt}
\keyword{internal} \alias{fa_glasses}
\keyword{internal} \alias{fa_glide}
\keyword{internal} \alias{fa_glide_g}
\keyword{internal} \alias{fa_globe}
\keyword{internal} \alias{fa_globe_africa}
\keyword{internal} \alias{fa_globe_americas}
\keyword{internal} \alias{fa_globe_asia}
\keyword{internal} \alias{fa_gofore}
\keyword{internal} \alias{fa_golf_ball}
\keyword{internal} \alias{fa_goodreads}
\keyword{internal} \alias{fa_goodreads_g}
\keyword{internal} \alias{fa_google}
\keyword{internal} \alias{fa_google_drive}
\keyword{internal} \alias{fa_google_play}
\keyword{internal} \alias{fa_google_plus}
\keyword{internal} \alias{fa_google_plus_g}
\keyword{internal} \alias{fa_google_plus_square}
\keyword{internal} \alias{fa_google_wallet}
\keyword{internal} \alias{fa_gopuram}
\keyword{internal} \alias{fa_graduation_cap}
\keyword{internal} \alias{fa_gratipay}
\keyword{internal} \alias{fa_grav}
\keyword{internal} \alias{fa_greater_than}
\keyword{internal} \alias{fa_greater_than_equal}
\keyword{internal} \alias{fa_grimace}
\keyword{internal} \alias{fa_grin}
\keyword{internal} \alias{fa_grin_alt}
\keyword{internal} \alias{fa_grin_beam}
\keyword{internal} \alias{fa_grin_beam_sweat}
\keyword{internal} \alias{fa_grin_hearts}
\keyword{internal} \alias{fa_grin_squint}
\keyword{internal} \alias{fa_grin_squint_tears}
\keyword{internal} \alias{fa_grin_stars}
\keyword{internal} \alias{fa_grin_tears}
\keyword{internal} \alias{fa_grin_tongue}
\keyword{internal} \alias{fa_grin_tongue_squint}
\keyword{internal} \alias{fa_grin_tongue_wink}
\keyword{internal} \alias{fa_grin_wink}
\keyword{internal} \alias{fa_grip_horizontal}
\keyword{internal} \alias{fa_grip_vertical}
\keyword{internal} \alias{fa_gripfire}
\keyword{internal} \alias{fa_grunt}
\keyword{internal} \alias{fa_gulp}
\keyword{internal} \alias{fa_h_square}
\keyword{internal} \alias{fa_hacker_news}
\keyword{internal} \alias{fa_hacker_news_square}
\keyword{internal} \alias{fa_hackerrank}
\keyword{internal} \alias{fa_hamsa}
\keyword{internal} \alias{fa_hand_holding}
\keyword{internal} \alias{fa_hand_holding_heart}
\keyword{internal} \alias{fa_hand_holding_usd}
\keyword{internal} \alias{fa_hand_lizard}
\keyword{internal} \alias{fa_hand_paper}
\keyword{internal} \alias{fa_hand_peace}
\keyword{internal} \alias{fa_hand_point_down}
\keyword{internal} \alias{fa_hand_point_left}
\keyword{internal} \alias{fa_hand_point_right}
\keyword{internal} \alias{fa_hand_point_up}
\keyword{internal} \alias{fa_hand_pointer}
\keyword{internal} \alias{fa_hand_rock}
\keyword{internal} \alias{fa_hand_scissors}
\keyword{internal} \alias{fa_hand_spock}
\keyword{internal} \alias{fa_hands}
\keyword{internal} \alias{fa_hands_helping}
\keyword{internal} \alias{fa_handshake}
\keyword{internal} \alias{fa_hashtag}
\keyword{internal} \alias{fa_haykal}
\keyword{internal} \alias{fa_hdd}
\keyword{internal} \alias{fa_heading}
\keyword{internal} \alias{fa_headphones}
\keyword{internal} \alias{fa_headphones_alt}
\keyword{internal} \alias{fa_headset}
\keyword{internal} \alias{fa_heart}
\keyword{internal} \alias{fa_heartbeat}
\keyword{internal} \alias{fa_helicopter}
\keyword{internal} \alias{fa_highlighter}
\keyword{internal} \alias{fa_hips}
\keyword{internal} \alias{fa_hire_a_helper}
\keyword{internal} \alias{fa_history}
\keyword{internal} \alias{fa_hockey_puck}
\keyword{internal} \alias{fa_home}
\keyword{internal} \alias{fa_hooli}
\keyword{internal} \alias{fa_hornbill}
\keyword{internal} \alias{fa_hospital}
\keyword{internal} \alias{fa_hospital_alt}
\keyword{internal} \alias{fa_hospital_symbol}
\keyword{internal} \alias{fa_hot_tub}
\keyword{internal} \alias{fa_hotel}
\keyword{internal} \alias{fa_hotjar}
\keyword{internal} \alias{fa_hourglass}
\keyword{internal} \alias{fa_hourglass_end}
\keyword{internal} \alias{fa_hourglass_half}
\keyword{internal} \alias{fa_hourglass_start}
\keyword{internal} \alias{fa_houzz}
\keyword{internal} \alias{fa_html5}
\keyword{internal} \alias{fa_hubspot}
\keyword{internal} \alias{fa_i_cursor}
\keyword{internal} \alias{fa_id_badge}
\keyword{internal} \alias{fa_id_card}
\keyword{internal} \alias{fa_id_card_alt}
\keyword{internal} \alias{fa_image}
\keyword{internal} \alias{fa_images}
\keyword{internal} \alias{fa_imdb}
\keyword{internal} \alias{fa_inbox}
\keyword{internal} \alias{fa_indent}
\keyword{internal} \alias{fa_industry}
\keyword{internal} \alias{fa_infinity}
\keyword{internal} \alias{fa_info}
\keyword{internal} \alias{fa_info_circle}
\keyword{internal} \alias{fa_instagram}
\keyword{internal} \alias{fa_internet_explorer}
\keyword{internal} \alias{fa_ioxhost}
\keyword{internal} \alias{fa_italic}
\keyword{internal} \alias{fa_itunes}
\keyword{internal} \alias{fa_itunes_note}
\keyword{internal} \alias{fa_java}
\keyword{internal} \alias{fa_jedi}
\keyword{internal} \alias{fa_jedi_order}
\keyword{internal} \alias{fa_jenkins}
\keyword{internal} \alias{fa_joget}
\keyword{internal} \alias{fa_joint}
\keyword{internal} \alias{fa_joomla}
\keyword{internal} \alias{fa_journal_whills}
\keyword{internal} \alias{fa_js}
\keyword{internal} \alias{fa_js_square}
\keyword{internal} \alias{fa_jsfiddle}
\keyword{internal} \alias{fa_kaaba}
\keyword{internal} \alias{fa_kaggle}
\keyword{internal} \alias{fa_key}
\keyword{internal} \alias{fa_keybase}
\keyword{internal} \alias{fa_keyboard}
\keyword{internal} \alias{fa_keycdn}
\keyword{internal} \alias{fa_khanda}
\keyword{internal} \alias{fa_kickstarter}
\keyword{internal} \alias{fa_kickstarter_k}
\keyword{internal} \alias{fa_kiss}
\keyword{internal} \alias{fa_kiss_beam}
\keyword{internal} \alias{fa_kiss_wink_heart}
\keyword{internal} \alias{fa_kiwi_bird}
\keyword{internal} \alias{fa_korvue}
\keyword{internal} \alias{fa_landmark}
\keyword{internal} \alias{fa_language}
\keyword{internal} \alias{fa_laptop}
\keyword{internal} \alias{fa_laptop_code}
\keyword{internal} \alias{fa_laravel}
\keyword{internal} \alias{fa_lastfm}
\keyword{internal} \alias{fa_lastfm_square}
\keyword{internal} \alias{fa_laugh}
\keyword{internal} \alias{fa_laugh_beam}
\keyword{internal} \alias{fa_laugh_squint}
\keyword{internal} \alias{fa_laugh_wink}
\keyword{internal} \alias{fa_layer_group}
\keyword{internal} \alias{fa_leaf}
\keyword{internal} \alias{fa_leanpub}
\keyword{internal} \alias{fa_lemon}
\keyword{internal} \alias{fa_less}
\keyword{internal} \alias{fa_less_than}
\keyword{internal} \alias{fa_less_than_equal}
\keyword{internal} \alias{fa_level_down_alt}
\keyword{internal} \alias{fa_level_up_alt}
\keyword{internal} \alias{fa_life_ring}
\keyword{internal} \alias{fa_lightbulb}
\keyword{internal} \alias{fa_line}
\keyword{internal} \alias{fa_link}
\keyword{internal} \alias{fa_linkedin}
\keyword{internal} \alias{fa_linkedin_in}
\keyword{internal} \alias{fa_linode}
\keyword{internal} \alias{fa_linux}
\keyword{internal} \alias{fa_lira_sign}
\keyword{internal} \alias{fa_list}
\keyword{internal} \alias{fa_list_alt}
\keyword{internal} \alias{fa_list_ol}
\keyword{internal} \alias{fa_list_ul}
\keyword{internal} \alias{fa_location_arrow}
\keyword{internal} \alias{fa_lock}
\keyword{internal} \alias{fa_lock_open}
\keyword{internal} \alias{fa_long_arrow_alt_down}
\keyword{internal} \alias{fa_long_arrow_alt_left}
\keyword{internal} \alias{fa_long_arrow_alt_right}
\keyword{internal} \alias{fa_long_arrow_alt_up}
\keyword{internal} \alias{fa_low_vision}
\keyword{internal} \alias{fa_luggage_cart}
\keyword{internal} \alias{fa_lyft}
\keyword{internal} \alias{fa_magento}
\keyword{internal} \alias{fa_magic}
\keyword{internal} \alias{fa_magnet}
\keyword{internal} \alias{fa_mail_bulk}
\keyword{internal} \alias{fa_mailchimp}
\keyword{internal} \alias{fa_male}
\keyword{internal} \alias{fa_mandalorian}
\keyword{internal} \alias{fa_map}
\keyword{internal} \alias{fa_map_marked}
\keyword{internal} \alias{fa_map_marked_alt}
\keyword{internal} \alias{fa_map_marker}
\keyword{internal} \alias{fa_map_marker_alt}
\keyword{internal} \alias{fa_map_pin}
\keyword{internal} \alias{fa_map_signs}
\keyword{internal} \alias{fa_markdown}
\keyword{internal} \alias{fa_marker}
\keyword{internal} \alias{fa_mars}
\keyword{internal} \alias{fa_mars_double}
\keyword{internal} \alias{fa_mars_stroke}
\keyword{internal} \alias{fa_mars_stroke_h}
\keyword{internal} \alias{fa_mars_stroke_v}
\keyword{internal} \alias{fa_mastodon}
\keyword{internal} \alias{fa_maxcdn}
\keyword{internal} \alias{fa_medal}
\keyword{internal} \alias{fa_medapps}
\keyword{internal} \alias{fa_medium}
\keyword{internal} \alias{fa_medium_m}
\keyword{internal} \alias{fa_medkit}
\keyword{internal} \alias{fa_medrt}
\keyword{internal} \alias{fa_meetup}
\keyword{internal} \alias{fa_megaport}
\keyword{internal} \alias{fa_meh}
\keyword{internal} \alias{fa_meh_blank}
\keyword{internal} \alias{fa_meh_rolling_eyes}
\keyword{internal} \alias{fa_memory}
\keyword{internal} \alias{fa_menorah}
\keyword{internal} \alias{fa_mercury}
\keyword{internal} \alias{fa_microchip}
\keyword{internal} \alias{fa_microphone}
\keyword{internal} \alias{fa_microphone_alt}
\keyword{internal} \alias{fa_microphone_alt_slash}
\keyword{internal} \alias{fa_microphone_slash}
\keyword{internal} \alias{fa_microscope}
\keyword{internal} \alias{fa_microsoft}
\keyword{internal} \alias{fa_minus}
\keyword{internal} \alias{fa_minus_circle}
\keyword{internal} \alias{fa_minus_square}
\keyword{internal} \alias{fa_mix}
\keyword{internal} \alias{fa_mixcloud}
\keyword{internal} \alias{fa_mizuni}
\keyword{internal} \alias{fa_mobile}
\keyword{internal} \alias{fa_mobile_alt}
\keyword{internal} \alias{fa_modx}
\keyword{internal} \alias{fa_monero}
\keyword{internal} \alias{fa_money_bill}
\keyword{internal} \alias{fa_money_bill_alt}
\keyword{internal} \alias{fa_money_bill_wave}
\keyword{internal} \alias{fa_money_bill_wave_alt}
\keyword{internal} \alias{fa_money_check}
\keyword{internal} \alias{fa_money_check_alt}
\keyword{internal} \alias{fa_monument}
\keyword{internal} \alias{fa_moon}
\keyword{internal} \alias{fa_mortar_pestle}
\keyword{internal} \alias{fa_mosque}
\keyword{internal} \alias{fa_motorcycle}
\keyword{internal} \alias{fa_mouse_pointer}
\keyword{internal} \alias{fa_music}
\keyword{internal} \alias{fa_napster}
\keyword{internal} \alias{fa_neos}
\keyword{internal} \alias{fa_neuter}
\keyword{internal} \alias{fa_newspaper}
\keyword{internal} \alias{fa_nimblr}
\keyword{internal} \alias{fa_nintendo_switch}
\keyword{internal} \alias{fa_node}
\keyword{internal} \alias{fa_node_js}
\keyword{internal} \alias{fa_not_equal}
\keyword{internal} \alias{fa_notes_medical}
\keyword{internal} \alias{fa_npm}
\keyword{internal} \alias{fa_ns8}
\keyword{internal} \alias{fa_nutritionix}
\keyword{internal} \alias{fa_object_group}
\keyword{internal} \alias{fa_object_ungroup}
\keyword{internal} \alias{fa_odnoklassniki}
\keyword{internal} \alias{fa_odnoklassniki_square}
\keyword{internal} \alias{fa_oil_can}
\keyword{internal} \alias{fa_old_republic}
\keyword{internal} \alias{fa_om}
\keyword{internal} \alias{fa_opencart}
\keyword{internal} \alias{fa_openid}
\keyword{internal} \alias{fa_opera}
\keyword{internal} \alias{fa_optin_monster}
\keyword{internal} \alias{fa_osi}
\keyword{internal} \alias{fa_outdent}
\keyword{internal} \alias{fa_page4}
\keyword{internal} \alias{fa_pagelines}
\keyword{internal} \alias{fa_paint_brush}
\keyword{internal} \alias{fa_paint_roller}
\keyword{internal} \alias{fa_palette}
\keyword{internal} \alias{fa_palfed}
\keyword{internal} \alias{fa_pallet}
\keyword{internal} \alias{fa_paper_plane}
\keyword{internal} \alias{fa_paperclip}
\keyword{internal} \alias{fa_parachute_box}
\keyword{internal} \alias{fa_paragraph}
\keyword{internal} \alias{fa_parking}
\keyword{internal} \alias{fa_passport}
\keyword{internal} \alias{fa_pastafarianism}
\keyword{internal} \alias{fa_paste}
\keyword{internal} \alias{fa_patreon}
\keyword{internal} \alias{fa_pause}
\keyword{internal} \alias{fa_pause_circle}
\keyword{internal} \alias{fa_paw}
\keyword{internal} \alias{fa_paypal}
\keyword{internal} \alias{fa_peace}
\keyword{internal} \alias{fa_pen}
\keyword{internal} \alias{fa_pen_alt}
\keyword{internal} \alias{fa_pen_fancy}
\keyword{internal} \alias{fa_pen_nib}
\keyword{internal} \alias{fa_pen_square}
\keyword{internal} \alias{fa_pencil_alt}
\keyword{internal} \alias{fa_pencil_ruler}
\keyword{internal} \alias{fa_people_carry}
\keyword{internal} \alias{fa_percent}
\keyword{internal} \alias{fa_percentage}
\keyword{internal} \alias{fa_periscope}
\keyword{internal} \alias{fa_phabricator}
\keyword{internal} \alias{fa_phoenix_framework}
\keyword{internal} \alias{fa_phoenix_squadron}
\keyword{internal} \alias{fa_phone}
\keyword{internal} \alias{fa_phone_slash}
\keyword{internal} \alias{fa_phone_square}
\keyword{internal} \alias{fa_phone_volume}
\keyword{internal} \alias{fa_php}
\keyword{internal} \alias{fa_pied_piper}
\keyword{internal} \alias{fa_pied_piper_alt}
\keyword{internal} \alias{fa_pied_piper_hat}
\keyword{internal} \alias{fa_pied_piper_pp}
\keyword{internal} \alias{fa_piggy_bank}
\keyword{internal} \alias{fa_pills}
\keyword{internal} \alias{fa_pinterest}
\keyword{internal} \alias{fa_pinterest_p}
\keyword{internal} \alias{fa_pinterest_square}
\keyword{internal} \alias{fa_place_of_worship}
\keyword{internal} \alias{fa_plane}
\keyword{internal} \alias{fa_plane_arrival}
\keyword{internal} \alias{fa_plane_departure}
\keyword{internal} \alias{fa_play}
\keyword{internal} \alias{fa_play_circle}
\keyword{internal} \alias{fa_playstation}
\keyword{internal} \alias{fa_plug}
\keyword{internal} \alias{fa_plus}
\keyword{internal} \alias{fa_plus_circle}
\keyword{internal} \alias{fa_plus_square}
\keyword{internal} \alias{fa_podcast}
\keyword{internal} \alias{fa_poll}
\keyword{internal} \alias{fa_poll_h}
\keyword{internal} \alias{fa_poo}
\keyword{internal} \alias{fa_poop}
\keyword{internal} \alias{fa_portrait}
\keyword{internal} \alias{fa_pound_sign}
\keyword{internal} \alias{fa_power_off}
\keyword{internal} \alias{fa_pray}
\keyword{internal} \alias{fa_praying_hands}
\keyword{internal} \alias{fa_prescription}
\keyword{internal} \alias{fa_prescription_bottle}
\keyword{internal} \alias{fa_prescription_bottle_alt}
\keyword{internal} \alias{fa_print}
\keyword{internal} \alias{fa_procedures}
\keyword{internal} \alias{fa_product_hunt}
\keyword{internal} \alias{fa_project_diagram}
\keyword{internal} \alias{fa_pushed}
\keyword{internal} \alias{fa_puzzle_piece}
\keyword{internal} \alias{fa_python}
\keyword{internal} \alias{fa_qq}
\keyword{internal} \alias{fa_qrcode}
\keyword{internal} \alias{fa_question}
\keyword{internal} \alias{fa_question_circle}
\keyword{internal} \alias{fa_quidditch}
\keyword{internal} \alias{fa_quinscape}
\keyword{internal} \alias{fa_quora}
\keyword{internal} \alias{fa_quote_left}
\keyword{internal} \alias{fa_quote_right}
\keyword{internal} \alias{fa_quran}
\keyword{internal} \alias{fa_r_project}
\keyword{internal} \alias{fa_random}
\keyword{internal} \alias{fa_ravelry}
\keyword{internal} \alias{fa_react}
\keyword{internal} \alias{fa_readme}
\keyword{internal} \alias{fa_rebel}
\keyword{internal} \alias{fa_receipt}
\keyword{internal} \alias{fa_recycle}
\keyword{internal} \alias{fa_red_river}
\keyword{internal} \alias{fa_reddit}
\keyword{internal} \alias{fa_reddit_alien}
\keyword{internal} \alias{fa_reddit_square}
\keyword{internal} \alias{fa_redo}
\keyword{internal} \alias{fa_redo_alt}
\keyword{internal} \alias{fa_registered}
\keyword{internal} \alias{fa_rendact}
\keyword{internal} \alias{fa_renren}
\keyword{internal} \alias{fa_reply}
\keyword{internal} \alias{fa_reply_all}
\keyword{internal} \alias{fa_replyd}
\keyword{internal} \alias{fa_researchgate}
\keyword{internal} \alias{fa_resolving}
\keyword{internal} \alias{fa_retweet}
\keyword{internal} \alias{fa_rev}
\keyword{internal} \alias{fa_ribbon}
\keyword{internal} \alias{fa_road}
\keyword{internal} \alias{fa_robot}
\keyword{internal} \alias{fa_rocket}
\keyword{internal} \alias{fa_rocketchat}
\keyword{internal} \alias{fa_rockrms}
\keyword{internal} \alias{fa_route}
\keyword{internal} \alias{fa_rss}
\keyword{internal} \alias{fa_rss_square}
\keyword{internal} \alias{fa_ruble_sign}
\keyword{internal} \alias{fa_ruler}
\keyword{internal} \alias{fa_ruler_combined}
\keyword{internal} \alias{fa_ruler_horizontal}
\keyword{internal} \alias{fa_ruler_vertical}
\keyword{internal} \alias{fa_rupee_sign}
\keyword{internal} \alias{fa_sad_cry}
\keyword{internal} \alias{fa_sad_tear}
\keyword{internal} \alias{fa_safari}
\keyword{internal} \alias{fa_sass}
\keyword{internal} \alias{fa_save}
\keyword{internal} \alias{fa_schlix}
\keyword{internal} \alias{fa_school}
\keyword{internal} \alias{fa_screwdriver}
\keyword{internal} \alias{fa_scribd}
\keyword{internal} \alias{fa_search}
\keyword{internal} \alias{fa_search_dollar}
\keyword{internal} \alias{fa_search_location}
\keyword{internal} \alias{fa_search_minus}
\keyword{internal} \alias{fa_search_plus}
\keyword{internal} \alias{fa_searchengin}
\keyword{internal} \alias{fa_seedling}
\keyword{internal} \alias{fa_sellcast}
\keyword{internal} \alias{fa_sellsy}
\keyword{internal} \alias{fa_server}
\keyword{internal} \alias{fa_servicestack}
\keyword{internal} \alias{fa_shapes}
\keyword{internal} \alias{fa_share}
\keyword{internal} \alias{fa_share_alt}
\keyword{internal} \alias{fa_share_alt_square}
\keyword{internal} \alias{fa_share_square}
\keyword{internal} \alias{fa_shekel_sign}
\keyword{internal} \alias{fa_shield_alt}
\keyword{internal} \alias{fa_ship}
\keyword{internal} \alias{fa_shipping_fast}
\keyword{internal} \alias{fa_shirtsinbulk}
\keyword{internal} \alias{fa_shoe_prints}
\keyword{internal} \alias{fa_shopping_bag}
\keyword{internal} \alias{fa_shopping_basket}
\keyword{internal} \alias{fa_shopping_cart}
\keyword{internal} \alias{fa_shopware}
\keyword{internal} \alias{fa_shower}
\keyword{internal} \alias{fa_shuttle_van}
\keyword{internal} \alias{fa_sign}
\keyword{internal} \alias{fa_sign_in_alt}
\keyword{internal} \alias{fa_sign_language}
\keyword{internal} \alias{fa_sign_out_alt}
\keyword{internal} \alias{fa_signal}
\keyword{internal} \alias{fa_signature}
\keyword{internal} \alias{fa_simplybuilt}
\keyword{internal} \alias{fa_sistrix}
\keyword{internal} \alias{fa_sitemap}
\keyword{internal} \alias{fa_sith}
\keyword{internal} \alias{fa_skull}
\keyword{internal} \alias{fa_skyatlas}
\keyword{internal} \alias{fa_skype}
\keyword{internal} \alias{fa_slack}
\keyword{internal} \alias{fa_slack_hash}
\keyword{internal} \alias{fa_sliders_h}
\keyword{internal} \alias{fa_slideshare}
\keyword{internal} \alias{fa_smile}
\keyword{internal} \alias{fa_smile_beam}
\keyword{internal} \alias{fa_smile_wink}
\keyword{internal} \alias{fa_smoking}
\keyword{internal} \alias{fa_smoking_ban}
\keyword{internal} \alias{fa_snapchat}
\keyword{internal} \alias{fa_snapchat_ghost}
\keyword{internal} \alias{fa_snapchat_square}
\keyword{internal} \alias{fa_snowflake}
\keyword{internal} \alias{fa_socks}
\keyword{internal} \alias{fa_solar_panel}
\keyword{internal} \alias{fa_sort}
\keyword{internal} \alias{fa_sort_alpha_down}
\keyword{internal} \alias{fa_sort_alpha_up}
\keyword{internal} \alias{fa_sort_amount_down}
\keyword{internal} \alias{fa_sort_amount_up}
\keyword{internal} \alias{fa_sort_down}
\keyword{internal} \alias{fa_sort_numeric_down}
\keyword{internal} \alias{fa_sort_numeric_up}
\keyword{internal} \alias{fa_sort_up}
\keyword{internal} \alias{fa_soundcloud}
\keyword{internal} \alias{fa_spa}
\keyword{internal} \alias{fa_space_shuttle}
\keyword{internal} \alias{fa_speakap}
\keyword{internal} \alias{fa_spinner}
\keyword{internal} \alias{fa_splotch}
\keyword{internal} \alias{fa_spotify}
\keyword{internal} \alias{fa_spray_can}
\keyword{internal} \alias{fa_square}
\keyword{internal} \alias{fa_square_full}
\keyword{internal} \alias{fa_square_root_alt}
\keyword{internal} \alias{fa_squarespace}
\keyword{internal} \alias{fa_stack_exchange}
\keyword{internal} \alias{fa_stack_overflow}
\keyword{internal} \alias{fa_stamp}
\keyword{internal} \alias{fa_star}
\keyword{internal} \alias{fa_star_and_crescent}
\keyword{internal} \alias{fa_star_half}
\keyword{internal} \alias{fa_star_half_alt}
\keyword{internal} \alias{fa_star_of_david}
\keyword{internal} \alias{fa_star_of_life}
\keyword{internal} \alias{fa_staylinked}
\keyword{internal} \alias{fa_steam}
\keyword{internal} \alias{fa_steam_square}
\keyword{internal} \alias{fa_steam_symbol}
\keyword{internal} \alias{fa_step_backward}
\keyword{internal} \alias{fa_step_forward}
\keyword{internal} \alias{fa_stethoscope}
\keyword{internal} \alias{fa_sticker_mule}
\keyword{internal} \alias{fa_sticky_note}
\keyword{internal} \alias{fa_stop}
\keyword{internal} \alias{fa_stop_circle}
\keyword{internal} \alias{fa_stopwatch}
\keyword{internal} \alias{fa_store}
\keyword{internal} \alias{fa_store_alt}
\keyword{internal} \alias{fa_strava}
\keyword{internal} \alias{fa_stream}
\keyword{internal} \alias{fa_street_view}
\keyword{internal} \alias{fa_strikethrough}
\keyword{internal} \alias{fa_stripe}
\keyword{internal} \alias{fa_stripe_s}
\keyword{internal} \alias{fa_stroopwafel}
\keyword{internal} \alias{fa_studiovinari}
\keyword{internal} \alias{fa_stumbleupon}
\keyword{internal} \alias{fa_stumbleupon_circle}
\keyword{internal} \alias{fa_subscript}
\keyword{internal} \alias{fa_subway}
\keyword{internal} \alias{fa_suitcase}
\keyword{internal} \alias{fa_suitcase_rolling}
\keyword{internal} \alias{fa_sun}
\keyword{internal} \alias{fa_superpowers}
\keyword{internal} \alias{fa_superscript}
\keyword{internal} \alias{fa_supple}
\keyword{internal} \alias{fa_surprise}
\keyword{internal} \alias{fa_swatchbook}
\keyword{internal} \alias{fa_swimmer}
\keyword{internal} \alias{fa_swimming_pool}
\keyword{internal} \alias{fa_synagogue}
\keyword{internal} \alias{fa_sync}
\keyword{internal} \alias{fa_sync_alt}
\keyword{internal} \alias{fa_syringe}
\keyword{internal} \alias{fa_table}
\keyword{internal} \alias{fa_table_tennis}
\keyword{internal} \alias{fa_tablet}
\keyword{internal} \alias{fa_tablet_alt}
\keyword{internal} \alias{fa_tablets}
\keyword{internal} \alias{fa_tachometer_alt}
\keyword{internal} \alias{fa_tag}
\keyword{internal} \alias{fa_tags}
\keyword{internal} \alias{fa_tape}
\keyword{internal} \alias{fa_tasks}
\keyword{internal} \alias{fa_taxi}
\keyword{internal} \alias{fa_teamspeak}
\keyword{internal} \alias{fa_teeth}
\keyword{internal} \alias{fa_teeth_open}
\keyword{internal} \alias{fa_telegram}
\keyword{internal} \alias{fa_telegram_plane}
\keyword{internal} \alias{fa_tencent_weibo}
\keyword{internal} \alias{fa_terminal}
\keyword{internal} \alias{fa_text_height}
\keyword{internal} \alias{fa_text_width}
\keyword{internal} \alias{fa_th}
\keyword{internal} \alias{fa_th_large}
\keyword{internal} \alias{fa_th_list}
\keyword{internal} \alias{fa_the_red_yeti}
\keyword{internal} \alias{fa_theater_masks}
\keyword{internal} \alias{fa_themeco}
\keyword{internal} \alias{fa_themeisle}
\keyword{internal} \alias{fa_thermometer}
\keyword{internal} \alias{fa_thermometer_empty}
\keyword{internal} \alias{fa_thermometer_full}
\keyword{internal} \alias{fa_thermometer_half}
\keyword{internal} \alias{fa_thermometer_quarter}
\keyword{internal} \alias{fa_thermometer_three_quarters}
\keyword{internal} \alias{fa_thumbs_down}
\keyword{internal} \alias{fa_thumbs_up}
\keyword{internal} \alias{fa_thumbtack}
\keyword{internal} \alias{fa_ticket_alt}
\keyword{internal} \alias{fa_times}
\keyword{internal} \alias{fa_times_circle}
\keyword{internal} \alias{fa_tint}
\keyword{internal} \alias{fa_tint_slash}
\keyword{internal} \alias{fa_tired}
\keyword{internal} \alias{fa_toggle_off}
\keyword{internal} \alias{fa_toggle_on}
\keyword{internal} \alias{fa_toolbox}
\keyword{internal} \alias{fa_tooth}
\keyword{internal} \alias{fa_torah}
\keyword{internal} \alias{fa_torii_gate}
\keyword{internal} \alias{fa_trade_federation}
\keyword{internal} \alias{fa_trademark}
\keyword{internal} \alias{fa_traffic_light}
\keyword{internal} \alias{fa_train}
\keyword{internal} \alias{fa_transgender}
\keyword{internal} \alias{fa_transgender_alt}
\keyword{internal} \alias{fa_trash}
\keyword{internal} \alias{fa_trash_alt}
\keyword{internal} \alias{fa_tree}
\keyword{internal} \alias{fa_trello}
\keyword{internal} \alias{fa_tripadvisor}
\keyword{internal} \alias{fa_trophy}
\keyword{internal} \alias{fa_truck}
\keyword{internal} \alias{fa_truck_loading}
\keyword{internal} \alias{fa_truck_monster}
\keyword{internal} \alias{fa_truck_moving}
\keyword{internal} \alias{fa_truck_pickup}
\keyword{internal} \alias{fa_tshirt}
\keyword{internal} \alias{fa_tty}
\keyword{internal} \alias{fa_tumblr}
\keyword{internal} \alias{fa_tumblr_square}
\keyword{internal} \alias{fa_tv}
\keyword{internal} \alias{fa_twitch}
\keyword{internal} \alias{fa_twitter}
\keyword{internal} \alias{fa_twitter_square}
\keyword{internal} \alias{fa_typo3}
\keyword{internal} \alias{fa_uber}
\keyword{internal} \alias{fa_uikit}
\keyword{internal} \alias{fa_umbrella}
\keyword{internal} \alias{fa_umbrella_beach}
\keyword{internal} \alias{fa_underline}
\keyword{internal} \alias{fa_undo}
\keyword{internal} \alias{fa_undo_alt}
\keyword{internal} \alias{fa_uniregistry}
\keyword{internal} \alias{fa_universal_access}
\keyword{internal} \alias{fa_university}
\keyword{internal} \alias{fa_unlink}
\keyword{internal} \alias{fa_unlock}
\keyword{internal} \alias{fa_unlock_alt}
\keyword{internal} \alias{fa_untappd}
\keyword{internal} \alias{fa_upload}
\keyword{internal} \alias{fa_usb}
\keyword{internal} \alias{fa_user}
\keyword{internal} \alias{fa_user_alt}
\keyword{internal} \alias{fa_user_alt_slash}
\keyword{internal} \alias{fa_user_astronaut}
\keyword{internal} \alias{fa_user_check}
\keyword{internal} \alias{fa_user_circle}
\keyword{internal} \alias{fa_user_clock}
\keyword{internal} \alias{fa_user_cog}
\keyword{internal} \alias{fa_user_edit}
\keyword{internal} \alias{fa_user_friends}
\keyword{internal} \alias{fa_user_graduate}
\keyword{internal} \alias{fa_user_lock}
\keyword{internal} \alias{fa_user_md}
\keyword{internal} \alias{fa_user_minus}
\keyword{internal} \alias{fa_user_ninja}
\keyword{internal} \alias{fa_user_plus}
\keyword{internal} \alias{fa_user_secret}
\keyword{internal} \alias{fa_user_shield}
\keyword{internal} \alias{fa_user_slash}
\keyword{internal} \alias{fa_user_tag}
\keyword{internal} \alias{fa_user_tie}
\keyword{internal} \alias{fa_user_times}
\keyword{internal} \alias{fa_users}
\keyword{internal} \alias{fa_users_cog}
\keyword{internal} \alias{fa_ussunnah}
\keyword{internal} \alias{fa_utensil_spoon}
\keyword{internal} \alias{fa_utensils}
\keyword{internal} \alias{fa_vaadin}
\keyword{internal} \alias{fa_vector_square}
\keyword{internal} \alias{fa_venus}
\keyword{internal} \alias{fa_venus_double}
\keyword{internal} \alias{fa_venus_mars}
\keyword{internal} \alias{fa_viacoin}
\keyword{internal} \alias{fa_viadeo}
\keyword{internal} \alias{fa_viadeo_square}
\keyword{internal} \alias{fa_vial}
\keyword{internal} \alias{fa_vials}
\keyword{internal} \alias{fa_viber}
\keyword{internal} \alias{fa_video}
\keyword{internal} \alias{fa_video_slash}
\keyword{internal} \alias{fa_vihara}
\keyword{internal} \alias{fa_vimeo}
\keyword{internal} \alias{fa_vimeo_square}
\keyword{internal} \alias{fa_vimeo_v}
\keyword{internal} \alias{fa_vine}
\keyword{internal} \alias{fa_vk}
\keyword{internal} \alias{fa_vnv}
\keyword{internal} \alias{fa_volleyball_ball}
\keyword{internal} \alias{fa_volume_down}
\keyword{internal} \alias{fa_volume_off}
\keyword{internal} \alias{fa_volume_up}
\keyword{internal} \alias{fa_vuejs}
\keyword{internal} \alias{fa_walking}
\keyword{internal} \alias{fa_wallet}
\keyword{internal} \alias{fa_warehouse}
\keyword{internal} \alias{fa_weebly}
\keyword{internal} \alias{fa_weibo}
\keyword{internal} \alias{fa_weight}
\keyword{internal} \alias{fa_weight_hanging}
\keyword{internal} \alias{fa_weixin}
\keyword{internal} \alias{fa_whatsapp}
\keyword{internal} \alias{fa_whatsapp_square}
\keyword{internal} \alias{fa_wheelchair}
\keyword{internal} \alias{fa_whmcs}
\keyword{internal} \alias{fa_wifi}
\keyword{internal} \alias{fa_wikipedia_w}
\keyword{internal} \alias{fa_window_close}
\keyword{internal} \alias{fa_window_maximize}
\keyword{internal} \alias{fa_window_minimize}
\keyword{internal} \alias{fa_window_restore}
\keyword{internal} \alias{fa_windows}
\keyword{internal} \alias{fa_wine_glass}
\keyword{internal} \alias{fa_wine_glass_alt}
\keyword{internal} \alias{fa_wix}
\keyword{internal} \alias{fa_wolf_pack_battalion}
\keyword{internal} \alias{fa_won_sign}
\keyword{internal} \alias{fa_wordpress}
\keyword{internal} \alias{fa_wordpress_simple}
\keyword{internal} \alias{fa_wpbeginner}
\keyword{internal} \alias{fa_wpexplorer}
\keyword{internal} \alias{fa_wpforms}
\keyword{internal} \alias{fa_wrench}
\keyword{internal} \alias{fa_x_ray}
\keyword{internal} \alias{fa_xbox}
\keyword{internal} \alias{fa_xing}
\keyword{internal} \alias{fa_xing_square}
\keyword{internal} \alias{fa_y_combinator}
\keyword{internal} \alias{fa_yahoo}
\keyword{internal} \alias{fa_yandex}
\keyword{internal} \alias{fa_yandex_international}
\keyword{internal} \alias{fa_yelp}
\keyword{internal} \alias{fa_yen_sign}
\keyword{internal} \alias{fa_yin_yang}
\keyword{internal} \alias{fa_yoast}
\keyword{internal} \alias{fa_youtube}
\keyword{internal} \alias{fa_youtube_square}
\keyword{internal} \alias{fa_zhihu}
