% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{download_fa}
\alias{download_fa}
\alias{fa}
\title{Font Awesome icons}
\usage{
download_fa(version = "dev")

fa(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\description{
Font Awesome icons
}
