% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome.R
\name{download_fontawesome}
\alias{download_fontawesome}
\alias{fontawesome}
\title{Font Awesome icons}
\usage{
download_fontawesome(version = "dev")

fontawesome(name, style = NULL)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}

\item{style}{Style of the icon. If NULL, a default style will be chosen for the specified icon.}
}
\value{
\code{download_fontawesome()} invisibly returns the \code{fontawesome}
icon set after downloading and installing its SVG files locally.

\code{fontawesome(name, style)} returns an \code{icon} object (an SVG tag) for
the requested icon.
}
\description{
Font Awesome icons
}
