% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_sigmau}
\alias{get_sigmau}
\title{Calculate \eqn{\Sigma_u}{Sigma_u}, the unexplained residual}
\usage{
get_sigmau(data, CID, Clist)
}
\arguments{
\item{data}{A data matrix.}

\item{CID}{A vector of cluster sizes.}

\item{Clist}{A vector of cluster assignments.}
}
\value{
A list containing \eqn{\Sigma_u}{Sigma_u} (\code{sigma_u}) and its Frobenius norm (\code{sigma_u_norm}).
}
\description{
This function calculates \eqn{\Sigma_u}{Sigma_u} and its Frobenius norm,
given the data and subnetwork detection results.
}
