% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_vectorform}
\alias{get_vectorform}
\title{Convert a Square-Form Matrix to a Vector}
\usage{
get_vectorform(dist)
}
\arguments{
\item{dist}{A square symmetric matrix.}
}
\value{
A vector containing the upper triangular part of the matrix.
}
\description{
Given a distance matrix, this function converts it into a half-vector form,
(a vector of the upper triangular part excluding the diagonal).
}
\examples{
data(sim)
get_vectorform(sim)

mat <- matrix(1:6, nrow = 3)
get_vectorform(mat)
}
