% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy}
\alias{greedy}
\title{Greedy algorithm}
\usage{
greedy(Wp, threshold_DSD, lambda)
}
\arguments{
\item{Wp}{Adjacency matrix}

\item{threshold_DSD}{Threshold value for filtering edges}

\item{lambda}{Parameter for greedy peeling algorithm}
}
\value{
A list containing the resulting adjacency matrix W_DSD_greedy,
cluster list Clist, and cluster ID list CID.
}
\description{
A greedy version of the greedy peeling algorithm.
}
\details{
The algorithm finds a cluster at each time until the average value in the
remainder adjacency matrix is below the mean of the input cluster.
}
\examples{
greedy()
}
