% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_objective.R
\name{greedy_objective}
\alias{greedy_objective}
\title{Objective Function Calculation}
\usage{
greedy_objective(C, lambda)
}
\arguments{
\item{C}{A matrix.}

\item{lambda}{The lambda parameter used in the objective function (default is 0.5).}
}
\value{
The value of the objective function.
}
\description{
This function calculates the objective function value based on the input squareform matrix.
}
\details{
The objective function is defined as:
\deqn{f(C) = \left(\frac{\sum_{i < j} C_{ij}}{n(n-1)/2}\right)^\lambda \cdot \left(\sum_{i < j} C_{ij}\right)^{1-\lambda}}
}
\examples{
C <- matrix(1:6, nrow = 3)
obj_func(C)
}
