% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{icon_path}
\alias{icon_path}
\title{Get the file path of an icon}
\usage{
icon_path(x)
}
\arguments{
\item{x}{An \code{icons} vector.}
}
\value{
A character vector giving the path(s) to the icon's source SVG
file(s) on disk.
}
\description{
Icons created from an SVG file (such as those from an \code{\link[=icon_set]{icon_set()}}, or
read directly with \code{\link[=read_icon]{read_icon()}}) retain the path to their source file.
This is useful for passing icons to other packages that work with SVG
files directly, such as \code{ggplot2} (via \code{grid::rasterGrob()} or similar) or
\code{gt}.
}
\examples{
path <- tempfile(fileext = ".svg")
writeLines(
  '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><path d="M0 0"/></svg>',
  path
)
icon_path(read_icon(path))

}
