% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{icon_set}
\alias{icon_set}
\title{Create a custom icon set}
\usage{
icon_set(path, meta = list(name = "Custom", version = NULL, license = NULL))
}
\arguments{
\item{path}{Path to the icons}

\item{meta}{Meta information for the icons}
}
\value{
An \code{icon_set} object providing access to the SVG files found at
\code{path}, for example via \code{$}.
}
\description{
Create a custom icon set
}
\examples{
dir <- tempfile()
dir.create(dir)
writeLines(
  '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><path d="M0 0"/></svg>',
  file.path(dir, "example.svg")
)

custom <- icon_set(dir)
custom$example

}
