% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{icon_style}
\alias{icon_style}
\title{Customise the style of an icon}
\usage{
icon_style(x, scale = NULL, fill = NULL, rotate = NULL, ...)
}
\arguments{
\item{x}{The icon to style.}

\item{scale}{Scaled size of the icon.}

\item{fill}{The colour for the fill.}

\item{rotate}{The angle to rotate the icon.}

\item{...}{Other CSS rules for the icon style, for example \code{float = "right"}.}
}
\value{
The input \code{x} (an \code{icon} or \code{icon_vec}), with its style updated to
reflect the requested styling.
}
\description{
Customise the style of an icon
}
