% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{icon_uri}
\alias{icon_uri}
\title{Encode a (styled) icon as a self-contained \verb{data:} URI}
\usage{
icon_uri(x)
}
\arguments{
\item{x}{An \code{icon} or \code{icon_vec} object.}
}
\value{
A character vector giving each icon's base64 encoded URI.
}
\description{
Renders an icon's SVG markup to a base64-encoded URI.

This enables self-contained and styled icon for consumers that use image
paths/URLs, including for example: an \verb{<img src>}, a \code{gt::html()} cell,
and a \code{leaflet} marker \code{iconUrl}.
}
