% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{iconset}
\alias{iconset}
\title{Create a custom icon set}
\usage{
iconset(path, meta = list(name = "Custom", version = NULL, license =
  NULL))
}
\arguments{
\item{path}{Path to the icons}

\item{meta}{Meta information for the icons}
}
\description{
Create a custom icon set
}
