% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ii.R
\name{ii-alias}
\alias{ii-alias}
\alias{ii-alias}
\title{Ionicons alias}
\description{
Ionicons alias
}
\keyword{internal} \alias{ii_alert}
\keyword{internal} \alias{ii_alert_circled}
\keyword{internal} \alias{ii_android_add}
\keyword{internal} \alias{ii_android_add_circle}
\keyword{internal} \alias{ii_android_alarm_clock}
\keyword{internal} \alias{ii_android_alert}
\keyword{internal} \alias{ii_android_apps}
\keyword{internal} \alias{ii_android_archive}
\keyword{internal} \alias{ii_android_arrow_back}
\keyword{internal} \alias{ii_android_arrow_down}
\keyword{internal} \alias{ii_android_arrow_dropdown}
\keyword{internal} \alias{ii_android_arrow_dropdown_circle}
\keyword{internal} \alias{ii_android_arrow_dropleft}
\keyword{internal} \alias{ii_android_arrow_dropleft_circle}
\keyword{internal} \alias{ii_android_arrow_dropright}
\keyword{internal} \alias{ii_android_arrow_dropright_circle}
\keyword{internal} \alias{ii_android_arrow_dropup}
\keyword{internal} \alias{ii_android_arrow_dropup_circle}
\keyword{internal} \alias{ii_android_arrow_forward}
\keyword{internal} \alias{ii_android_arrow_up}
\keyword{internal} \alias{ii_android_attach}
\keyword{internal} \alias{ii_android_bar}
\keyword{internal} \alias{ii_android_bicycle}
\keyword{internal} \alias{ii_android_boat}
\keyword{internal} \alias{ii_android_bookmark}
\keyword{internal} \alias{ii_android_bulb}
\keyword{internal} \alias{ii_android_bus}
\keyword{internal} \alias{ii_android_calendar}
\keyword{internal} \alias{ii_android_call}
\keyword{internal} \alias{ii_android_camera}
\keyword{internal} \alias{ii_android_cancel}
\keyword{internal} \alias{ii_android_car}
\keyword{internal} \alias{ii_android_cart}
\keyword{internal} \alias{ii_android_chat}
\keyword{internal} \alias{ii_android_checkbox}
\keyword{internal} \alias{ii_android_checkbox_blank}
\keyword{internal} \alias{ii_android_checkbox_outline}
\keyword{internal} \alias{ii_android_checkbox_outline_blank}
\keyword{internal} \alias{ii_android_checkmark_circle}
\keyword{internal} \alias{ii_android_clipboard}
\keyword{internal} \alias{ii_android_close}
\keyword{internal} \alias{ii_android_cloud}
\keyword{internal} \alias{ii_android_cloud_circle}
\keyword{internal} \alias{ii_android_cloud_done}
\keyword{internal} \alias{ii_android_cloud_outline}
\keyword{internal} \alias{ii_android_color_palette}
\keyword{internal} \alias{ii_android_compass}
\keyword{internal} \alias{ii_android_contact}
\keyword{internal} \alias{ii_android_contacts}
\keyword{internal} \alias{ii_android_contract}
\keyword{internal} \alias{ii_android_create}
\keyword{internal} \alias{ii_android_delete}
\keyword{internal} \alias{ii_android_desktop}
\keyword{internal} \alias{ii_android_document}
\keyword{internal} \alias{ii_android_done}
\keyword{internal} \alias{ii_android_done_all}
\keyword{internal} \alias{ii_android_download}
\keyword{internal} \alias{ii_android_drafts}
\keyword{internal} \alias{ii_android_exit}
\keyword{internal} \alias{ii_android_expand}
\keyword{internal} \alias{ii_android_favorite}
\keyword{internal} \alias{ii_android_favorite_outline}
\keyword{internal} \alias{ii_android_film}
\keyword{internal} \alias{ii_android_folder}
\keyword{internal} \alias{ii_android_folder_open}
\keyword{internal} \alias{ii_android_funnel}
\keyword{internal} \alias{ii_android_globe}
\keyword{internal} \alias{ii_android_hand}
\keyword{internal} \alias{ii_android_hangout}
\keyword{internal} \alias{ii_android_happy}
\keyword{internal} \alias{ii_android_home}
\keyword{internal} \alias{ii_android_image}
\keyword{internal} \alias{ii_android_laptop}
\keyword{internal} \alias{ii_android_list}
\keyword{internal} \alias{ii_android_locate}
\keyword{internal} \alias{ii_android_lock}
\keyword{internal} \alias{ii_android_mail}
\keyword{internal} \alias{ii_android_map}
\keyword{internal} \alias{ii_android_menu}
\keyword{internal} \alias{ii_android_microphone}
\keyword{internal} \alias{ii_android_microphone_off}
\keyword{internal} \alias{ii_android_more_horizontal}
\keyword{internal} \alias{ii_android_more_vertical}
\keyword{internal} \alias{ii_android_navigate}
\keyword{internal} \alias{ii_android_notifications}
\keyword{internal} \alias{ii_android_notifications_none}
\keyword{internal} \alias{ii_android_notifications_off}
\keyword{internal} \alias{ii_android_open}
\keyword{internal} \alias{ii_android_options}
\keyword{internal} \alias{ii_android_people}
\keyword{internal} \alias{ii_android_person}
\keyword{internal} \alias{ii_android_person_add}
\keyword{internal} \alias{ii_android_phone_landscape}
\keyword{internal} \alias{ii_android_phone_portrait}
\keyword{internal} \alias{ii_android_pin}
\keyword{internal} \alias{ii_android_plane}
\keyword{internal} \alias{ii_android_playstore}
\keyword{internal} \alias{ii_android_print}
\keyword{internal} \alias{ii_android_radio_button_off}
\keyword{internal} \alias{ii_android_radio_button_on}
\keyword{internal} \alias{ii_android_refresh}
\keyword{internal} \alias{ii_android_remove}
\keyword{internal} \alias{ii_android_remove_circle}
\keyword{internal} \alias{ii_android_restaurant}
\keyword{internal} \alias{ii_android_sad}
\keyword{internal} \alias{ii_android_search}
\keyword{internal} \alias{ii_android_send}
\keyword{internal} \alias{ii_android_settings}
\keyword{internal} \alias{ii_android_share}
\keyword{internal} \alias{ii_android_share_alt}
\keyword{internal} \alias{ii_android_star}
\keyword{internal} \alias{ii_android_star_half}
\keyword{internal} \alias{ii_android_star_outline}
\keyword{internal} \alias{ii_android_stopwatch}
\keyword{internal} \alias{ii_android_subway}
\keyword{internal} \alias{ii_android_sunny}
\keyword{internal} \alias{ii_android_sync}
\keyword{internal} \alias{ii_android_textsms}
\keyword{internal} \alias{ii_android_time}
\keyword{internal} \alias{ii_android_train}
\keyword{internal} \alias{ii_android_unlock}
\keyword{internal} \alias{ii_android_upload}
\keyword{internal} \alias{ii_android_volume_down}
\keyword{internal} \alias{ii_android_volume_mute}
\keyword{internal} \alias{ii_android_volume_off}
\keyword{internal} \alias{ii_android_volume_up}
\keyword{internal} \alias{ii_android_walk}
\keyword{internal} \alias{ii_android_warning}
\keyword{internal} \alias{ii_android_watch}
\keyword{internal} \alias{ii_android_wifi}
\keyword{internal} \alias{ii_aperture}
\keyword{internal} \alias{ii_archive}
\keyword{internal} \alias{ii_arrow_down_a}
\keyword{internal} \alias{ii_arrow_down_b}
\keyword{internal} \alias{ii_arrow_down_c}
\keyword{internal} \alias{ii_arrow_expand}
\keyword{internal} \alias{ii_arrow_graph_down_left}
\keyword{internal} \alias{ii_arrow_graph_down_right}
\keyword{internal} \alias{ii_arrow_graph_up_left}
\keyword{internal} \alias{ii_arrow_graph_up_right}
\keyword{internal} \alias{ii_arrow_left_a}
\keyword{internal} \alias{ii_arrow_left_b}
\keyword{internal} \alias{ii_arrow_left_c}
\keyword{internal} \alias{ii_arrow_move}
\keyword{internal} \alias{ii_arrow_resize}
\keyword{internal} \alias{ii_arrow_return_left}
\keyword{internal} \alias{ii_arrow_return_right}
\keyword{internal} \alias{ii_arrow_right_a}
\keyword{internal} \alias{ii_arrow_right_b}
\keyword{internal} \alias{ii_arrow_right_c}
\keyword{internal} \alias{ii_arrow_shrink}
\keyword{internal} \alias{ii_arrow_swap}
\keyword{internal} \alias{ii_arrow_up_a}
\keyword{internal} \alias{ii_arrow_up_b}
\keyword{internal} \alias{ii_arrow_up_c}
\keyword{internal} \alias{ii_asterisk}
\keyword{internal} \alias{ii_at}
\keyword{internal} \alias{ii_backspace}
\keyword{internal} \alias{ii_backspace_outline}
\keyword{internal} \alias{ii_bag}
\keyword{internal} \alias{ii_battery_charging}
\keyword{internal} \alias{ii_battery_empty}
\keyword{internal} \alias{ii_battery_full}
\keyword{internal} \alias{ii_battery_half}
\keyword{internal} \alias{ii_battery_low}
\keyword{internal} \alias{ii_beaker}
\keyword{internal} \alias{ii_beer}
\keyword{internal} \alias{ii_bluetooth}
\keyword{internal} \alias{ii_bonfire}
\keyword{internal} \alias{ii_bookmark}
\keyword{internal} \alias{ii_bowtie}
\keyword{internal} \alias{ii_briefcase}
\keyword{internal} \alias{ii_bug}
\keyword{internal} \alias{ii_calculator}
\keyword{internal} \alias{ii_calendar}
\keyword{internal} \alias{ii_camera}
\keyword{internal} \alias{ii_card}
\keyword{internal} \alias{ii_cash}
\keyword{internal} \alias{ii_chatbox}
\keyword{internal} \alias{ii_chatbox_working}
\keyword{internal} \alias{ii_chatboxes}
\keyword{internal} \alias{ii_chatbubble}
\keyword{internal} \alias{ii_chatbubble_working}
\keyword{internal} \alias{ii_chatbubbles}
\keyword{internal} \alias{ii_checkmark}
\keyword{internal} \alias{ii_checkmark_circled}
\keyword{internal} \alias{ii_checkmark_round}
\keyword{internal} \alias{ii_chevron_down}
\keyword{internal} \alias{ii_chevron_left}
\keyword{internal} \alias{ii_chevron_right}
\keyword{internal} \alias{ii_chevron_up}
\keyword{internal} \alias{ii_clipboard}
\keyword{internal} \alias{ii_clock}
\keyword{internal} \alias{ii_close}
\keyword{internal} \alias{ii_close_circled}
\keyword{internal} \alias{ii_close_round}
\keyword{internal} \alias{ii_closed_captioning}
\keyword{internal} \alias{ii_cloud}
\keyword{internal} \alias{ii_code}
\keyword{internal} \alias{ii_code_download}
\keyword{internal} \alias{ii_code_working}
\keyword{internal} \alias{ii_coffee}
\keyword{internal} \alias{ii_compass}
\keyword{internal} \alias{ii_compose}
\keyword{internal} \alias{ii_connection_bars}
\keyword{internal} \alias{ii_contrast}
\keyword{internal} \alias{ii_crop}
\keyword{internal} \alias{ii_cube}
\keyword{internal} \alias{ii_disc}
\keyword{internal} \alias{ii_document}
\keyword{internal} \alias{ii_document_text}
\keyword{internal} \alias{ii_drag}
\keyword{internal} \alias{ii_earth}
\keyword{internal} \alias{ii_easel}
\keyword{internal} \alias{ii_edit}
\keyword{internal} \alias{ii_egg}
\keyword{internal} \alias{ii_eject}
\keyword{internal} \alias{ii_email}
\keyword{internal} \alias{ii_email_unread}
\keyword{internal} \alias{ii_erlenmeyer_flask}
\keyword{internal} \alias{ii_erlenmeyer_flask_bubbles}
\keyword{internal} \alias{ii_eye}
\keyword{internal} \alias{ii_eye_disabled}
\keyword{internal} \alias{ii_female}
\keyword{internal} \alias{ii_filing}
\keyword{internal} \alias{ii_film_marker}
\keyword{internal} \alias{ii_fireball}
\keyword{internal} \alias{ii_flag}
\keyword{internal} \alias{ii_flame}
\keyword{internal} \alias{ii_flash}
\keyword{internal} \alias{ii_flash_off}
\keyword{internal} \alias{ii_folder}
\keyword{internal} \alias{ii_fork}
\keyword{internal} \alias{ii_fork_repo}
\keyword{internal} \alias{ii_forward}
\keyword{internal} \alias{ii_funnel}
\keyword{internal} \alias{ii_gear_a}
\keyword{internal} \alias{ii_gear_b}
\keyword{internal} \alias{ii_grid}
\keyword{internal} \alias{ii_hammer}
\keyword{internal} \alias{ii_happy}
\keyword{internal} \alias{ii_happy_outline}
\keyword{internal} \alias{ii_headphone}
\keyword{internal} \alias{ii_heart}
\keyword{internal} \alias{ii_heart_broken}
\keyword{internal} \alias{ii_help}
\keyword{internal} \alias{ii_help_buoy}
\keyword{internal} \alias{ii_help_circled}
\keyword{internal} \alias{ii_home}
\keyword{internal} \alias{ii_icecream}
\keyword{internal} \alias{ii_image}
\keyword{internal} \alias{ii_images}
\keyword{internal} \alias{ii_information}
\keyword{internal} \alias{ii_information_circled}
\keyword{internal} \alias{ii_ionic}
\keyword{internal} \alias{ii_ios_alarm}
\keyword{internal} \alias{ii_ios_alarm_outline}
\keyword{internal} \alias{ii_ios_albums}
\keyword{internal} \alias{ii_ios_albums_outline}
\keyword{internal} \alias{ii_ios_americanfootball}
\keyword{internal} \alias{ii_ios_americanfootball_outline}
\keyword{internal} \alias{ii_ios_analytics}
\keyword{internal} \alias{ii_ios_analytics_outline}
\keyword{internal} \alias{ii_ios_arrow_back}
\keyword{internal} \alias{ii_ios_arrow_down}
\keyword{internal} \alias{ii_ios_arrow_forward}
\keyword{internal} \alias{ii_ios_arrow_left}
\keyword{internal} \alias{ii_ios_arrow_right}
\keyword{internal} \alias{ii_ios_arrow_thin_down}
\keyword{internal} \alias{ii_ios_arrow_thin_left}
\keyword{internal} \alias{ii_ios_arrow_thin_right}
\keyword{internal} \alias{ii_ios_arrow_thin_up}
\keyword{internal} \alias{ii_ios_arrow_up}
\keyword{internal} \alias{ii_ios_at}
\keyword{internal} \alias{ii_ios_at_outline}
\keyword{internal} \alias{ii_ios_barcode}
\keyword{internal} \alias{ii_ios_barcode_outline}
\keyword{internal} \alias{ii_ios_baseball}
\keyword{internal} \alias{ii_ios_baseball_outline}
\keyword{internal} \alias{ii_ios_basketball}
\keyword{internal} \alias{ii_ios_basketball_outline}
\keyword{internal} \alias{ii_ios_bell}
\keyword{internal} \alias{ii_ios_bell_outline}
\keyword{internal} \alias{ii_ios_body}
\keyword{internal} \alias{ii_ios_body_outline}
\keyword{internal} \alias{ii_ios_bolt}
\keyword{internal} \alias{ii_ios_bolt_outline}
\keyword{internal} \alias{ii_ios_book}
\keyword{internal} \alias{ii_ios_book_outline}
\keyword{internal} \alias{ii_ios_bookmarks}
\keyword{internal} \alias{ii_ios_bookmarks_outline}
\keyword{internal} \alias{ii_ios_box}
\keyword{internal} \alias{ii_ios_box_outline}
\keyword{internal} \alias{ii_ios_briefcase}
\keyword{internal} \alias{ii_ios_briefcase_outline}
\keyword{internal} \alias{ii_ios_browsers}
\keyword{internal} \alias{ii_ios_browsers_outline}
\keyword{internal} \alias{ii_ios_calculator}
\keyword{internal} \alias{ii_ios_calculator_outline}
\keyword{internal} \alias{ii_ios_calendar}
\keyword{internal} \alias{ii_ios_calendar_outline}
\keyword{internal} \alias{ii_ios_camera}
\keyword{internal} \alias{ii_ios_camera_outline}
\keyword{internal} \alias{ii_ios_cart}
\keyword{internal} \alias{ii_ios_cart_outline}
\keyword{internal} \alias{ii_ios_chatboxes}
\keyword{internal} \alias{ii_ios_chatboxes_outline}
\keyword{internal} \alias{ii_ios_chatbubble}
\keyword{internal} \alias{ii_ios_chatbubble_outline}
\keyword{internal} \alias{ii_ios_checkmark}
\keyword{internal} \alias{ii_ios_checkmark_empty}
\keyword{internal} \alias{ii_ios_checkmark_outline}
\keyword{internal} \alias{ii_ios_circle_filled}
\keyword{internal} \alias{ii_ios_circle_outline}
\keyword{internal} \alias{ii_ios_clock}
\keyword{internal} \alias{ii_ios_clock_outline}
\keyword{internal} \alias{ii_ios_close}
\keyword{internal} \alias{ii_ios_close_empty}
\keyword{internal} \alias{ii_ios_close_outline}
\keyword{internal} \alias{ii_ios_cloud}
\keyword{internal} \alias{ii_ios_cloud_download}
\keyword{internal} \alias{ii_ios_cloud_download_outline}
\keyword{internal} \alias{ii_ios_cloud_outline}
\keyword{internal} \alias{ii_ios_cloud_upload}
\keyword{internal} \alias{ii_ios_cloud_upload_outline}
\keyword{internal} \alias{ii_ios_cloudy}
\keyword{internal} \alias{ii_ios_cloudy_night}
\keyword{internal} \alias{ii_ios_cloudy_night_outline}
\keyword{internal} \alias{ii_ios_cloudy_outline}
\keyword{internal} \alias{ii_ios_cog}
\keyword{internal} \alias{ii_ios_cog_outline}
\keyword{internal} \alias{ii_ios_color_filter}
\keyword{internal} \alias{ii_ios_color_filter_outline}
\keyword{internal} \alias{ii_ios_color_wand}
\keyword{internal} \alias{ii_ios_color_wand_outline}
\keyword{internal} \alias{ii_ios_compose}
\keyword{internal} \alias{ii_ios_compose_outline}
\keyword{internal} \alias{ii_ios_contact}
\keyword{internal} \alias{ii_ios_contact_outline}
\keyword{internal} \alias{ii_ios_copy}
\keyword{internal} \alias{ii_ios_copy_outline}
\keyword{internal} \alias{ii_ios_crop}
\keyword{internal} \alias{ii_ios_crop_strong}
\keyword{internal} \alias{ii_ios_download}
\keyword{internal} \alias{ii_ios_download_outline}
\keyword{internal} \alias{ii_ios_drag}
\keyword{internal} \alias{ii_ios_email}
\keyword{internal} \alias{ii_ios_email_outline}
\keyword{internal} \alias{ii_ios_eye}
\keyword{internal} \alias{ii_ios_eye_outline}
\keyword{internal} \alias{ii_ios_fastforward}
\keyword{internal} \alias{ii_ios_fastforward_outline}
\keyword{internal} \alias{ii_ios_filing}
\keyword{internal} \alias{ii_ios_filing_outline}
\keyword{internal} \alias{ii_ios_film}
\keyword{internal} \alias{ii_ios_film_outline}
\keyword{internal} \alias{ii_ios_flag}
\keyword{internal} \alias{ii_ios_flag_outline}
\keyword{internal} \alias{ii_ios_flame}
\keyword{internal} \alias{ii_ios_flame_outline}
\keyword{internal} \alias{ii_ios_flask}
\keyword{internal} \alias{ii_ios_flask_outline}
\keyword{internal} \alias{ii_ios_flower}
\keyword{internal} \alias{ii_ios_flower_outline}
\keyword{internal} \alias{ii_ios_folder}
\keyword{internal} \alias{ii_ios_folder_outline}
\keyword{internal} \alias{ii_ios_football}
\keyword{internal} \alias{ii_ios_football_outline}
\keyword{internal} \alias{ii_ios_game_controller_a}
\keyword{internal} \alias{ii_ios_game_controller_a_outline}
\keyword{internal} \alias{ii_ios_game_controller_b}
\keyword{internal} \alias{ii_ios_game_controller_b_outline}
\keyword{internal} \alias{ii_ios_gear}
\keyword{internal} \alias{ii_ios_gear_outline}
\keyword{internal} \alias{ii_ios_glasses}
\keyword{internal} \alias{ii_ios_glasses_outline}
\keyword{internal} \alias{ii_ios_grid_view}
\keyword{internal} \alias{ii_ios_grid_view_outline}
\keyword{internal} \alias{ii_ios_heart}
\keyword{internal} \alias{ii_ios_heart_outline}
\keyword{internal} \alias{ii_ios_help}
\keyword{internal} \alias{ii_ios_help_empty}
\keyword{internal} \alias{ii_ios_help_outline}
\keyword{internal} \alias{ii_ios_home}
\keyword{internal} \alias{ii_ios_home_outline}
\keyword{internal} \alias{ii_ios_infinite}
\keyword{internal} \alias{ii_ios_infinite_outline}
\keyword{internal} \alias{ii_ios_information}
\keyword{internal} \alias{ii_ios_information_empty}
\keyword{internal} \alias{ii_ios_information_outline}
\keyword{internal} \alias{ii_ios_ionic_outline}
\keyword{internal} \alias{ii_ios_keypad}
\keyword{internal} \alias{ii_ios_keypad_outline}
\keyword{internal} \alias{ii_ios_lightbulb}
\keyword{internal} \alias{ii_ios_lightbulb_outline}
\keyword{internal} \alias{ii_ios_list}
\keyword{internal} \alias{ii_ios_list_outline}
\keyword{internal} \alias{ii_ios_location}
\keyword{internal} \alias{ii_ios_location_outline}
\keyword{internal} \alias{ii_ios_locked}
\keyword{internal} \alias{ii_ios_locked_outline}
\keyword{internal} \alias{ii_ios_loop}
\keyword{internal} \alias{ii_ios_loop_strong}
\keyword{internal} \alias{ii_ios_medical}
\keyword{internal} \alias{ii_ios_medical_outline}
\keyword{internal} \alias{ii_ios_medkit}
\keyword{internal} \alias{ii_ios_medkit_outline}
\keyword{internal} \alias{ii_ios_mic}
\keyword{internal} \alias{ii_ios_mic_off}
\keyword{internal} \alias{ii_ios_mic_outline}
\keyword{internal} \alias{ii_ios_minus}
\keyword{internal} \alias{ii_ios_minus_empty}
\keyword{internal} \alias{ii_ios_minus_outline}
\keyword{internal} \alias{ii_ios_monitor}
\keyword{internal} \alias{ii_ios_monitor_outline}
\keyword{internal} \alias{ii_ios_moon}
\keyword{internal} \alias{ii_ios_moon_outline}
\keyword{internal} \alias{ii_ios_more}
\keyword{internal} \alias{ii_ios_more_outline}
\keyword{internal} \alias{ii_ios_musical_note}
\keyword{internal} \alias{ii_ios_musical_notes}
\keyword{internal} \alias{ii_ios_navigate}
\keyword{internal} \alias{ii_ios_navigate_outline}
\keyword{internal} \alias{ii_ios_nutrition}
\keyword{internal} \alias{ii_ios_nutrition_outline}
\keyword{internal} \alias{ii_ios_paper}
\keyword{internal} \alias{ii_ios_paper_outline}
\keyword{internal} \alias{ii_ios_paperplane}
\keyword{internal} \alias{ii_ios_paperplane_outline}
\keyword{internal} \alias{ii_ios_partlysunny}
\keyword{internal} \alias{ii_ios_partlysunny_outline}
\keyword{internal} \alias{ii_ios_pause}
\keyword{internal} \alias{ii_ios_pause_outline}
\keyword{internal} \alias{ii_ios_paw}
\keyword{internal} \alias{ii_ios_paw_outline}
\keyword{internal} \alias{ii_ios_people}
\keyword{internal} \alias{ii_ios_people_outline}
\keyword{internal} \alias{ii_ios_person}
\keyword{internal} \alias{ii_ios_person_outline}
\keyword{internal} \alias{ii_ios_personadd}
\keyword{internal} \alias{ii_ios_personadd_outline}
\keyword{internal} \alias{ii_ios_photos}
\keyword{internal} \alias{ii_ios_photos_outline}
\keyword{internal} \alias{ii_ios_pie}
\keyword{internal} \alias{ii_ios_pie_outline}
\keyword{internal} \alias{ii_ios_pint}
\keyword{internal} \alias{ii_ios_pint_outline}
\keyword{internal} \alias{ii_ios_play}
\keyword{internal} \alias{ii_ios_play_outline}
\keyword{internal} \alias{ii_ios_plus}
\keyword{internal} \alias{ii_ios_plus_empty}
\keyword{internal} \alias{ii_ios_plus_outline}
\keyword{internal} \alias{ii_ios_pricetag}
\keyword{internal} \alias{ii_ios_pricetag_outline}
\keyword{internal} \alias{ii_ios_pricetags}
\keyword{internal} \alias{ii_ios_pricetags_outline}
\keyword{internal} \alias{ii_ios_printer}
\keyword{internal} \alias{ii_ios_printer_outline}
\keyword{internal} \alias{ii_ios_pulse}
\keyword{internal} \alias{ii_ios_pulse_strong}
\keyword{internal} \alias{ii_ios_rainy}
\keyword{internal} \alias{ii_ios_rainy_outline}
\keyword{internal} \alias{ii_ios_recording}
\keyword{internal} \alias{ii_ios_recording_outline}
\keyword{internal} \alias{ii_ios_redo}
\keyword{internal} \alias{ii_ios_redo_outline}
\keyword{internal} \alias{ii_ios_refresh}
\keyword{internal} \alias{ii_ios_refresh_empty}
\keyword{internal} \alias{ii_ios_refresh_outline}
\keyword{internal} \alias{ii_ios_reload}
\keyword{internal} \alias{ii_ios_reverse_camera}
\keyword{internal} \alias{ii_ios_reverse_camera_outline}
\keyword{internal} \alias{ii_ios_rewind}
\keyword{internal} \alias{ii_ios_rewind_outline}
\keyword{internal} \alias{ii_ios_rose}
\keyword{internal} \alias{ii_ios_rose_outline}
\keyword{internal} \alias{ii_ios_search}
\keyword{internal} \alias{ii_ios_search_strong}
\keyword{internal} \alias{ii_ios_settings}
\keyword{internal} \alias{ii_ios_settings_strong}
\keyword{internal} \alias{ii_ios_shuffle}
\keyword{internal} \alias{ii_ios_shuffle_strong}
\keyword{internal} \alias{ii_ios_skipbackward}
\keyword{internal} \alias{ii_ios_skipbackward_outline}
\keyword{internal} \alias{ii_ios_skipforward}
\keyword{internal} \alias{ii_ios_skipforward_outline}
\keyword{internal} \alias{ii_ios_snowy}
\keyword{internal} \alias{ii_ios_speedometer}
\keyword{internal} \alias{ii_ios_speedometer_outline}
\keyword{internal} \alias{ii_ios_star}
\keyword{internal} \alias{ii_ios_star_half}
\keyword{internal} \alias{ii_ios_star_outline}
\keyword{internal} \alias{ii_ios_stopwatch}
\keyword{internal} \alias{ii_ios_stopwatch_outline}
\keyword{internal} \alias{ii_ios_sunny}
\keyword{internal} \alias{ii_ios_sunny_outline}
\keyword{internal} \alias{ii_ios_telephone}
\keyword{internal} \alias{ii_ios_telephone_outline}
\keyword{internal} \alias{ii_ios_tennisball}
\keyword{internal} \alias{ii_ios_tennisball_outline}
\keyword{internal} \alias{ii_ios_thunderstorm}
\keyword{internal} \alias{ii_ios_thunderstorm_outline}
\keyword{internal} \alias{ii_ios_time}
\keyword{internal} \alias{ii_ios_time_outline}
\keyword{internal} \alias{ii_ios_timer}
\keyword{internal} \alias{ii_ios_timer_outline}
\keyword{internal} \alias{ii_ios_toggle}
\keyword{internal} \alias{ii_ios_toggle_outline}
\keyword{internal} \alias{ii_ios_trash}
\keyword{internal} \alias{ii_ios_trash_outline}
\keyword{internal} \alias{ii_ios_undo}
\keyword{internal} \alias{ii_ios_undo_outline}
\keyword{internal} \alias{ii_ios_unlocked}
\keyword{internal} \alias{ii_ios_unlocked_outline}
\keyword{internal} \alias{ii_ios_upload}
\keyword{internal} \alias{ii_ios_upload_outline}
\keyword{internal} \alias{ii_ios_videocam}
\keyword{internal} \alias{ii_ios_videocam_outline}
\keyword{internal} \alias{ii_ios_volume_high}
\keyword{internal} \alias{ii_ios_volume_low}
\keyword{internal} \alias{ii_ios_wineglass}
\keyword{internal} \alias{ii_ios_wineglass_outline}
\keyword{internal} \alias{ii_ios_world}
\keyword{internal} \alias{ii_ios_world_outline}
\keyword{internal} \alias{ii_ipad}
\keyword{internal} \alias{ii_iphone}
\keyword{internal} \alias{ii_ipod}
\keyword{internal} \alias{ii_jet}
\keyword{internal} \alias{ii_key}
\keyword{internal} \alias{ii_knife}
\keyword{internal} \alias{ii_laptop}
\keyword{internal} \alias{ii_leaf}
\keyword{internal} \alias{ii_levels}
\keyword{internal} \alias{ii_lightbulb}
\keyword{internal} \alias{ii_link}
\keyword{internal} \alias{ii_load_a}
\keyword{internal} \alias{ii_load_b}
\keyword{internal} \alias{ii_load_c}
\keyword{internal} \alias{ii_load_d}
\keyword{internal} \alias{ii_location}
\keyword{internal} \alias{ii_lock_combination}
\keyword{internal} \alias{ii_locked}
\keyword{internal} \alias{ii_log_in}
\keyword{internal} \alias{ii_log_out}
\keyword{internal} \alias{ii_loop}
\keyword{internal} \alias{ii_magnet}
\keyword{internal} \alias{ii_male}
\keyword{internal} \alias{ii_man}
\keyword{internal} \alias{ii_map}
\keyword{internal} \alias{ii_medkit}
\keyword{internal} \alias{ii_merge}
\keyword{internal} \alias{ii_mic_a}
\keyword{internal} \alias{ii_mic_b}
\keyword{internal} \alias{ii_mic_c}
\keyword{internal} \alias{ii_minus}
\keyword{internal} \alias{ii_minus_circled}
\keyword{internal} \alias{ii_minus_round}
\keyword{internal} \alias{ii_model_s}
\keyword{internal} \alias{ii_monitor}
\keyword{internal} \alias{ii_more}
\keyword{internal} \alias{ii_mouse}
\keyword{internal} \alias{ii_music_note}
\keyword{internal} \alias{ii_navicon}
\keyword{internal} \alias{ii_navicon_round}
\keyword{internal} \alias{ii_navigate}
\keyword{internal} \alias{ii_network}
\keyword{internal} \alias{ii_no_smoking}
\keyword{internal} \alias{ii_nuclear}
\keyword{internal} \alias{ii_outlet}
\keyword{internal} \alias{ii_paintbrush}
\keyword{internal} \alias{ii_paintbucket}
\keyword{internal} \alias{ii_paper_airplane}
\keyword{internal} \alias{ii_paperclip}
\keyword{internal} \alias{ii_pause}
\keyword{internal} \alias{ii_person}
\keyword{internal} \alias{ii_person_add}
\keyword{internal} \alias{ii_person_stalker}
\keyword{internal} \alias{ii_pie_graph}
\keyword{internal} \alias{ii_pin}
\keyword{internal} \alias{ii_pinpoint}
\keyword{internal} \alias{ii_pizza}
\keyword{internal} \alias{ii_plane}
\keyword{internal} \alias{ii_planet}
\keyword{internal} \alias{ii_play}
\keyword{internal} \alias{ii_playstation}
\keyword{internal} \alias{ii_plus}
\keyword{internal} \alias{ii_plus_circled}
\keyword{internal} \alias{ii_plus_round}
\keyword{internal} \alias{ii_podium}
\keyword{internal} \alias{ii_pound}
\keyword{internal} \alias{ii_power}
\keyword{internal} \alias{ii_pricetag}
\keyword{internal} \alias{ii_pricetags}
\keyword{internal} \alias{ii_printer}
\keyword{internal} \alias{ii_pull_request}
\keyword{internal} \alias{ii_qr_scanner}
\keyword{internal} \alias{ii_quote}
\keyword{internal} \alias{ii_radio_waves}
\keyword{internal} \alias{ii_record}
\keyword{internal} \alias{ii_refresh}
\keyword{internal} \alias{ii_reply}
\keyword{internal} \alias{ii_reply_all}
\keyword{internal} \alias{ii_ribbon_a}
\keyword{internal} \alias{ii_ribbon_b}
\keyword{internal} \alias{ii_sad}
\keyword{internal} \alias{ii_sad_outline}
\keyword{internal} \alias{ii_scissors}
\keyword{internal} \alias{ii_search}
\keyword{internal} \alias{ii_settings}
\keyword{internal} \alias{ii_share}
\keyword{internal} \alias{ii_shuffle}
\keyword{internal} \alias{ii_skip_backward}
\keyword{internal} \alias{ii_skip_forward}
\keyword{internal} \alias{ii_social_android}
\keyword{internal} \alias{ii_social_android_outline}
\keyword{internal} \alias{ii_social_angular}
\keyword{internal} \alias{ii_social_angular_outline}
\keyword{internal} \alias{ii_social_apple}
\keyword{internal} \alias{ii_social_apple_outline}
\keyword{internal} \alias{ii_social_bitcoin}
\keyword{internal} \alias{ii_social_bitcoin_outline}
\keyword{internal} \alias{ii_social_buffer}
\keyword{internal} \alias{ii_social_buffer_outline}
\keyword{internal} \alias{ii_social_chrome}
\keyword{internal} \alias{ii_social_chrome_outline}
\keyword{internal} \alias{ii_social_codepen}
\keyword{internal} \alias{ii_social_codepen_outline}
\keyword{internal} \alias{ii_social_css3}
\keyword{internal} \alias{ii_social_css3_outline}
\keyword{internal} \alias{ii_social_designernews}
\keyword{internal} \alias{ii_social_designernews_outline}
\keyword{internal} \alias{ii_social_dribbble}
\keyword{internal} \alias{ii_social_dribbble_outline}
\keyword{internal} \alias{ii_social_dropbox}
\keyword{internal} \alias{ii_social_dropbox_outline}
\keyword{internal} \alias{ii_social_euro}
\keyword{internal} \alias{ii_social_euro_outline}
\keyword{internal} \alias{ii_social_facebook}
\keyword{internal} \alias{ii_social_facebook_outline}
\keyword{internal} \alias{ii_social_foursquare}
\keyword{internal} \alias{ii_social_foursquare_outline}
\keyword{internal} \alias{ii_social_freebsd_devil}
\keyword{internal} \alias{ii_social_github}
\keyword{internal} \alias{ii_social_github_outline}
\keyword{internal} \alias{ii_social_google}
\keyword{internal} \alias{ii_social_google_outline}
\keyword{internal} \alias{ii_social_googleplus}
\keyword{internal} \alias{ii_social_googleplus_outline}
\keyword{internal} \alias{ii_social_hackernews}
\keyword{internal} \alias{ii_social_hackernews_outline}
\keyword{internal} \alias{ii_social_html5}
\keyword{internal} \alias{ii_social_html5_outline}
\keyword{internal} \alias{ii_social_instagram}
\keyword{internal} \alias{ii_social_instagram_outline}
\keyword{internal} \alias{ii_social_javascript}
\keyword{internal} \alias{ii_social_javascript_outline}
\keyword{internal} \alias{ii_social_linkedin}
\keyword{internal} \alias{ii_social_linkedin_outline}
\keyword{internal} \alias{ii_social_markdown}
\keyword{internal} \alias{ii_social_nodejs}
\keyword{internal} \alias{ii_social_octocat}
\keyword{internal} \alias{ii_social_pinterest}
\keyword{internal} \alias{ii_social_pinterest_outline}
\keyword{internal} \alias{ii_social_python}
\keyword{internal} \alias{ii_social_reddit}
\keyword{internal} \alias{ii_social_reddit_outline}
\keyword{internal} \alias{ii_social_rss}
\keyword{internal} \alias{ii_social_rss_outline}
\keyword{internal} \alias{ii_social_sass}
\keyword{internal} \alias{ii_social_skype}
\keyword{internal} \alias{ii_social_skype_outline}
\keyword{internal} \alias{ii_social_snapchat}
\keyword{internal} \alias{ii_social_snapchat_outline}
\keyword{internal} \alias{ii_social_tumblr}
\keyword{internal} \alias{ii_social_tumblr_outline}
\keyword{internal} \alias{ii_social_tux}
\keyword{internal} \alias{ii_social_twitch}
\keyword{internal} \alias{ii_social_twitch_outline}
\keyword{internal} \alias{ii_social_twitter}
\keyword{internal} \alias{ii_social_twitter_outline}
\keyword{internal} \alias{ii_social_usd}
\keyword{internal} \alias{ii_social_usd_outline}
\keyword{internal} \alias{ii_social_vimeo}
\keyword{internal} \alias{ii_social_vimeo_outline}
\keyword{internal} \alias{ii_social_whatsapp}
\keyword{internal} \alias{ii_social_whatsapp_outline}
\keyword{internal} \alias{ii_social_windows}
\keyword{internal} \alias{ii_social_windows_outline}
\keyword{internal} \alias{ii_social_wordpress}
\keyword{internal} \alias{ii_social_wordpress_outline}
\keyword{internal} \alias{ii_social_yahoo}
\keyword{internal} \alias{ii_social_yahoo_outline}
\keyword{internal} \alias{ii_social_yen}
\keyword{internal} \alias{ii_social_yen_outline}
\keyword{internal} \alias{ii_social_youtube}
\keyword{internal} \alias{ii_social_youtube_outline}
\keyword{internal} \alias{ii_soup_can}
\keyword{internal} \alias{ii_soup_can_outline}
\keyword{internal} \alias{ii_speakerphone}
\keyword{internal} \alias{ii_speedometer}
\keyword{internal} \alias{ii_spoon}
\keyword{internal} \alias{ii_star}
\keyword{internal} \alias{ii_stats_bars}
\keyword{internal} \alias{ii_steam}
\keyword{internal} \alias{ii_stop}
\keyword{internal} \alias{ii_thermometer}
\keyword{internal} \alias{ii_thumbsdown}
\keyword{internal} \alias{ii_thumbsup}
\keyword{internal} \alias{ii_toggle}
\keyword{internal} \alias{ii_toggle_filled}
\keyword{internal} \alias{ii_transgender}
\keyword{internal} \alias{ii_trash_a}
\keyword{internal} \alias{ii_trash_b}
\keyword{internal} \alias{ii_trophy}
\keyword{internal} \alias{ii_tshirt}
\keyword{internal} \alias{ii_tshirt_outline}
\keyword{internal} \alias{ii_umbrella}
\keyword{internal} \alias{ii_university}
\keyword{internal} \alias{ii_unlocked}
\keyword{internal} \alias{ii_upload}
\keyword{internal} \alias{ii_usb}
\keyword{internal} \alias{ii_videocamera}
\keyword{internal} \alias{ii_volume_high}
\keyword{internal} \alias{ii_volume_low}
\keyword{internal} \alias{ii_volume_medium}
\keyword{internal} \alias{ii_volume_mute}
\keyword{internal} \alias{ii_wand}
\keyword{internal} \alias{ii_waterdrop}
\keyword{internal} \alias{ii_wifi}
\keyword{internal} \alias{ii_wineglass}
\keyword{internal} \alias{ii_woman}
\keyword{internal} \alias{ii_wrench}
\keyword{internal} \alias{ii_xbox}
