% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ii.R
\name{ii}
\alias{ii}
\title{Insert Icon From ionicons}
\usage{
ii(name = "ionic", size = 1, fixed_width = FALSE, animate = "still",
  rotate = 0, flip = "none", border = FALSE, pull = NULL,
  other = NULL)
}
\arguments{
\item{name}{The name of the icon}

\item{size}{Size of the icon relative to font size. Options are 1, lg (33%
increase), 2, 3, 4, or 5.}

\item{fixed_width}{If true, icons are set to a fixed width}

\item{animate}{Options are 'still', 'spin', or 'pulse'.}

\item{rotate}{Options are 0, 90, 180, or 270.}

\item{flip}{Options are 'none', 'horizontal', 'vertical'.}

\item{border}{If true, draws a border around the icon.}

\item{pull}{Pulls icon to either 'left' or 'right' and wraps proceeding text
around it.}

\item{other}{Character vector of other parameters directly added to the icon classes}
}
\description{
Insert Icon From ionicons
}
\alias{ii_alert}
\alias{ii_alert_circled}
\alias{ii_android_add}
\alias{ii_android_add_circle}
\alias{ii_android_alarm_clock}
\alias{ii_android_alert}
\alias{ii_android_apps}
\alias{ii_android_archive}
\alias{ii_android_arrow_back}
\alias{ii_android_arrow_down}
\alias{ii_android_arrow_dropdown}
\alias{ii_android_arrow_dropdown_circle}
\alias{ii_android_arrow_dropleft}
\alias{ii_android_arrow_dropleft_circle}
\alias{ii_android_arrow_dropright}
\alias{ii_android_arrow_dropright_circle}
\alias{ii_android_arrow_dropup}
\alias{ii_android_arrow_dropup_circle}
\alias{ii_android_arrow_forward}
\alias{ii_android_arrow_up}
\alias{ii_android_attach}
\alias{ii_android_bar}
\alias{ii_android_bicycle}
\alias{ii_android_boat}
\alias{ii_android_bookmark}
\alias{ii_android_bulb}
\alias{ii_android_bus}
\alias{ii_android_calendar}
\alias{ii_android_call}
\alias{ii_android_camera}
\alias{ii_android_cancel}
\alias{ii_android_car}
\alias{ii_android_cart}
\alias{ii_android_chat}
\alias{ii_android_checkbox}
\alias{ii_android_checkbox_blank}
\alias{ii_android_checkbox_outline}
\alias{ii_android_checkbox_outline_blank}
\alias{ii_android_checkmark_circle}
\alias{ii_android_clipboard}
\alias{ii_android_close}
\alias{ii_android_cloud}
\alias{ii_android_cloud_circle}
\alias{ii_android_cloud_done}
\alias{ii_android_cloud_outline}
\alias{ii_android_color_palette}
\alias{ii_android_compass}
\alias{ii_android_contact}
\alias{ii_android_contacts}
\alias{ii_android_contract}
\alias{ii_android_create}
\alias{ii_android_delete}
\alias{ii_android_desktop}
\alias{ii_android_document}
\alias{ii_android_done}
\alias{ii_android_done_all}
\alias{ii_android_download}
\alias{ii_android_drafts}
\alias{ii_android_exit}
\alias{ii_android_expand}
\alias{ii_android_favorite}
\alias{ii_android_favorite_outline}
\alias{ii_android_film}
\alias{ii_android_folder}
\alias{ii_android_folder_open}
\alias{ii_android_funnel}
\alias{ii_android_globe}
\alias{ii_android_hand}
\alias{ii_android_hangout}
\alias{ii_android_happy}
\alias{ii_android_home}
\alias{ii_android_image}
\alias{ii_android_laptop}
\alias{ii_android_list}
\alias{ii_android_locate}
\alias{ii_android_lock}
\alias{ii_android_mail}
\alias{ii_android_map}
\alias{ii_android_menu}
\alias{ii_android_microphone}
\alias{ii_android_microphone_off}
\alias{ii_android_more_horizontal}
\alias{ii_android_more_vertical}
\alias{ii_android_navigate}
\alias{ii_android_notifications}
\alias{ii_android_notifications_none}
\alias{ii_android_notifications_off}
\alias{ii_android_open}
\alias{ii_android_options}
\alias{ii_android_people}
\alias{ii_android_person}
\alias{ii_android_person_add}
\alias{ii_android_phone_landscape}
\alias{ii_android_phone_portrait}
\alias{ii_android_pin}
\alias{ii_android_plane}
\alias{ii_android_playstore}
\alias{ii_android_print}
\alias{ii_android_radio_button_off}
\alias{ii_android_radio_button_on}
\alias{ii_android_refresh}
\alias{ii_android_remove}
\alias{ii_android_remove_circle}
\alias{ii_android_restaurant}
\alias{ii_android_sad}
\alias{ii_android_search}
\alias{ii_android_send}
\alias{ii_android_settings}
\alias{ii_android_share}
\alias{ii_android_share_alt}
\alias{ii_android_star}
\alias{ii_android_star_half}
\alias{ii_android_star_outline}
\alias{ii_android_stopwatch}
\alias{ii_android_subway}
\alias{ii_android_sunny}
\alias{ii_android_sync}
\alias{ii_android_textsms}
\alias{ii_android_time}
\alias{ii_android_train}
\alias{ii_android_unlock}
\alias{ii_android_upload}
\alias{ii_android_volume_down}
\alias{ii_android_volume_mute}
\alias{ii_android_volume_off}
\alias{ii_android_volume_up}
\alias{ii_android_walk}
\alias{ii_android_warning}
\alias{ii_android_watch}
\alias{ii_android_wifi}
\alias{ii_aperture}
\alias{ii_archive}
\alias{ii_arrow_down_a}
\alias{ii_arrow_down_b}
\alias{ii_arrow_down_c}
\alias{ii_arrow_expand}
\alias{ii_arrow_graph_down_left}
\alias{ii_arrow_graph_down_right}
\alias{ii_arrow_graph_up_left}
\alias{ii_arrow_graph_up_right}
\alias{ii_arrow_left_a}
\alias{ii_arrow_left_b}
\alias{ii_arrow_left_c}
\alias{ii_arrow_move}
\alias{ii_arrow_resize}
\alias{ii_arrow_return_left}
\alias{ii_arrow_return_right}
\alias{ii_arrow_right_a}
\alias{ii_arrow_right_b}
\alias{ii_arrow_right_c}
\alias{ii_arrow_shrink}
\alias{ii_arrow_swap}
\alias{ii_arrow_up_a}
\alias{ii_arrow_up_b}
\alias{ii_arrow_up_c}
\alias{ii_asterisk}
\alias{ii_at}
\alias{ii_backspace}
\alias{ii_backspace_outline}
\alias{ii_bag}
\alias{ii_battery_charging}
\alias{ii_battery_empty}
\alias{ii_battery_full}
\alias{ii_battery_half}
\alias{ii_battery_low}
\alias{ii_beaker}
\alias{ii_beer}
\alias{ii_bluetooth}
\alias{ii_bonfire}
\alias{ii_bookmark}
\alias{ii_bowtie}
\alias{ii_briefcase}
\alias{ii_bug}
\alias{ii_calculator}
\alias{ii_calendar}
\alias{ii_camera}
\alias{ii_card}
\alias{ii_cash}
\alias{ii_chatbox}
\alias{ii_chatbox_working}
\alias{ii_chatboxes}
\alias{ii_chatbubble}
\alias{ii_chatbubble_working}
\alias{ii_chatbubbles}
\alias{ii_checkmark}
\alias{ii_checkmark_circled}
\alias{ii_checkmark_round}
\alias{ii_chevron_down}
\alias{ii_chevron_left}
\alias{ii_chevron_right}
\alias{ii_chevron_up}
\alias{ii_clipboard}
\alias{ii_clock}
\alias{ii_close}
\alias{ii_close_circled}
\alias{ii_close_round}
\alias{ii_closed_captioning}
\alias{ii_cloud}
\alias{ii_code}
\alias{ii_code_download}
\alias{ii_code_working}
\alias{ii_coffee}
\alias{ii_compass}
\alias{ii_compose}
\alias{ii_connection_bars}
\alias{ii_contrast}
\alias{ii_crop}
\alias{ii_cube}
\alias{ii_disc}
\alias{ii_document}
\alias{ii_document_text}
\alias{ii_drag}
\alias{ii_earth}
\alias{ii_easel}
\alias{ii_edit}
\alias{ii_egg}
\alias{ii_eject}
\alias{ii_email}
\alias{ii_email_unread}
\alias{ii_erlenmeyer_flask}
\alias{ii_erlenmeyer_flask_bubbles}
\alias{ii_eye}
\alias{ii_eye_disabled}
\alias{ii_female}
\alias{ii_filing}
\alias{ii_film_marker}
\alias{ii_fireball}
\alias{ii_flag}
\alias{ii_flame}
\alias{ii_flash}
\alias{ii_flash_off}
\alias{ii_folder}
\alias{ii_fork}
\alias{ii_fork_repo}
\alias{ii_forward}
\alias{ii_funnel}
\alias{ii_gear_a}
\alias{ii_gear_b}
\alias{ii_grid}
\alias{ii_hammer}
\alias{ii_happy}
\alias{ii_happy_outline}
\alias{ii_headphone}
\alias{ii_heart}
\alias{ii_heart_broken}
\alias{ii_help}
\alias{ii_help_buoy}
\alias{ii_help_circled}
\alias{ii_home}
\alias{ii_icecream}
\alias{ii_image}
\alias{ii_images}
\alias{ii_information}
\alias{ii_information_circled}
\alias{ii_ionic}
\alias{ii_ios_alarm}
\alias{ii_ios_alarm_outline}
\alias{ii_ios_albums}
\alias{ii_ios_albums_outline}
\alias{ii_ios_americanfootball}
\alias{ii_ios_americanfootball_outline}
\alias{ii_ios_analytics}
\alias{ii_ios_analytics_outline}
\alias{ii_ios_arrow_back}
\alias{ii_ios_arrow_down}
\alias{ii_ios_arrow_forward}
\alias{ii_ios_arrow_left}
\alias{ii_ios_arrow_right}
\alias{ii_ios_arrow_thin_down}
\alias{ii_ios_arrow_thin_left}
\alias{ii_ios_arrow_thin_right}
\alias{ii_ios_arrow_thin_up}
\alias{ii_ios_arrow_up}
\alias{ii_ios_at}
\alias{ii_ios_at_outline}
\alias{ii_ios_barcode}
\alias{ii_ios_barcode_outline}
\alias{ii_ios_baseball}
\alias{ii_ios_baseball_outline}
\alias{ii_ios_basketball}
\alias{ii_ios_basketball_outline}
\alias{ii_ios_bell}
\alias{ii_ios_bell_outline}
\alias{ii_ios_body}
\alias{ii_ios_body_outline}
\alias{ii_ios_bolt}
\alias{ii_ios_bolt_outline}
\alias{ii_ios_book}
\alias{ii_ios_book_outline}
\alias{ii_ios_bookmarks}
\alias{ii_ios_bookmarks_outline}
\alias{ii_ios_box}
\alias{ii_ios_box_outline}
\alias{ii_ios_briefcase}
\alias{ii_ios_briefcase_outline}
\alias{ii_ios_browsers}
\alias{ii_ios_browsers_outline}
\alias{ii_ios_calculator}
\alias{ii_ios_calculator_outline}
\alias{ii_ios_calendar}
\alias{ii_ios_calendar_outline}
\alias{ii_ios_camera}
\alias{ii_ios_camera_outline}
\alias{ii_ios_cart}
\alias{ii_ios_cart_outline}
\alias{ii_ios_chatboxes}
\alias{ii_ios_chatboxes_outline}
\alias{ii_ios_chatbubble}
\alias{ii_ios_chatbubble_outline}
\alias{ii_ios_checkmark}
\alias{ii_ios_checkmark_empty}
\alias{ii_ios_checkmark_outline}
\alias{ii_ios_circle_filled}
\alias{ii_ios_circle_outline}
\alias{ii_ios_clock}
\alias{ii_ios_clock_outline}
\alias{ii_ios_close}
\alias{ii_ios_close_empty}
\alias{ii_ios_close_outline}
\alias{ii_ios_cloud}
\alias{ii_ios_cloud_download}
\alias{ii_ios_cloud_download_outline}
\alias{ii_ios_cloud_outline}
\alias{ii_ios_cloud_upload}
\alias{ii_ios_cloud_upload_outline}
\alias{ii_ios_cloudy}
\alias{ii_ios_cloudy_night}
\alias{ii_ios_cloudy_night_outline}
\alias{ii_ios_cloudy_outline}
\alias{ii_ios_cog}
\alias{ii_ios_cog_outline}
\alias{ii_ios_color_filter}
\alias{ii_ios_color_filter_outline}
\alias{ii_ios_color_wand}
\alias{ii_ios_color_wand_outline}
\alias{ii_ios_compose}
\alias{ii_ios_compose_outline}
\alias{ii_ios_contact}
\alias{ii_ios_contact_outline}
\alias{ii_ios_copy}
\alias{ii_ios_copy_outline}
\alias{ii_ios_crop}
\alias{ii_ios_crop_strong}
\alias{ii_ios_download}
\alias{ii_ios_download_outline}
\alias{ii_ios_drag}
\alias{ii_ios_email}
\alias{ii_ios_email_outline}
\alias{ii_ios_eye}
\alias{ii_ios_eye_outline}
\alias{ii_ios_fastforward}
\alias{ii_ios_fastforward_outline}
\alias{ii_ios_filing}
\alias{ii_ios_filing_outline}
\alias{ii_ios_film}
\alias{ii_ios_film_outline}
\alias{ii_ios_flag}
\alias{ii_ios_flag_outline}
\alias{ii_ios_flame}
\alias{ii_ios_flame_outline}
\alias{ii_ios_flask}
\alias{ii_ios_flask_outline}
\alias{ii_ios_flower}
\alias{ii_ios_flower_outline}
\alias{ii_ios_folder}
\alias{ii_ios_folder_outline}
\alias{ii_ios_football}
\alias{ii_ios_football_outline}
\alias{ii_ios_game_controller_a}
\alias{ii_ios_game_controller_a_outline}
\alias{ii_ios_game_controller_b}
\alias{ii_ios_game_controller_b_outline}
\alias{ii_ios_gear}
\alias{ii_ios_gear_outline}
\alias{ii_ios_glasses}
\alias{ii_ios_glasses_outline}
\alias{ii_ios_grid_view}
\alias{ii_ios_grid_view_outline}
\alias{ii_ios_heart}
\alias{ii_ios_heart_outline}
\alias{ii_ios_help}
\alias{ii_ios_help_empty}
\alias{ii_ios_help_outline}
\alias{ii_ios_home}
\alias{ii_ios_home_outline}
\alias{ii_ios_infinite}
\alias{ii_ios_infinite_outline}
\alias{ii_ios_information}
\alias{ii_ios_information_empty}
\alias{ii_ios_information_outline}
\alias{ii_ios_ionic_outline}
\alias{ii_ios_keypad}
\alias{ii_ios_keypad_outline}
\alias{ii_ios_lightbulb}
\alias{ii_ios_lightbulb_outline}
\alias{ii_ios_list}
\alias{ii_ios_list_outline}
\alias{ii_ios_location}
\alias{ii_ios_location_outline}
\alias{ii_ios_locked}
\alias{ii_ios_locked_outline}
\alias{ii_ios_loop}
\alias{ii_ios_loop_strong}
\alias{ii_ios_medical}
\alias{ii_ios_medical_outline}
\alias{ii_ios_medkit}
\alias{ii_ios_medkit_outline}
\alias{ii_ios_mic}
\alias{ii_ios_mic_off}
\alias{ii_ios_mic_outline}
\alias{ii_ios_minus}
\alias{ii_ios_minus_empty}
\alias{ii_ios_minus_outline}
\alias{ii_ios_monitor}
\alias{ii_ios_monitor_outline}
\alias{ii_ios_moon}
\alias{ii_ios_moon_outline}
\alias{ii_ios_more}
\alias{ii_ios_more_outline}
\alias{ii_ios_musical_note}
\alias{ii_ios_musical_notes}
\alias{ii_ios_navigate}
\alias{ii_ios_navigate_outline}
\alias{ii_ios_nutrition}
\alias{ii_ios_nutrition_outline}
\alias{ii_ios_paper}
\alias{ii_ios_paper_outline}
\alias{ii_ios_paperplane}
\alias{ii_ios_paperplane_outline}
\alias{ii_ios_partlysunny}
\alias{ii_ios_partlysunny_outline}
\alias{ii_ios_pause}
\alias{ii_ios_pause_outline}
\alias{ii_ios_paw}
\alias{ii_ios_paw_outline}
\alias{ii_ios_people}
\alias{ii_ios_people_outline}
\alias{ii_ios_person}
\alias{ii_ios_person_outline}
\alias{ii_ios_personadd}
\alias{ii_ios_personadd_outline}
\alias{ii_ios_photos}
\alias{ii_ios_photos_outline}
\alias{ii_ios_pie}
\alias{ii_ios_pie_outline}
\alias{ii_ios_pint}
\alias{ii_ios_pint_outline}
\alias{ii_ios_play}
\alias{ii_ios_play_outline}
\alias{ii_ios_plus}
\alias{ii_ios_plus_empty}
\alias{ii_ios_plus_outline}
\alias{ii_ios_pricetag}
\alias{ii_ios_pricetag_outline}
\alias{ii_ios_pricetags}
\alias{ii_ios_pricetags_outline}
\alias{ii_ios_printer}
\alias{ii_ios_printer_outline}
\alias{ii_ios_pulse}
\alias{ii_ios_pulse_strong}
\alias{ii_ios_rainy}
\alias{ii_ios_rainy_outline}
\alias{ii_ios_recording}
\alias{ii_ios_recording_outline}
\alias{ii_ios_redo}
\alias{ii_ios_redo_outline}
\alias{ii_ios_refresh}
\alias{ii_ios_refresh_empty}
\alias{ii_ios_refresh_outline}
\alias{ii_ios_reload}
\alias{ii_ios_reverse_camera}
\alias{ii_ios_reverse_camera_outline}
\alias{ii_ios_rewind}
\alias{ii_ios_rewind_outline}
\alias{ii_ios_rose}
\alias{ii_ios_rose_outline}
\alias{ii_ios_search}
\alias{ii_ios_search_strong}
\alias{ii_ios_settings}
\alias{ii_ios_settings_strong}
\alias{ii_ios_shuffle}
\alias{ii_ios_shuffle_strong}
\alias{ii_ios_skipbackward}
\alias{ii_ios_skipbackward_outline}
\alias{ii_ios_skipforward}
\alias{ii_ios_skipforward_outline}
\alias{ii_ios_snowy}
\alias{ii_ios_speedometer}
\alias{ii_ios_speedometer_outline}
\alias{ii_ios_star}
\alias{ii_ios_star_half}
\alias{ii_ios_star_outline}
\alias{ii_ios_stopwatch}
\alias{ii_ios_stopwatch_outline}
\alias{ii_ios_sunny}
\alias{ii_ios_sunny_outline}
\alias{ii_ios_telephone}
\alias{ii_ios_telephone_outline}
\alias{ii_ios_tennisball}
\alias{ii_ios_tennisball_outline}
\alias{ii_ios_thunderstorm}
\alias{ii_ios_thunderstorm_outline}
\alias{ii_ios_time}
\alias{ii_ios_time_outline}
\alias{ii_ios_timer}
\alias{ii_ios_timer_outline}
\alias{ii_ios_toggle}
\alias{ii_ios_toggle_outline}
\alias{ii_ios_trash}
\alias{ii_ios_trash_outline}
\alias{ii_ios_undo}
\alias{ii_ios_undo_outline}
\alias{ii_ios_unlocked}
\alias{ii_ios_unlocked_outline}
\alias{ii_ios_upload}
\alias{ii_ios_upload_outline}
\alias{ii_ios_videocam}
\alias{ii_ios_videocam_outline}
\alias{ii_ios_volume_high}
\alias{ii_ios_volume_low}
\alias{ii_ios_wineglass}
\alias{ii_ios_wineglass_outline}
\alias{ii_ios_world}
\alias{ii_ios_world_outline}
\alias{ii_ipad}
\alias{ii_iphone}
\alias{ii_ipod}
\alias{ii_jet}
\alias{ii_key}
\alias{ii_knife}
\alias{ii_laptop}
\alias{ii_leaf}
\alias{ii_levels}
\alias{ii_lightbulb}
\alias{ii_link}
\alias{ii_load_a}
\alias{ii_load_b}
\alias{ii_load_c}
\alias{ii_load_d}
\alias{ii_location}
\alias{ii_lock_combination}
\alias{ii_locked}
\alias{ii_log_in}
\alias{ii_log_out}
\alias{ii_loop}
\alias{ii_magnet}
\alias{ii_male}
\alias{ii_man}
\alias{ii_map}
\alias{ii_medkit}
\alias{ii_merge}
\alias{ii_mic_a}
\alias{ii_mic_b}
\alias{ii_mic_c}
\alias{ii_minus}
\alias{ii_minus_circled}
\alias{ii_minus_round}
\alias{ii_model_s}
\alias{ii_monitor}
\alias{ii_more}
\alias{ii_mouse}
\alias{ii_music_note}
\alias{ii_navicon}
\alias{ii_navicon_round}
\alias{ii_navigate}
\alias{ii_network}
\alias{ii_no_smoking}
\alias{ii_nuclear}
\alias{ii_outlet}
\alias{ii_paintbrush}
\alias{ii_paintbucket}
\alias{ii_paper_airplane}
\alias{ii_paperclip}
\alias{ii_pause}
\alias{ii_person}
\alias{ii_person_add}
\alias{ii_person_stalker}
\alias{ii_pie_graph}
\alias{ii_pin}
\alias{ii_pinpoint}
\alias{ii_pizza}
\alias{ii_plane}
\alias{ii_planet}
\alias{ii_play}
\alias{ii_playstation}
\alias{ii_plus}
\alias{ii_plus_circled}
\alias{ii_plus_round}
\alias{ii_podium}
\alias{ii_pound}
\alias{ii_power}
\alias{ii_pricetag}
\alias{ii_pricetags}
\alias{ii_printer}
\alias{ii_pull_request}
\alias{ii_qr_scanner}
\alias{ii_quote}
\alias{ii_radio_waves}
\alias{ii_record}
\alias{ii_refresh}
\alias{ii_reply}
\alias{ii_reply_all}
\alias{ii_ribbon_a}
\alias{ii_ribbon_b}
\alias{ii_sad}
\alias{ii_sad_outline}
\alias{ii_scissors}
\alias{ii_search}
\alias{ii_settings}
\alias{ii_share}
\alias{ii_shuffle}
\alias{ii_skip_backward}
\alias{ii_skip_forward}
\alias{ii_social_android}
\alias{ii_social_android_outline}
\alias{ii_social_angular}
\alias{ii_social_angular_outline}
\alias{ii_social_apple}
\alias{ii_social_apple_outline}
\alias{ii_social_bitcoin}
\alias{ii_social_bitcoin_outline}
\alias{ii_social_buffer}
\alias{ii_social_buffer_outline}
\alias{ii_social_chrome}
\alias{ii_social_chrome_outline}
\alias{ii_social_codepen}
\alias{ii_social_codepen_outline}
\alias{ii_social_css3}
\alias{ii_social_css3_outline}
\alias{ii_social_designernews}
\alias{ii_social_designernews_outline}
\alias{ii_social_dribbble}
\alias{ii_social_dribbble_outline}
\alias{ii_social_dropbox}
\alias{ii_social_dropbox_outline}
\alias{ii_social_euro}
\alias{ii_social_euro_outline}
\alias{ii_social_facebook}
\alias{ii_social_facebook_outline}
\alias{ii_social_foursquare}
\alias{ii_social_foursquare_outline}
\alias{ii_social_freebsd_devil}
\alias{ii_social_github}
\alias{ii_social_github_outline}
\alias{ii_social_google}
\alias{ii_social_google_outline}
\alias{ii_social_googleplus}
\alias{ii_social_googleplus_outline}
\alias{ii_social_hackernews}
\alias{ii_social_hackernews_outline}
\alias{ii_social_html5}
\alias{ii_social_html5_outline}
\alias{ii_social_instagram}
\alias{ii_social_instagram_outline}
\alias{ii_social_javascript}
\alias{ii_social_javascript_outline}
\alias{ii_social_linkedin}
\alias{ii_social_linkedin_outline}
\alias{ii_social_markdown}
\alias{ii_social_nodejs}
\alias{ii_social_octocat}
\alias{ii_social_pinterest}
\alias{ii_social_pinterest_outline}
\alias{ii_social_python}
\alias{ii_social_reddit}
\alias{ii_social_reddit_outline}
\alias{ii_social_rss}
\alias{ii_social_rss_outline}
\alias{ii_social_sass}
\alias{ii_social_skype}
\alias{ii_social_skype_outline}
\alias{ii_social_snapchat}
\alias{ii_social_snapchat_outline}
\alias{ii_social_tumblr}
\alias{ii_social_tumblr_outline}
\alias{ii_social_tux}
\alias{ii_social_twitch}
\alias{ii_social_twitch_outline}
\alias{ii_social_twitter}
\alias{ii_social_twitter_outline}
\alias{ii_social_usd}
\alias{ii_social_usd_outline}
\alias{ii_social_vimeo}
\alias{ii_social_vimeo_outline}
\alias{ii_social_whatsapp}
\alias{ii_social_whatsapp_outline}
\alias{ii_social_windows}
\alias{ii_social_windows_outline}
\alias{ii_social_wordpress}
\alias{ii_social_wordpress_outline}
\alias{ii_social_yahoo}
\alias{ii_social_yahoo_outline}
\alias{ii_social_yen}
\alias{ii_social_yen_outline}
\alias{ii_social_youtube}
\alias{ii_social_youtube_outline}
\alias{ii_soup_can}
\alias{ii_soup_can_outline}
\alias{ii_speakerphone}
\alias{ii_speedometer}
\alias{ii_spoon}
\alias{ii_star}
\alias{ii_stats_bars}
\alias{ii_steam}
\alias{ii_stop}
\alias{ii_thermometer}
\alias{ii_thumbsdown}
\alias{ii_thumbsup}
\alias{ii_toggle}
\alias{ii_toggle_filled}
\alias{ii_transgender}
\alias{ii_trash_a}
\alias{ii_trash_b}
\alias{ii_trophy}
\alias{ii_tshirt}
\alias{ii_tshirt_outline}
\alias{ii_umbrella}
\alias{ii_university}
\alias{ii_unlocked}
\alias{ii_upload}
\alias{ii_usb}
\alias{ii_videocamera}
\alias{ii_volume_high}
\alias{ii_volume_low}
\alias{ii_volume_medium}
\alias{ii_volume_mute}
\alias{ii_wand}
\alias{ii_waterdrop}
\alias{ii_wifi}
\alias{ii_wineglass}
\alias{ii_woman}
\alias{ii_wrench}
\alias{ii_xbox}
