% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_design.R
\name{download_material_design}
\alias{download_material_design}
\alias{material_design}
\title{Material design icons}
\usage{
download_material_design(version = "dev")

material_design(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\description{
Material design icons
}
