% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{fa}
\alias{fa}
\alias{ii}
\alias{ai}
\title{Migration tools}
\usage{
fa(...)

ii(...)

ai(...)
}
\arguments{
\item{...}{Unused.}
}
\value{
These functions do not return; they always signal an error
directing you to the new interface.
}
\description{
The icon package has been overhauled to be more efficient and extensible.
Unfortunately this has led to several necessary interface changes. To
minimise impact, please use the \code{migrate_icon()} function to update your files
to use the new interface.
}
