% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfactors.R
\name{plot.icons_nfactors}
\alias{plot.icons_nfactors}
\title{Plot the factor-count criterion path}
\usage{
\method{plot}{icons_nfactors}(x, relative = FALSE, ...)
}
\arguments{
\item{x}{An \code{"icons_nfactors"} object.}

\item{relative}{Logical; plot the criterion relative to \eqn{\|S\|_F}.}

\item{...}{Passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Plot the factor-count criterion path
}
\examples{
data(sim)
nf <- n_factors(sim, icons_detect(cor(sim), threshold = 0.6))
plot(nf)

}
