% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMatrix}
\alias{plotMatrix}
\title{Visualize a matrix as a heatmap and save it to a file}
\usage{
plotMatrix(
  z,
  filepath = file.path(getwd(), "plot.tiff"),
  save.image = FALSE,
  width = 2500,
  height = 2350,
  res = 300,
  format = "tiff",
  colorbar.range = NULL,
  ...
)
}
\arguments{
\item{z}{A numeric matrix to be plotted. Must be numeric and two-dimensional.}

\item{filepath}{Character. The file path where the plot image will be saved if \code{save.image} is \code{TRUE}. Default is "plot.tiff" in the working directory.}

\item{save.image}{Logical. If \code{TRUE}, saves the plot to a file specified by \code{filepath}. Default is \code{FALSE}.}

\item{width}{Integer. Width of the saved image, in pixels (for raster formats) or inches (for vector formats). Default is 2500.}

\item{height}{Integer. Height of the saved image, in pixels (for raster formats) or inches (for vector formats). Default is 2350.}

\item{res}{Integer. Resolution of the saved image, in ppi (for raster formats only). Default is 300.}

\item{format}{Character. Format of the saved image. Options are \code{"tiff"}, \code{"png"}, \code{"jpeg"}, \code{"pdf"}, \code{"eps"}, or \code{"svg"}. Default is \code{"tiff"}.}

\item{colorbar.range}{Numeric vector of length 2. Specifies the minimum and maximum values for the color bar range. If \code{NULL} (default), the range is automatically determined from the input matrix \code{z}.}

\item{...}{Additional graphical parameters to pass to the \code{image} function, such as axis properties.}
}
\value{
The function does not return any values. If \code{save.image} is \code{TRUE}, a message is printed to the console indicating the file path where the image was saved.
}
\description{
Creates a heatmap visualization of a numeric matrix with a jet-like color palette.
}
\details{
This function plots a heatmap of a matrix with a color palette similar to the jet color map in MATLAB.
If \code{save.image} is \code{TRUE}, the plot is saved in the format and location specified by \code{filepath} and \code{format}.
The \code{colorbar.range} parameter allows users to manually set the range of the color bar, which is useful for ensuring consistent scaling across multiple plots.
}
\examples{
mat <- matrix(rnorm(100), nrow = 10)
plotMatrix(mat)
}
