% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmau.R
\name{sigmau}
\alias{sigmau}
\title{Compute SigmaU and its Norm}
\usage{
sigmau(data, cid, clist)
}
\arguments{
\item{data}{A data matrix.}

\item{cid}{A vector of cluster sizes.}

\item{clist}{A vector of cluster assignments.}
}
\value{
A list containing the SigmaU matrix (\code{sigma_u}) and its Frobenius norm (\code{sigma_u_norm}).
}
\description{
This function calculates the SigmaU matrix and its Frobenius norm for given data and clustering results.
}
\examples{
data <- matrix(runif(200), nrow = 20)
cid <- c(5, 5)
clist <- sample(1:2, 20, replace = TRUE)
result <- sigmau(data, cid, clist)
print(result)

}
