% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{squareform}
\alias{squareform}
\title{Convert a Squareform Matrix to a Vector}
\usage{
squareform(mat)
}
\arguments{
\item{mat}{A square-form matrix.}
}
\value{
A vector containing the lower triangular part of the matrix.
}
\description{
Given a squareform matrix, this function converts it into a vector containing
the lower triangular part of the matrix (excluding the diagonal).
}
\examples{
data(sim)
squareform(sim)

mat <- matrix(1:6, nrow = 3)
squareform(mat)
}
