% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{style}
\alias{style}
\title{Customise the style of an icon}
\usage{
style(x, scale = NULL, fill = NULL)
}
\arguments{
\item{scale}{Scaled size of the icon.}

\item{fill}{The colour for the fill.}
}
\description{
Customise the style of an icon
}
