% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetwork_visualize.R
\name{subnetwork_visualize}
\alias{subnetwork_visualize}
\title{Visualize a Matrix with ggplot2}
\usage{
subnetwork_visualize(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
A ggplot2 object representing the heatmap of the matrix.
}
\description{
This function takes a matrix as input and creates a heatmap to visualize it using ggplot2.
}
\examples{
mat <- matrix(rnorm(25), nrow = 5)
visualize_matrix(mat)
}
