% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academicons.R
\name{download_academicons}
\alias{download_academicons}
\alias{academicons}
\title{Academicons icons}
\usage{
download_academicons(version = "dev")

academicons(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\description{
Academicons icons
}
