% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioicons.R
\name{download_bioicons}
\alias{download_bioicons}
\alias{bioicons}
\title{Bioicons icons}
\usage{
download_bioicons()

bioicons(name, category = NULL)
}
\arguments{
\item{name}{Name of the icon.}

\item{category}{The icon's category, one of "Animals", "Blood_immunology",
"Cell_lines", "Cell_membrane", "Chemistry", "Chemo-_and_Bioinformatics",
"General_items", "Genetics", "Human_physiology", "Intracellular_components"
, "Lab_apparatus", "Machine_Learning", "Microbiology",
"Nucleic_Acids", "Oncology", "Parasites", "Receptor_channels",
"Safety_symbols", "Scientific_graphs", "Tissues", or "Viruses".
If \code{NULL}, it will default to the first available icon with this name.}
}
\value{
\code{download_bioicons()} invisibly returns the \code{bioicons} icon set
after downloading and installing its SVG files locally.

\code{bioicons(name, category)} returns an \code{icon} object (an SVG tag) for
the requested icon.
}
\description{
Bioicons is a free library of open source icons for scientific
illustrations using vector graphics software such as Inkscape or Adobe
Illustrator.
}
\seealso{
\url{https://bioicons.com/}
}
