% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioicons.R
\name{download_bioicons}
\alias{download_bioicons}
\alias{bioicons}
\title{Bioicons icons}
\usage{
download_bioicons()

bioicons(name)
}
\arguments{
\item{name}{Name of the icon.}
}
\description{
Bioicons is a free library of open source icons for scientific
illustrations using vector graphics software such as Inkscape or Adobe
Illustrator.
}
\seealso{
\url{https://bioicons.com/}
}
