% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioicons.R
\name{download_bioicons}
\alias{download_bioicons}
\alias{bioicons}
\title{Bioicons icons}
\usage{
download_bioicons()

bioicons(name, category = NULL)
}
\arguments{
\item{name}{Name of the icon.}

\item{category}{The icon's category, one of "Animals", "Blood  immunology",
"Cell lines", "Cell membrane", "Chemistry", "Chemo- and Bioinformatics",
"General items", "Genetics", "Human physiology", "Intracellular
components", "Lab apparatus", "Machine Learning", "Microbiology",
"Nucleic Acids", "Oncology", "Parasites", "Receptor channels", "Safety
symbols", "Scientific graphs", "Tissues", or "Viruses". \code{category} is not
case-sensitive. If \code{NULL}, it will default to the first
available icon with this name.}
}
\description{
Bioicons is a free library of open source icons for scientific
illustrations using vector graphics software such as Inkscape or Adobe
Illustrator.
}
\seealso{
\url{https://bioicons.com/}
}
