% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{fa-alias}
\alias{fa-alias}
\alias{fa-alias}
\title{Font awesome alias}
\description{
Font awesome alias
}
\keyword{internal} \alias{fa_glass}
\keyword{internal} \alias{fa_music}
\keyword{internal} \alias{fa_search}
\keyword{internal} \alias{fa_envelope_o}
\keyword{internal} \alias{fa_heart}
\keyword{internal} \alias{fa_star}
\keyword{internal} \alias{fa_star_o}
\keyword{internal} \alias{fa_user}
\keyword{internal} \alias{fa_film}
\keyword{internal} \alias{fa_th_large}
\keyword{internal} \alias{fa_th}
\keyword{internal} \alias{fa_th_list}
\keyword{internal} \alias{fa_check}
\keyword{internal} \alias{fa_times}
\keyword{internal} \alias{fa_search_plus}
\keyword{internal} \alias{fa_search_minus}
\keyword{internal} \alias{fa_power_off}
\keyword{internal} \alias{fa_signal}
\keyword{internal} \alias{fa_cog}
\keyword{internal} \alias{fa_trash_o}
\keyword{internal} \alias{fa_home}
\keyword{internal} \alias{fa_file_o}
\keyword{internal} \alias{fa_clock_o}
\keyword{internal} \alias{fa_road}
\keyword{internal} \alias{fa_download}
\keyword{internal} \alias{fa_arrow_circle_o_down}
\keyword{internal} \alias{fa_arrow_circle_o_up}
\keyword{internal} \alias{fa_inbox}
\keyword{internal} \alias{fa_play_circle_o}
\keyword{internal} \alias{fa_repeat}
\keyword{internal} \alias{fa_refresh}
\keyword{internal} \alias{fa_list_alt}
\keyword{internal} \alias{fa_lock}
\keyword{internal} \alias{fa_flag}
\keyword{internal} \alias{fa_headphones}
\keyword{internal} \alias{fa_volume_off}
\keyword{internal} \alias{fa_volume_down}
\keyword{internal} \alias{fa_volume_up}
\keyword{internal} \alias{fa_qrcode}
\keyword{internal} \alias{fa_barcode}
\keyword{internal} \alias{fa_tag}
\keyword{internal} \alias{fa_tags}
\keyword{internal} \alias{fa_book}
\keyword{internal} \alias{fa_bookmark}
\keyword{internal} \alias{fa_print}
\keyword{internal} \alias{fa_camera}
\keyword{internal} \alias{fa_font}
\keyword{internal} \alias{fa_bold}
\keyword{internal} \alias{fa_italic}
\keyword{internal} \alias{fa_text_height}
\keyword{internal} \alias{fa_text_width}
\keyword{internal} \alias{fa_align_left}
\keyword{internal} \alias{fa_align_center}
\keyword{internal} \alias{fa_align_right}
\keyword{internal} \alias{fa_align_justify}
\keyword{internal} \alias{fa_list}
\keyword{internal} \alias{fa_outdent}
\keyword{internal} \alias{fa_indent}
\keyword{internal} \alias{fa_video_camera}
\keyword{internal} \alias{fa_picture_o}
\keyword{internal} \alias{fa_pencil}
\keyword{internal} \alias{fa_map_marker}
\keyword{internal} \alias{fa_adjust}
\keyword{internal} \alias{fa_tint}
\keyword{internal} \alias{fa_pencil_square_o}
\keyword{internal} \alias{fa_share_square_o}
\keyword{internal} \alias{fa_check_square_o}
\keyword{internal} \alias{fa_arrows}
\keyword{internal} \alias{fa_step_backward}
\keyword{internal} \alias{fa_fast_backward}
\keyword{internal} \alias{fa_backward}
\keyword{internal} \alias{fa_play}
\keyword{internal} \alias{fa_pause}
\keyword{internal} \alias{fa_stop}
\keyword{internal} \alias{fa_forward}
\keyword{internal} \alias{fa_fast_forward}
\keyword{internal} \alias{fa_step_forward}
\keyword{internal} \alias{fa_eject}
\keyword{internal} \alias{fa_chevron_left}
\keyword{internal} \alias{fa_chevron_right}
\keyword{internal} \alias{fa_plus_circle}
\keyword{internal} \alias{fa_minus_circle}
\keyword{internal} \alias{fa_times_circle}
\keyword{internal} \alias{fa_check_circle}
\keyword{internal} \alias{fa_question_circle}
\keyword{internal} \alias{fa_info_circle}
\keyword{internal} \alias{fa_crosshairs}
\keyword{internal} \alias{fa_times_circle_o}
\keyword{internal} \alias{fa_check_circle_o}
\keyword{internal} \alias{fa_ban}
\keyword{internal} \alias{fa_arrow_left}
\keyword{internal} \alias{fa_arrow_right}
\keyword{internal} \alias{fa_arrow_up}
\keyword{internal} \alias{fa_arrow_down}
\keyword{internal} \alias{fa_share}
\keyword{internal} \alias{fa_expand}
\keyword{internal} \alias{fa_compress}
\keyword{internal} \alias{fa_plus}
\keyword{internal} \alias{fa_minus}
\keyword{internal} \alias{fa_asterisk}
\keyword{internal} \alias{fa_exclamation_circle}
\keyword{internal} \alias{fa_gift}
\keyword{internal} \alias{fa_leaf}
\keyword{internal} \alias{fa_fire}
\keyword{internal} \alias{fa_eye}
\keyword{internal} \alias{fa_eye_slash}
\keyword{internal} \alias{fa_exclamation_triangle}
\keyword{internal} \alias{fa_plane}
\keyword{internal} \alias{fa_calendar}
\keyword{internal} \alias{fa_random}
\keyword{internal} \alias{fa_comment}
\keyword{internal} \alias{fa_magnet}
\keyword{internal} \alias{fa_chevron_up}
\keyword{internal} \alias{fa_chevron_down}
\keyword{internal} \alias{fa_retweet}
\keyword{internal} \alias{fa_shopping_cart}
\keyword{internal} \alias{fa_folder}
\keyword{internal} \alias{fa_folder_open}
\keyword{internal} \alias{fa_arrows_v}
\keyword{internal} \alias{fa_arrows_h}
\keyword{internal} \alias{fa_bar_chart}
\keyword{internal} \alias{fa_twitter_square}
\keyword{internal} \alias{fa_facebook_square}
\keyword{internal} \alias{fa_camera_retro}
\keyword{internal} \alias{fa_key}
\keyword{internal} \alias{fa_cogs}
\keyword{internal} \alias{fa_comments}
\keyword{internal} \alias{fa_thumbs_o_up}
\keyword{internal} \alias{fa_thumbs_o_down}
\keyword{internal} \alias{fa_star_half}
\keyword{internal} \alias{fa_heart_o}
\keyword{internal} \alias{fa_sign_out}
\keyword{internal} \alias{fa_linkedin_square}
\keyword{internal} \alias{fa_thumb_tack}
\keyword{internal} \alias{fa_external_link}
\keyword{internal} \alias{fa_sign_in}
\keyword{internal} \alias{fa_trophy}
\keyword{internal} \alias{fa_github_square}
\keyword{internal} \alias{fa_upload}
\keyword{internal} \alias{fa_lemon_o}
\keyword{internal} \alias{fa_phone}
\keyword{internal} \alias{fa_square_o}
\keyword{internal} \alias{fa_bookmark_o}
\keyword{internal} \alias{fa_phone_square}
\keyword{internal} \alias{fa_twitter}
\keyword{internal} \alias{fa_facebook}
\keyword{internal} \alias{fa_github}
\keyword{internal} \alias{fa_unlock}
\keyword{internal} \alias{fa_credit_card}
\keyword{internal} \alias{fa_rss}
\keyword{internal} \alias{fa_hdd_o}
\keyword{internal} \alias{fa_bullhorn}
\keyword{internal} \alias{fa_bell}
\keyword{internal} \alias{fa_certificate}
\keyword{internal} \alias{fa_hand_o_right}
\keyword{internal} \alias{fa_hand_o_left}
\keyword{internal} \alias{fa_hand_o_up}
\keyword{internal} \alias{fa_hand_o_down}
\keyword{internal} \alias{fa_arrow_circle_left}
\keyword{internal} \alias{fa_arrow_circle_right}
\keyword{internal} \alias{fa_arrow_circle_up}
\keyword{internal} \alias{fa_arrow_circle_down}
\keyword{internal} \alias{fa_globe}
\keyword{internal} \alias{fa_wrench}
\keyword{internal} \alias{fa_tasks}
\keyword{internal} \alias{fa_filter}
\keyword{internal} \alias{fa_briefcase}
\keyword{internal} \alias{fa_arrows_alt}
\keyword{internal} \alias{fa_users}
\keyword{internal} \alias{fa_link}
\keyword{internal} \alias{fa_cloud}
\keyword{internal} \alias{fa_flask}
\keyword{internal} \alias{fa_scissors}
\keyword{internal} \alias{fa_files_o}
\keyword{internal} \alias{fa_paperclip}
\keyword{internal} \alias{fa_floppy_o}
\keyword{internal} \alias{fa_square}
\keyword{internal} \alias{fa_bars}
\keyword{internal} \alias{fa_list_ul}
\keyword{internal} \alias{fa_list_ol}
\keyword{internal} \alias{fa_strikethrough}
\keyword{internal} \alias{fa_underline}
\keyword{internal} \alias{fa_table}
\keyword{internal} \alias{fa_magic}
\keyword{internal} \alias{fa_truck}
\keyword{internal} \alias{fa_pinterest}
\keyword{internal} \alias{fa_pinterest_square}
\keyword{internal} \alias{fa_google_plus_square}
\keyword{internal} \alias{fa_google_plus}
\keyword{internal} \alias{fa_money}
\keyword{internal} \alias{fa_caret_down}
\keyword{internal} \alias{fa_caret_up}
\keyword{internal} \alias{fa_caret_left}
\keyword{internal} \alias{fa_caret_right}
\keyword{internal} \alias{fa_columns}
\keyword{internal} \alias{fa_sort}
\keyword{internal} \alias{fa_sort_desc}
\keyword{internal} \alias{fa_sort_asc}
\keyword{internal} \alias{fa_envelope}
\keyword{internal} \alias{fa_linkedin}
\keyword{internal} \alias{fa_undo}
\keyword{internal} \alias{fa_gavel}
\keyword{internal} \alias{fa_tachometer}
\keyword{internal} \alias{fa_comment_o}
\keyword{internal} \alias{fa_comments_o}
\keyword{internal} \alias{fa_bolt}
\keyword{internal} \alias{fa_sitemap}
\keyword{internal} \alias{fa_umbrella}
\keyword{internal} \alias{fa_clipboard}
\keyword{internal} \alias{fa_lightbulb_o}
\keyword{internal} \alias{fa_exchange}
\keyword{internal} \alias{fa_cloud_download}
\keyword{internal} \alias{fa_cloud_upload}
\keyword{internal} \alias{fa_user_md}
\keyword{internal} \alias{fa_stethoscope}
\keyword{internal} \alias{fa_suitcase}
\keyword{internal} \alias{fa_bell_o}
\keyword{internal} \alias{fa_coffee}
\keyword{internal} \alias{fa_cutlery}
\keyword{internal} \alias{fa_file_text_o}
\keyword{internal} \alias{fa_building_o}
\keyword{internal} \alias{fa_hospital_o}
\keyword{internal} \alias{fa_ambulance}
\keyword{internal} \alias{fa_medkit}
\keyword{internal} \alias{fa_fighter_jet}
\keyword{internal} \alias{fa_beer}
\keyword{internal} \alias{fa_h_square}
\keyword{internal} \alias{fa_plus_square}
\keyword{internal} \alias{fa_angle_double_left}
\keyword{internal} \alias{fa_angle_double_right}
\keyword{internal} \alias{fa_angle_double_up}
\keyword{internal} \alias{fa_angle_double_down}
\keyword{internal} \alias{fa_angle_left}
\keyword{internal} \alias{fa_angle_right}
\keyword{internal} \alias{fa_angle_up}
\keyword{internal} \alias{fa_angle_down}
\keyword{internal} \alias{fa_desktop}
\keyword{internal} \alias{fa_laptop}
\keyword{internal} \alias{fa_tablet}
\keyword{internal} \alias{fa_mobile}
\keyword{internal} \alias{fa_circle_o}
\keyword{internal} \alias{fa_quote_left}
\keyword{internal} \alias{fa_quote_right}
\keyword{internal} \alias{fa_spinner}
\keyword{internal} \alias{fa_circle}
\keyword{internal} \alias{fa_reply}
\keyword{internal} \alias{fa_github_alt}
\keyword{internal} \alias{fa_folder_o}
\keyword{internal} \alias{fa_folder_open_o}
\keyword{internal} \alias{fa_smile_o}
\keyword{internal} \alias{fa_frown_o}
\keyword{internal} \alias{fa_meh_o}
\keyword{internal} \alias{fa_gamepad}
\keyword{internal} \alias{fa_keyboard_o}
\keyword{internal} \alias{fa_flag_o}
\keyword{internal} \alias{fa_flag_checkered}
\keyword{internal} \alias{fa_terminal}
\keyword{internal} \alias{fa_code}
\keyword{internal} \alias{fa_reply_all}
\keyword{internal} \alias{fa_star_half_o}
\keyword{internal} \alias{fa_location_arrow}
\keyword{internal} \alias{fa_crop}
\keyword{internal} \alias{fa_code_fork}
\keyword{internal} \alias{fa_chain_broken}
\keyword{internal} \alias{fa_question}
\keyword{internal} \alias{fa_info}
\keyword{internal} \alias{fa_exclamation}
\keyword{internal} \alias{fa_superscript}
\keyword{internal} \alias{fa_subscript}
\keyword{internal} \alias{fa_eraser}
\keyword{internal} \alias{fa_puzzle_piece}
\keyword{internal} \alias{fa_microphone}
\keyword{internal} \alias{fa_microphone_slash}
\keyword{internal} \alias{fa_shield}
\keyword{internal} \alias{fa_calendar_o}
\keyword{internal} \alias{fa_fire_extinguisher}
\keyword{internal} \alias{fa_rocket}
\keyword{internal} \alias{fa_maxcdn}
\keyword{internal} \alias{fa_chevron_circle_left}
\keyword{internal} \alias{fa_chevron_circle_right}
\keyword{internal} \alias{fa_chevron_circle_up}
\keyword{internal} \alias{fa_chevron_circle_down}
\keyword{internal} \alias{fa_html5}
\keyword{internal} \alias{fa_css3}
\keyword{internal} \alias{fa_anchor}
\keyword{internal} \alias{fa_unlock_alt}
\keyword{internal} \alias{fa_bullseye}
\keyword{internal} \alias{fa_ellipsis_h}
\keyword{internal} \alias{fa_ellipsis_v}
\keyword{internal} \alias{fa_rss_square}
\keyword{internal} \alias{fa_play_circle}
\keyword{internal} \alias{fa_ticket}
\keyword{internal} \alias{fa_minus_square}
\keyword{internal} \alias{fa_minus_square_o}
\keyword{internal} \alias{fa_level_up}
\keyword{internal} \alias{fa_level_down}
\keyword{internal} \alias{fa_check_square}
\keyword{internal} \alias{fa_pencil_square}
\keyword{internal} \alias{fa_external_link_square}
\keyword{internal} \alias{fa_share_square}
\keyword{internal} \alias{fa_compass}
\keyword{internal} \alias{fa_caret_square_o_down}
\keyword{internal} \alias{fa_caret_square_o_up}
\keyword{internal} \alias{fa_caret_square_o_right}
\keyword{internal} \alias{fa_eur}
\keyword{internal} \alias{fa_gbp}
\keyword{internal} \alias{fa_usd}
\keyword{internal} \alias{fa_inr}
\keyword{internal} \alias{fa_jpy}
\keyword{internal} \alias{fa_rub}
\keyword{internal} \alias{fa_krw}
\keyword{internal} \alias{fa_btc}
\keyword{internal} \alias{fa_file}
\keyword{internal} \alias{fa_file_text}
\keyword{internal} \alias{fa_sort_alpha_asc}
\keyword{internal} \alias{fa_sort_alpha_desc}
\keyword{internal} \alias{fa_sort_amount_asc}
\keyword{internal} \alias{fa_sort_amount_desc}
\keyword{internal} \alias{fa_sort_numeric_asc}
\keyword{internal} \alias{fa_sort_numeric_desc}
\keyword{internal} \alias{fa_thumbs_up}
\keyword{internal} \alias{fa_thumbs_down}
\keyword{internal} \alias{fa_youtube_square}
\keyword{internal} \alias{fa_youtube}
\keyword{internal} \alias{fa_xing}
\keyword{internal} \alias{fa_xing_square}
\keyword{internal} \alias{fa_youtube_play}
\keyword{internal} \alias{fa_dropbox}
\keyword{internal} \alias{fa_stack_overflow}
\keyword{internal} \alias{fa_instagram}
\keyword{internal} \alias{fa_flickr}
\keyword{internal} \alias{fa_adn}
\keyword{internal} \alias{fa_bitbucket}
\keyword{internal} \alias{fa_bitbucket_square}
\keyword{internal} \alias{fa_tumblr}
\keyword{internal} \alias{fa_tumblr_square}
\keyword{internal} \alias{fa_long_arrow_down}
\keyword{internal} \alias{fa_long_arrow_up}
\keyword{internal} \alias{fa_long_arrow_left}
\keyword{internal} \alias{fa_long_arrow_right}
\keyword{internal} \alias{fa_apple}
\keyword{internal} \alias{fa_windows}
\keyword{internal} \alias{fa_android}
\keyword{internal} \alias{fa_linux}
\keyword{internal} \alias{fa_dribbble}
\keyword{internal} \alias{fa_skype}
\keyword{internal} \alias{fa_foursquare}
\keyword{internal} \alias{fa_trello}
\keyword{internal} \alias{fa_female}
\keyword{internal} \alias{fa_male}
\keyword{internal} \alias{fa_gratipay}
\keyword{internal} \alias{fa_sun_o}
\keyword{internal} \alias{fa_moon_o}
\keyword{internal} \alias{fa_archive}
\keyword{internal} \alias{fa_bug}
\keyword{internal} \alias{fa_vk}
\keyword{internal} \alias{fa_weibo}
\keyword{internal} \alias{fa_renren}
\keyword{internal} \alias{fa_pagelines}
\keyword{internal} \alias{fa_stack_exchange}
\keyword{internal} \alias{fa_arrow_circle_o_right}
\keyword{internal} \alias{fa_arrow_circle_o_left}
\keyword{internal} \alias{fa_caret_square_o_left}
\keyword{internal} \alias{fa_dot_circle_o}
\keyword{internal} \alias{fa_wheelchair}
\keyword{internal} \alias{fa_vimeo_square}
\keyword{internal} \alias{fa_try}
\keyword{internal} \alias{fa_plus_square_o}
\keyword{internal} \alias{fa_space_shuttle}
\keyword{internal} \alias{fa_slack}
\keyword{internal} \alias{fa_envelope_square}
\keyword{internal} \alias{fa_wordpress}
\keyword{internal} \alias{fa_openid}
\keyword{internal} \alias{fa_university}
\keyword{internal} \alias{fa_graduation_cap}
\keyword{internal} \alias{fa_yahoo}
\keyword{internal} \alias{fa_google}
\keyword{internal} \alias{fa_reddit}
\keyword{internal} \alias{fa_reddit_square}
\keyword{internal} \alias{fa_stumbleupon_circle}
\keyword{internal} \alias{fa_stumbleupon}
\keyword{internal} \alias{fa_delicious}
\keyword{internal} \alias{fa_digg}
\keyword{internal} \alias{fa_pied_piper_pp}
\keyword{internal} \alias{fa_pied_piper_alt}
\keyword{internal} \alias{fa_drupal}
\keyword{internal} \alias{fa_joomla}
\keyword{internal} \alias{fa_language}
\keyword{internal} \alias{fa_fax}
\keyword{internal} \alias{fa_building}
\keyword{internal} \alias{fa_child}
\keyword{internal} \alias{fa_paw}
\keyword{internal} \alias{fa_spoon}
\keyword{internal} \alias{fa_cube}
\keyword{internal} \alias{fa_cubes}
\keyword{internal} \alias{fa_behance}
\keyword{internal} \alias{fa_behance_square}
\keyword{internal} \alias{fa_steam}
\keyword{internal} \alias{fa_steam_square}
\keyword{internal} \alias{fa_recycle}
\keyword{internal} \alias{fa_car}
\keyword{internal} \alias{fa_taxi}
\keyword{internal} \alias{fa_tree}
\keyword{internal} \alias{fa_spotify}
\keyword{internal} \alias{fa_deviantart}
\keyword{internal} \alias{fa_soundcloud}
\keyword{internal} \alias{fa_database}
\keyword{internal} \alias{fa_file_pdf_o}
\keyword{internal} \alias{fa_file_word_o}
\keyword{internal} \alias{fa_file_excel_o}
\keyword{internal} \alias{fa_file_powerpoint_o}
\keyword{internal} \alias{fa_file_image_o}
\keyword{internal} \alias{fa_file_archive_o}
\keyword{internal} \alias{fa_file_audio_o}
\keyword{internal} \alias{fa_file_video_o}
\keyword{internal} \alias{fa_file_code_o}
\keyword{internal} \alias{fa_vine}
\keyword{internal} \alias{fa_codepen}
\keyword{internal} \alias{fa_jsfiddle}
\keyword{internal} \alias{fa_life_ring}
\keyword{internal} \alias{fa_circle_o_notch}
\keyword{internal} \alias{fa_rebel}
\keyword{internal} \alias{fa_empire}
\keyword{internal} \alias{fa_git_square}
\keyword{internal} \alias{fa_git}
\keyword{internal} \alias{fa_hacker_news}
\keyword{internal} \alias{fa_tencent_weibo}
\keyword{internal} \alias{fa_qq}
\keyword{internal} \alias{fa_weixin}
\keyword{internal} \alias{fa_paper_plane}
\keyword{internal} \alias{fa_paper_plane_o}
\keyword{internal} \alias{fa_history}
\keyword{internal} \alias{fa_circle_thin}
\keyword{internal} \alias{fa_header}
\keyword{internal} \alias{fa_paragraph}
\keyword{internal} \alias{fa_sliders}
\keyword{internal} \alias{fa_share_alt}
\keyword{internal} \alias{fa_share_alt_square}
\keyword{internal} \alias{fa_bomb}
\keyword{internal} \alias{fa_futbol_o}
\keyword{internal} \alias{fa_tty}
\keyword{internal} \alias{fa_binoculars}
\keyword{internal} \alias{fa_plug}
\keyword{internal} \alias{fa_slideshare}
\keyword{internal} \alias{fa_twitch}
\keyword{internal} \alias{fa_yelp}
\keyword{internal} \alias{fa_newspaper_o}
\keyword{internal} \alias{fa_wifi}
\keyword{internal} \alias{fa_calculator}
\keyword{internal} \alias{fa_paypal}
\keyword{internal} \alias{fa_google_wallet}
\keyword{internal} \alias{fa_cc_visa}
\keyword{internal} \alias{fa_cc_mastercard}
\keyword{internal} \alias{fa_cc_discover}
\keyword{internal} \alias{fa_cc_amex}
\keyword{internal} \alias{fa_cc_paypal}
\keyword{internal} \alias{fa_cc_stripe}
\keyword{internal} \alias{fa_bell_slash}
\keyword{internal} \alias{fa_bell_slash_o}
\keyword{internal} \alias{fa_trash}
\keyword{internal} \alias{fa_copyright}
\keyword{internal} \alias{fa_at}
\keyword{internal} \alias{fa_eyedropper}
\keyword{internal} \alias{fa_paint_brush}
\keyword{internal} \alias{fa_birthday_cake}
\keyword{internal} \alias{fa_area_chart}
\keyword{internal} \alias{fa_pie_chart}
\keyword{internal} \alias{fa_line_chart}
\keyword{internal} \alias{fa_lastfm}
\keyword{internal} \alias{fa_lastfm_square}
\keyword{internal} \alias{fa_toggle_off}
\keyword{internal} \alias{fa_toggle_on}
\keyword{internal} \alias{fa_bicycle}
\keyword{internal} \alias{fa_bus}
\keyword{internal} \alias{fa_ioxhost}
\keyword{internal} \alias{fa_angellist}
\keyword{internal} \alias{fa_cc}
\keyword{internal} \alias{fa_ils}
\keyword{internal} \alias{fa_meanpath}
\keyword{internal} \alias{fa_buysellads}
\keyword{internal} \alias{fa_connectdevelop}
\keyword{internal} \alias{fa_dashcube}
\keyword{internal} \alias{fa_forumbee}
\keyword{internal} \alias{fa_leanpub}
\keyword{internal} \alias{fa_sellsy}
\keyword{internal} \alias{fa_shirtsinbulk}
\keyword{internal} \alias{fa_simplybuilt}
\keyword{internal} \alias{fa_skyatlas}
\keyword{internal} \alias{fa_cart_plus}
\keyword{internal} \alias{fa_cart_arrow_down}
\keyword{internal} \alias{fa_diamond}
\keyword{internal} \alias{fa_ship}
\keyword{internal} \alias{fa_user_secret}
\keyword{internal} \alias{fa_motorcycle}
\keyword{internal} \alias{fa_street_view}
\keyword{internal} \alias{fa_heartbeat}
\keyword{internal} \alias{fa_venus}
\keyword{internal} \alias{fa_mars}
\keyword{internal} \alias{fa_mercury}
\keyword{internal} \alias{fa_transgender}
\keyword{internal} \alias{fa_transgender_alt}
\keyword{internal} \alias{fa_venus_double}
\keyword{internal} \alias{fa_mars_double}
\keyword{internal} \alias{fa_venus_mars}
\keyword{internal} \alias{fa_mars_stroke}
\keyword{internal} \alias{fa_mars_stroke_v}
\keyword{internal} \alias{fa_mars_stroke_h}
\keyword{internal} \alias{fa_neuter}
\keyword{internal} \alias{fa_genderless}
\keyword{internal} \alias{fa_facebook_official}
\keyword{internal} \alias{fa_pinterest_p}
\keyword{internal} \alias{fa_whatsapp}
\keyword{internal} \alias{fa_server}
\keyword{internal} \alias{fa_user_plus}
\keyword{internal} \alias{fa_user_times}
\keyword{internal} \alias{fa_bed}
\keyword{internal} \alias{fa_viacoin}
\keyword{internal} \alias{fa_train}
\keyword{internal} \alias{fa_subway}
\keyword{internal} \alias{fa_medium}
\keyword{internal} \alias{fa_y_combinator}
\keyword{internal} \alias{fa_optin_monster}
\keyword{internal} \alias{fa_opencart}
\keyword{internal} \alias{fa_expeditedssl}
\keyword{internal} \alias{fa_battery_full}
\keyword{internal} \alias{fa_battery_three_quarters}
\keyword{internal} \alias{fa_battery_half}
\keyword{internal} \alias{fa_battery_quarter}
\keyword{internal} \alias{fa_battery_empty}
\keyword{internal} \alias{fa_mouse_pointer}
\keyword{internal} \alias{fa_i_cursor}
\keyword{internal} \alias{fa_object_group}
\keyword{internal} \alias{fa_object_ungroup}
\keyword{internal} \alias{fa_sticky_note}
\keyword{internal} \alias{fa_sticky_note_o}
\keyword{internal} \alias{fa_cc_jcb}
\keyword{internal} \alias{fa_cc_diners_club}
\keyword{internal} \alias{fa_clone}
\keyword{internal} \alias{fa_balance_scale}
\keyword{internal} \alias{fa_hourglass_o}
\keyword{internal} \alias{fa_hourglass_start}
\keyword{internal} \alias{fa_hourglass_half}
\keyword{internal} \alias{fa_hourglass_end}
\keyword{internal} \alias{fa_hourglass}
\keyword{internal} \alias{fa_hand_rock_o}
\keyword{internal} \alias{fa_hand_paper_o}
\keyword{internal} \alias{fa_hand_scissors_o}
\keyword{internal} \alias{fa_hand_lizard_o}
\keyword{internal} \alias{fa_hand_spock_o}
\keyword{internal} \alias{fa_hand_pointer_o}
\keyword{internal} \alias{fa_hand_peace_o}
\keyword{internal} \alias{fa_trademark}
\keyword{internal} \alias{fa_registered}
\keyword{internal} \alias{fa_creative_commons}
\keyword{internal} \alias{fa_gg}
\keyword{internal} \alias{fa_gg_circle}
\keyword{internal} \alias{fa_tripadvisor}
\keyword{internal} \alias{fa_odnoklassniki}
\keyword{internal} \alias{fa_odnoklassniki_square}
\keyword{internal} \alias{fa_get_pocket}
\keyword{internal} \alias{fa_wikipedia_w}
\keyword{internal} \alias{fa_safari}
\keyword{internal} \alias{fa_chrome}
\keyword{internal} \alias{fa_firefox}
\keyword{internal} \alias{fa_opera}
\keyword{internal} \alias{fa_internet_explorer}
\keyword{internal} \alias{fa_television}
\keyword{internal} \alias{fa_contao}
\keyword{internal} \alias{fa_500px}
\keyword{internal} \alias{fa_amazon}
\keyword{internal} \alias{fa_calendar_plus_o}
\keyword{internal} \alias{fa_calendar_minus_o}
\keyword{internal} \alias{fa_calendar_times_o}
\keyword{internal} \alias{fa_calendar_check_o}
\keyword{internal} \alias{fa_industry}
\keyword{internal} \alias{fa_map_pin}
\keyword{internal} \alias{fa_map_signs}
\keyword{internal} \alias{fa_map_o}
\keyword{internal} \alias{fa_map}
\keyword{internal} \alias{fa_commenting}
\keyword{internal} \alias{fa_commenting_o}
\keyword{internal} \alias{fa_houzz}
\keyword{internal} \alias{fa_vimeo}
\keyword{internal} \alias{fa_black_tie}
\keyword{internal} \alias{fa_fonticons}
\keyword{internal} \alias{fa_reddit_alien}
\keyword{internal} \alias{fa_edge}
\keyword{internal} \alias{fa_credit_card_alt}
\keyword{internal} \alias{fa_codiepie}
\keyword{internal} \alias{fa_modx}
\keyword{internal} \alias{fa_fort_awesome}
\keyword{internal} \alias{fa_usb}
\keyword{internal} \alias{fa_product_hunt}
\keyword{internal} \alias{fa_mixcloud}
\keyword{internal} \alias{fa_scribd}
\keyword{internal} \alias{fa_pause_circle}
\keyword{internal} \alias{fa_pause_circle_o}
\keyword{internal} \alias{fa_stop_circle}
\keyword{internal} \alias{fa_stop_circle_o}
\keyword{internal} \alias{fa_shopping_bag}
\keyword{internal} \alias{fa_shopping_basket}
\keyword{internal} \alias{fa_hashtag}
\keyword{internal} \alias{fa_bluetooth}
\keyword{internal} \alias{fa_bluetooth_b}
\keyword{internal} \alias{fa_percent}
\keyword{internal} \alias{fa_gitlab}
\keyword{internal} \alias{fa_wpbeginner}
\keyword{internal} \alias{fa_wpforms}
\keyword{internal} \alias{fa_envira}
\keyword{internal} \alias{fa_universal_access}
\keyword{internal} \alias{fa_wheelchair_alt}
\keyword{internal} \alias{fa_question_circle_o}
\keyword{internal} \alias{fa_blind}
\keyword{internal} \alias{fa_audio_description}
\keyword{internal} \alias{fa_volume_control_phone}
\keyword{internal} \alias{fa_braille}
\keyword{internal} \alias{fa_assistive_listening_systems}
\keyword{internal} \alias{fa_american_sign_language_interpreting}
\keyword{internal} \alias{fa_deaf}
\keyword{internal} \alias{fa_glide}
\keyword{internal} \alias{fa_glide_g}
\keyword{internal} \alias{fa_sign_language}
\keyword{internal} \alias{fa_low_vision}
\keyword{internal} \alias{fa_viadeo}
\keyword{internal} \alias{fa_viadeo_square}
\keyword{internal} \alias{fa_snapchat}
\keyword{internal} \alias{fa_snapchat_ghost}
\keyword{internal} \alias{fa_snapchat_square}
\keyword{internal} \alias{fa_pied_piper}
\keyword{internal} \alias{fa_first_order}
\keyword{internal} \alias{fa_yoast}
\keyword{internal} \alias{fa_themeisle}
\keyword{internal} \alias{fa_google_plus_official}
\keyword{internal} \alias{fa_font_awesome}
\keyword{internal} \alias{fa_handshake_o}
\keyword{internal} \alias{fa_envelope_open}
\keyword{internal} \alias{fa_envelope_open_o}
\keyword{internal} \alias{fa_linode}
\keyword{internal} \alias{fa_address_book}
\keyword{internal} \alias{fa_address_book_o}
\keyword{internal} \alias{fa_address_card}
\keyword{internal} \alias{fa_address_card_o}
\keyword{internal} \alias{fa_user_circle}
\keyword{internal} \alias{fa_user_circle_o}
\keyword{internal} \alias{fa_user_o}
\keyword{internal} \alias{fa_id_badge}
\keyword{internal} \alias{fa_id_card}
\keyword{internal} \alias{fa_id_card_o}
\keyword{internal} \alias{fa_quora}
\keyword{internal} \alias{fa_free_code_camp}
\keyword{internal} \alias{fa_telegram}
\keyword{internal} \alias{fa_thermometer_full}
\keyword{internal} \alias{fa_thermometer_three_quarters}
\keyword{internal} \alias{fa_thermometer_half}
\keyword{internal} \alias{fa_thermometer_quarter}
\keyword{internal} \alias{fa_thermometer_empty}
\keyword{internal} \alias{fa_shower}
\keyword{internal} \alias{fa_bath}
\keyword{internal} \alias{fa_podcast}
\keyword{internal} \alias{fa_window_maximize}
\keyword{internal} \alias{fa_window_minimize}
\keyword{internal} \alias{fa_window_restore}
\keyword{internal} \alias{fa_window_close}
\keyword{internal} \alias{fa_window_close_o}
\keyword{internal} \alias{fa_bandcamp}
\keyword{internal} \alias{fa_grav}
\keyword{internal} \alias{fa_etsy}
\keyword{internal} \alias{fa_imdb}
\keyword{internal} \alias{fa_ravelry}
\keyword{internal} \alias{fa_eercast}
\keyword{internal} \alias{fa_microchip}
\keyword{internal} \alias{fa_snowflake_o}
\keyword{internal} \alias{fa_superpowers}
\keyword{internal} \alias{fa_wpexplorer}
\keyword{internal} \alias{fa_meetup}
