% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\docType{data}
\name{fa_iconList}
\alias{fa_iconList}
\alias{fa}
\title{Insert icon from font awesome v4.7.0}
\format{An object of class \code{character} of length 912.}
\usage{
fa_iconList

fa(name = "font-awesome", size = 1, fixed_width = FALSE,
  animate = "still", rotate = 0, flip = "none", border = FALSE,
  pull = NULL, other = NULL)
}
\arguments{
\item{name}{A string indicating the icon name.}

\item{size}{Size of the icon relative to font size. Options are 1, lg (33%
increase), 2, 3, 4, or 5.}

\item{fixed_width}{If TRUE, the icon is set to a fixed width}

\item{animate}{'still', 'spin', or 'pulse'.}

\item{rotate}{Rotate degree: 0, 90, 180, or 270.}

\item{flip}{'none', 'horizontal', 'vertical'.}

\item{border}{If TRUE, draws a border around the icon.}

\item{pull}{Pulls icon to either 'left' or 'right' and wraps proceeding text
around it.}

\item{other}{Character vector of other parameters directly added to the icon classes}
}
\description{
Insert icon from font awesome v4.7.0
}
\details{
\code{fa_*} is equivalent to \code{fa(name = *)}, which utilises the auto completion.
}
\references{
\href{http://fontawesome.io/icons/}{Font awesome}
}
\keyword{datasets}
