% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather_icons.R
\name{download_feather_icons}
\alias{download_feather_icons}
\alias{feather_icons}
\title{Feather Icons}
\usage{
download_feather_icons(version = "dev")

feather_icons(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\value{
\code{download_feather_icons()} invisibly returns the \code{feather_icons}
icon set after downloading and installing its SVG files locally.

\code{feather_icons(name)} returns an \code{icons} vector (an SVG tag) for the
requested icon.
}
\description{
Feather Icons
}
\section{License}{

Feather Icons are distributed under the \href{https://github.com/feathericons/feather/blob/main/LICENSE}{MIT License}.
}

