% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather_icons.R
\name{download_feather_icons}
\alias{download_feather_icons}
\alias{feather_icons}
\title{Feather Icons}
\usage{
download_feather_icons(version = "dev")

feather_icons(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\description{
Feather Icons
}
