% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\docType{data}
\name{download_fontawesome}
\alias{download_fontawesome}
\alias{fontawesome}
\title{Font Awesome icons}
\format{An object of class \code{NULL} of length 0.}
\usage{
download_fontawesome(version = "dev")

fontawesome
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\description{
Font Awesome icons
}
\keyword{datasets}
