% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome.R
\name{download_fontawesome}
\alias{download_fontawesome}
\alias{fontawesome}
\title{Font Awesome icons}
\usage{
download_fontawesome(version = "dev")

fontawesome(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\description{
Font Awesome icons
}
