% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_sigmau}
\alias{get_sigmau}
\title{Calculate \eqn{\Sigma_u}{Sigma_u}, the unexplained residual}
\usage{
get_sigmau(data, cid, clist)
}
\arguments{
\item{data}{A data matrix.}

\item{cid}{A vector of cluster sizes.}

\item{clist}{A vector of cluster assignments.}
}
\value{
A list containing the SigmaU matrix (\code{sigma_u}) and its Frobenius norm (\code{sigma_u_norm}).
}
\description{
This function calculates \eqn{\Sigma_u}{Sigma_u} and its Frobenius norm,
given the data and subnetwork detection results.
}
\examples{
data <- matrix(runif(200), nrow = 20)
cid <- c(5, 5)
clist <- sample(1:2, 20, replace = TRUE)
result <- sigmau(data, cid, clist)
print(result)

}
