% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_material.R
\name{download_google_material}
\alias{download_google_material}
\alias{google_material}
\title{Google's Material design icons}
\usage{
download_google_material(version = "dev")

google_material(name, category = NULL, theme = NULL)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}

\item{category}{The icon's category, either "action", "alert", "av",
"communication", "content", "device", "editor", "file", "hardware", "home",
"image", "maps", "navigation", "notification", "places", "social", or
"toggle"}

\item{theme}{The style variant for the icon, requires v4.0.0 or greater.
Available themes vary by icon, but can be either "filled", "outlined",
"round", "sharp", or "twotone". If NULL, it will default to the
first available variant in this order.}
}
\value{
\code{download_google_material()} invisibly returns the
\code{google_material} icon set after downloading and installing its SVG
files locally.

\code{google_material(name, category, theme)} returns an \code{icon} object (an
SVG tag) for the requested icon.
}
\description{
Google's Material design icons
}
\section{License}{

Material Design Icons are distributed under the
\href{https://github.com/google/material-design-icons/blob/master/LICENSE}{Apache License 2.0}.
}

\seealso{
https://material.io/resources/icons/
}
