% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_material.R
\name{download_google_material}
\alias{download_google_material}
\alias{google_material}
\title{Google's Material design icons}
\usage{
download_google_material(version = "dev")

google_material(name, category = NULL, theme = NULL)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}

\item{category}{The icon's category, either "action", "alert", "av",
"communication", "content", "device", "editor", "file", "hardware", "home",
"image", "maps", "navigation", "notification", "places", "social", or
"toggle"}

\item{theme}{The style variant for the icon, requires v4.0.0 or greater.
Available themes vary by icon, but can be either "filled", "outlined",
"round", "sharp", or "twotone". If NULL, it will default to the
first available variant in this order.}
}
\description{
Google's Material design icons
}
\seealso{
https://material.io/resources/icons/
}
