% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense.R
\name{greedy_peeling}
\alias{greedy_peeling}
\title{Greedy Peeling}
\usage{
greedy_peeling(Wp_DSD, lambda)
}
\arguments{
\item{Wp_DSD}{Input adjacency matrix}

\item{lambda}{Parameter for the objective function}
}
\value{
A list containing:\tabular{ll}{
\code{W_DSD_greedy} \tab Reordered adjacency matrix \cr
\tab \cr
\code{Clist} \tab Cluster list \cr
\tab \cr
\code{Node_Seq} \tab A vector of reordered nodes \cr
\tab \cr
\code{removing_node} \tab Nodes to be removed \cr
}
}
\description{
A simple implementation of the greedy algorithm from the "Denser than the Densest subgraph" paper (Tsourakakis 2013)
using the generalized objective function (Shuo 2023).
}
\details{
Note that this function only extracts ONE dense subgraph.
}
\examples{
data(sim)
greedy_peeling(sim, 1.4)
greedy_peeling(sim, 1.6)
}
