% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_peeling.R
\name{greedy_peeling}
\alias{greedy_peeling}
\title{Greedy Peeling}
\usage{
greedy_peeling(Wp_DSD, lambda)
}
\arguments{
\item{Wp_DSD}{Input adjacency matrix}

\item{lambda}{Parameter for the objective function}
}
\value{
A list with components
\itemize{
\item{W_DSD_greedy}{reordered adjacency matrix}
\item{Clist}{Cluster list}
\item{Node_Seq}{A vector of reordered nodes}
\item{removing_node}{Nodes to be removed}
}
}
\description{
A simple implementation of the greedy algorithm from the "Denser than the Densest subgraph" paper (Tsourakakis 2013)
using the generalized objective function (Shuo 2023).
}
\details{
Note that this function only extracts ONE dense subgraph.
}
\examples{
data(sim)
greedy_peeling(sim, 1.4)
greedy_peeling(sim, 1.6)
}
