% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_icons.R
\name{download_health_icons}
\alias{download_health_icons}
\alias{health_icons}
\title{Health icons}
\usage{
download_health_icons()

health_icons(name, category = NULL, theme = NULL)
}
\arguments{
\item{name}{Name of the icon}

\item{category}{The icon's category, either "blood", "body", "conditions",
"devices", "diagnostics", "emotions", "family", "medications", "objects",
"people", "places", "shapes", "specialties", "symbols", "typography", or
"vehicles"}

\item{theme}{The style variant for the icon, requires v4.0.0 or greater.
Available themes vary by icon, but can be either "filled", "negative",
"outline". If NULL, it will default to the
first available variant in this order.}
}
\description{
Health icons
}
