% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\docType{package}
\name{icon-package}
\alias{icon}
\alias{icon-package}
\title{icon: Use different icon fonts in R Markdown}
\description{
This package allows easy insertion of icons from font awesome into R Markdown.
}
\author{
\strong{Maintainer}: Earo Wang \email{earo.wang@gmail.com}

Authors:
\itemize{
  \item Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}
  \item Timothy Hyndman \email{timothy.hyndman@gmail.com}
}

}
\keyword{internal}
