% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\docType{package}
\name{icon-package}
\alias{icon}
\alias{icon-package}
\title{icon: SVG Icons for R Documents and Apps}
\description{
Easily embed SVG icons into R Markdown documents or shiny apps. A collection of popular icon sets are supported by the package, including Font Awesome, Academicons and Ionicons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mitchelloharawild/icon}
  \item Report bugs at \url{https://github.com/mitchelloharawild/icon/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Other contributors:
\itemize{
  \item Earo Wang [contributor]
  \item Timothy Hyndman [contributor]
}

}
\keyword{internal}
