% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\docType{package}
\name{icon-package}
\alias{icon}
\alias{icon-package}
\title{icon: Use different icon fonts in RMarkdown}
\description{
This package allows easy insertion of icons from font awesome, academicons and ionicons into RMarkdown.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropenscilabs/icon}
  \item Report bugs at \url{https://github.com/ropenscilabs/icon/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Earo Wang \email{earo.wang@gmail.com}
  \item Timothy Hyndman \email{timothy.hyndman@gmail.com}
}

}
\keyword{internal}
