% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{icon_find}
\alias{icon_find}
\title{Find icons in installed sets by name}
\usage{
icon_find(name, set = NULL)
}
\arguments{
\item{name}{The name of the icon}

\item{set}{Icon sets to search. If NULL, all available icons will be searched.}
}
\description{
Find icons in installed sets by name
}
