% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{icon_installed}
\alias{icon_installed}
\title{Check if an icon set is installed.}
\usage{
icon_installed(x)
}
\arguments{
\item{x}{An icon set (such as \code{\link{fontawesome}}).}
}
\description{
This function will return \code{TRUE} if the icons for an icon set are installed.
If they aren't, they can be installed using the appropriate \verb{download_*()}
function.
}
