% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{icon_save}
\alias{icon_save}
\title{Save icons into a local path}
\usage{
icon_save(icons, path = ".")
}
\arguments{
\item{icons}{A named list of icons, the names specify the icon's name in the new icon set.}

\item{path}{A directory to save the icons into.}
}
\value{
Invisibly, the \code{icon_set} created from the newly saved SVG files
at \code{path}.
}
\description{
Store icons in a local directory to simplify sharing of documents containing
icons. Bundling icons with your document or app avoids the need for
collaborators or servers to have all icons from a library installed.
}
\examples{
if(icon_installed(fontawesome) && icon_installed(ionicons)){

icon_save(
  list(arrow = fontawesome$solid$`arrow-right`, alarm = ionicons$alarm),
  path = "icons"
)

app_icons <- icon_set("icons")
app_icons$arrow

}

}
