% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{icon_style}
\alias{icon_style}
\title{Customise the style of an icon}
\usage{
icon_style(x, scale = NULL, fill = NULL, ...)
}
\arguments{
\item{scale}{Scaled size of the icon.}

\item{fill}{The colour for the fill.}

\item{...}{Other CSS rules for the icon style.}
}
\description{
Customise the style of an icon
}
