% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons-package.R
\docType{package}
\name{icons-package}
\alias{icons}
\alias{icons-package}
\title{icons: SVG Icons for R Documents and Apps}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Easily embed SVG icons into 'R Markdown' documents or 'Shiny' apps. A collection of popular icon sets are supported by the package, including Font Awesome, Academicons and Ionicons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.mitchelloharawild.com/icons/}
  \item \url{https://github.com/mitchelloharawild/icons}
  \item Report bugs at \url{https://github.com/mitchelloharawild/icons/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}
}

Other contributors:
\itemize{
  \item Earo Wang [contributor]
  \item Timothy Hyndman [contributor]
  \item Matthew T. Warkentin [contributor]
  \item Tom Jemmett \email{thomas.jemmett@nhs.net} (\href{https://orcid.org/0000-0002-6943-2990}{ORCID}) [contributor]
}

}
\keyword{internal}
