% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICONS-package.R
\docType{package}
\name{ICONS-package}
\alias{ICONS}
\alias{ICONS-package}
\title{ICONS: Integrative analysis of covariance matrix and network structure}
\description{
This package provides tools for performing community detection and semi-confirmatory factor analysis in statistical computing environments.
}
\author{
\strong{Maintainer}: Yezhi Pan \email{yezhipan7@gmail.com} (\href{https://orcid.org/0009-0001-4014-0302}{ORCID})

}
\keyword{internal}
