% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ii.R
\name{download_ii}
\alias{download_ii}
\alias{ii}
\title{Ionicons icons}
\usage{
download_ii(x, version = "dev", ...)

ii(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\description{
Ionicons icons
}
