% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ii.R
\docType{data}
\name{ii_iconList}
\alias{ii_iconList}
\alias{ii}
\title{Insert icon from ionicons v2.0.1}
\format{An object of class \code{character} of length 733.}
\usage{
ii_iconList

ii(name = "ionic", size = 1, fixed_width = FALSE,
  animate = "still", rotate = 0, flip = "none", border = FALSE,
  pull = NULL, other = NULL)
}
\arguments{
\item{name}{A string indicating the icon name.}

\item{size}{Size of the icon relative to font size. Options are 1, lg (33%
increase), 2, 3, 4, or 5.}

\item{fixed_width}{If TRUE, the icon is set to a fixed width}

\item{animate}{'still', 'spin', or 'pulse'.}

\item{rotate}{Rotate degree: 0, 90, 180, or 270.}

\item{flip}{'none', 'horizontal', 'vertical'.}

\item{border}{If TRUE, draws a border around the icon.}

\item{pull}{Pulls icon to either 'left' or 'right' and wraps proceeding text
around it.}

\item{other}{Character vector of other parameters directly added to the icon classes}
}
\description{
Insert icon from ionicons v2.0.1
}
\keyword{datasets}
