% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionicons.R
\name{download_ionicons}
\alias{download_ionicons}
\alias{ionicons}
\title{Ionicons icons}
\usage{
download_ionicons(version = "dev")

ionicons(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\value{
\code{download_ionicons()} invisibly returns the \code{ionicons} icon set
after downloading and installing its SVG files locally.

\code{ionicons(name)} returns an \code{icon} object (an SVG tag) for the
requested icon.
}
\description{
Ionicons icons
}
\section{License}{

Ionicons are distributed under the \href{https://github.com/ionic-team/ionicons/blob/main/LICENSE}{MIT License}.
}

