% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionicons.R
\name{download_ionicons}
\alias{download_ionicons}
\alias{ionicons}
\title{Ionicons icons}
\usage{
download_ionicons(version = "dev")

ionicons(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\description{
Ionicons icons
}
