% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ii.R
\docType{data}
\name{download_ionicons}
\alias{download_ionicons}
\alias{ionicons}
\title{Ionicons icons}
\format{An object of class \code{NULL} of length 0.}
\usage{
download_ionicons(version = "dev")

ionicons
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\description{
Ionicons icons
}
\keyword{datasets}
