% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{fa}
\alias{fa}
\alias{ii}
\alias{ai}
\alias{migrate_icon}
\title{Migration tools}
\usage{
fa(...)

ii(...)

ai(...)

migrate_icon(path, ext = c(".Rmd", ".R"))
}
\arguments{
\item{...}{Unused.}

\item{path}{A directory or file to migrate}

\item{ext}{File extensions to update}
}
\description{
The icon package has been overhauled to be more efficient and extensible.
Unfortunately this has led to several necessary interface changes. To
minimise impact, please use the \code{migrate_icon()} function to update your files
to use the new interface.
}
