% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octicons.R
\name{download_octicons}
\alias{download_octicons}
\alias{octicons}
\title{GitHub's Octicons}
\usage{
download_octicons(version = "dev")

octicons(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\value{
\code{download_octicons()} invisibly returns the \code{octicons} icon set
after downloading and installing its SVG files locally.

\code{octicons(name)} returns an \code{icons} vector (an SVG tag) for the
requested icon.
}
\description{
GitHub's Octicons
}
\section{License}{

Octicons are distributed under the \href{https://github.com/primer/octicons/blob/main/LICENSE}{MIT License}.
}

\seealso{
https://primer.style/octicons/
}
