% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octicons.R
\name{download_octicons}
\alias{download_octicons}
\alias{octicons}
\title{GitHub's Octicons}
\usage{
download_octicons(version = "dev")

octicons(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\description{
GitHub's Octicons
}
\seealso{
https://primer.style/octicons/
}
