% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_tuning_sigmau.R
\name{param_tuning_sigmau}
\alias{param_tuning_sigmau}
\title{Parameter Tuning for SigmaU}
\usage{
param_tuning_sigmau(Wp, data, prctile_vec, lam_vec)
}
\arguments{
\item{Wp}{A symmetric matrix representing pairwise potentials (weights).}

\item{data}{A data matrix used in the SigmaU function.}

\item{prctile_vec}{A vector of percentiles for the pairwise potentials to consider.}

\item{lam_vec}{A vector of lambda values to test.}
}
\value{
A list containing the optimal lambda value (\code{lambda_out}) and the corresponding cut-off value (\code{cut_out}).
}
\description{
This function tunes parameters for the SigmaU function by evaluating different
combinations of lambda and cut-off values.
}
\examples{
Wp <- matrix(runif(100), nrow = 10)
data <- matrix(runif(200), nrow = 20)
prctile_vec <- c(10, 20, 30)
lam_vec <- c(0.1, 0.5, 1.0)
result <- param_tuning_sigmau(Wp, data, prctile_vec, lam_vec)
print(result)

}
