% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMatrix}
\alias{plotMatrix}
\title{Visualize a matrix as a heatmap and save it to a file}
\usage{
plotMatrix(
  data,
  filepath = file.path(getwd(), "plot.tiff"),
  width = 850,
  height = 800,
  palette = "viridis",
  format = "tiff"
)
}
\arguments{
\item{data}{A numeric matrix.}

\item{filepath}{The output file path, default is the current working directory}

\item{width}{Width of the output image in pixels, default is 850.}

\item{height}{Height of the output image in pixels, default is 800.}

\item{palette}{Type of color palette to use: "jet" or "viridis", default is "viridis".}

\item{format}{File format for saving the image, default is "tiff". Supported formats: "tiff", "png", "jpeg", "svg".}
}
\value{
Nothing is explicitly returned; an image file is saved to the designated file path.
}
\description{
This function creates a heatmap to visualize a numeric matrix. The heatmap is saved to a file
at the specified filepath in one of several supported formats.
}
\examples{
mat <- matrix(rnorm(100), nrow = 10)
plotMatrix(mat)
}
