% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{read_icon}
\alias{read_icon}
\title{Read an individual icon}
\usage{
read_icon(x)
}
\arguments{
\item{x}{Path to the icon}
}
\value{
An \code{icons} vector: an SVG icon that can be printed, embedded
in R Markdown, or styled with \code{\link[=icon_style]{icon_style()}}. If \code{x} is a file path, it
is retained and can be retrieved with \code{\link[=icon_path]{icon_path()}}.
}
\description{
Read an individual icon
}
