% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfa.R
\name{scfa}
\alias{scfa}
\title{Semi-Confirmatory Factor Analysis}
\usage{
scfa(data, CID, Clist)
}
\arguments{
\item{data}{A data frame or matrix where rows represent observations and columns represent variables.}

\item{CID}{A numeric vector indicating the number of variables associated with each factor.}

\item{Clist}{A numeric vector indicating the column indices of the variables to be used in the analysis.}
}
\value{
A list containing:
\item{loading}{A matrix of factor loadings.}
\item{factorscore}{A matrix of estimated factor scores.}
\item{n.network}{Integer. Number of factors.}
\item{n.var}{Integer. Number of variables.}
\item{n.obs}{Integer. Number of observations.}
}
\description{
This function performs a semi-confirmatory factor analysis (SCFA) on the provided dataset.
}
