% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfa.R
\name{scfa}
\alias{scfa}
\title{Semi-Confirmatory Factor Analysis}
\usage{
scfa(
  data,
  CID,
  Clist,
  method = "Sample",
  epsilon = 1e-06,
  remove.singletons = TRUE
)
}
\arguments{
\item{data}{A scaled data frame or matrix where rows represent observations and columns represent variables.}

\item{CID}{A numeric vector indicating the number of variables associated with each factor.}

\item{Clist}{A numeric vector indicating the column indices of the variables to be used in the analysis.}

\item{method}{Character. Method of residual term estimation: "Sample" or "MLE".}

\item{epsilon}{A small regularization value for solving the inverse of Sigma_U.}

\item{remove.singletons}{Logical. Remove the singleton set for SCFA or not.}
}
\value{
A list containing:
\describe{
\item{loading}{A matrix of factor loadings.}
\item{factorscore}{A matrix of estimated factor scores.}
\item{sigma_u}{Matrix of off-diagonal residuals (Sample or MLE).}
\item{sigma_u_norm}{Frobenius norm of sigma_u.}
\item{sigma_u_diag}{Diagonalized residuals (Sample or MLE).}
\item{sigma_u_diag_norm}{Frobenius norm of sigma_u_diag.}
\item{n.network}{Integer. Number of factors.}
\item{n.var}{Integer. Number of variables.}
\item{n.obs}{Integer. Number of observations.}
\item{method}{The chosen estimation method.}
}
}
\description{
This function performs a semi-confirmatory factor analysis (SCFA) on the provided dataset.
}
