% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim}
\alias{sim}
\title{Simulated data}
\format{
A data frame with 100 subjects and 200 variables.
}
\source{
null
}
\usage{
sim
}
\description{
The data contains variables with complex inter-community and intra-community
correlations.
}
\examples{
data(sim)
matrix <- cor(sim)
plotMatrix(z = matrix)
results <- dense(matrix, 0.5, 0.5)
results$CID
}
\keyword{datasets}
