% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim}
\alias{sim}
\title{Simulated adjacency matrix}
\format{
\subsection{\code{sim}}{

A data frame with 100 rows and 100 columns:
\describe{
\item{V1}{}
\item{V2}{}
\item{V3}{}
...
}
}
}
\source{
null
}
\usage{
sim
}
\description{
details to be filled in
}
\keyword{datasets}
