% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_icons.R
\name{download_simple_icons}
\alias{download_simple_icons}
\alias{simple_icons}
\title{Simple Icons}
\usage{
download_simple_icons(version = "dev")

simple_icons(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\value{
\code{download_simple_icons()} invisibly returns the \code{simple_icons}
icon set after downloading and installing its SVG files locally.

\code{simple_icons(name)} returns an \code{icon} object (an SVG tag) for the
requested icon.
}
\description{
Simple Icons
}
\section{License}{

The Simple Icons artwork is distributed under the
\href{https://github.com/simple-icons/simple-icons/blob/develop/LICENSE.md}{CC0 1.0 Universal License}.
Many of the icons depict third-party brands and logos which may be
protected as trademarks; read Simple Icons'
\href{https://github.com/simple-icons/simple-icons/blob/develop/DISCLAIMER.md}{legal disclaimer}
before using them.
}

