% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_icons.R
\name{download_simple_icons}
\alias{download_simple_icons}
\alias{simple_icons}
\title{Simple Icons}
\usage{
download_simple_icons(version = "dev")

simple_icons(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\description{
Simple Icons
}
