% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_tiny_icons.R
\name{download_super_tiny_icons}
\alias{download_super_tiny_icons}
\alias{super_tiny_icons}
\title{Super Tiny Icons}
\usage{
download_super_tiny_icons(version = "dev")

super_tiny_icons(name)
}
\arguments{
\item{version}{Version of the library}

\item{name}{Name of the icon}
}
\value{
\code{download_super_tiny_icons()} invisibly returns the
\code{super_tiny_icons} icon set after downloading and installing its SVG
files locally.

\code{super_tiny_icons(name)} returns an \code{icon} object (an SVG tag) for the
requested icon.
}
\description{
Super Tiny Icons are miniscule SVG versions of popular website
and app logos, each under 1KB in size.
}
\seealso{
\url{https://github.com/edent/SuperTinyIcons}
}
